% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_describe_environment_status}
\alias{cloud9_describe_environment_status}
\title{Gets status information for an AWS Cloud9 development environment}
\usage{
cloud9_describe_environment_status(environmentId)
}
\arguments{
\item{environmentId}{[required] The ID of the environment to get status information about.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  status = "error"|"creating"|"connecting"|"ready"|"stopping"|"stopped"|"deleting",
  message = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets status information for an AWS Cloud9 development environment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_environment_status(
  environmentId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# 
svc$describe_environment_status(
  environmentId = "8d9967e2f0624182b74e7690ad69ebEX"
)
}

}
\keyword{internal}
