% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_tag_resource}
\alias{cloud9_tag_resource}
\title{Adds tags to an Cloud9 development environment}
\usage{
cloud9_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the Cloud9 development environment to
add tags to.}

\item{Tags}{[required] The list of tags to add to the given Cloud9 development environment.}
}
\description{
Adds tags to an Cloud9 development environment.

See \url{https://paws-r.github.io/docs/cloud9/tag_resource.html} for full documentation.
}
\keyword{internal}
