% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_create_package_group}
\alias{codeartifact_create_package_group}
\title{Creates a package group}
\usage{
codeartifact_create_package_group(
  domain,
  domainOwner = NULL,
  packageGroup,
  contactInfo = NULL,
  description = NULL,
  tags = NULL
)
}
\arguments{
\item{domain}{[required] The name of the domain in which you want to create a package group.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{packageGroup}{[required] The pattern of the package group to create. The pattern is also the
identifier of the package group.}

\item{contactInfo}{The contact information for the created package group.}

\item{description}{A description of the package group.}

\item{tags}{One or more tag key-value pairs for the package group.}
}
\description{
Creates a package group. For more information about creating package groups, including example CLI commands, see \href{https://docs.aws.amazon.com/codeartifact/latest/ug/create-package-group.html}{Create a package group} in the \emph{CodeArtifact User Guide}.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_create_package_group/} for full documentation.
}
\keyword{internal}
