% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_delete_domain}
\alias{codeartifact_delete_domain}
\title{Deletes a domain}
\usage{
codeartifact_delete_domain(domain, domainOwner = NULL)
}
\arguments{
\item{domain}{[required] The name of the domain to delete.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}
}
\description{
Deletes a domain. You cannot delete a domain that contains repositories. If you want to delete a domain with repositories, first delete its repositories.

See \url{https://paws-r.github.io/docs/codeartifact/delete_domain.html} for full documentation.
}
\keyword{internal}
