% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_delete_package_group}
\alias{codeartifact_delete_package_group}
\title{Deletes a package group}
\usage{
codeartifact_delete_package_group(domain, domainOwner = NULL, packageGroup)
}
\arguments{
\item{domain}{[required] The domain that contains the package group to be deleted.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{packageGroup}{[required] The pattern of the package group to be deleted.}
}
\description{
Deletes a package group. Deleting a package group does not delete packages or package versions associated with the package group. When a package group is deleted, the direct child package groups will become children of the package group's direct parent package group. Therefore, if any of the child groups are inheriting any settings from the parent, those settings could change.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_delete_package_group/} for full documentation.
}
\keyword{internal}
