% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_delete_repository}
\alias{codeartifact_delete_repository}
\title{Deletes a repository}
\usage{
codeartifact_delete_repository(domain, domainOwner = NULL, repository)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository to delete.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{repository}{[required] The name of the repository to delete.}
}
\description{
Deletes a repository.

See \url{https://paws-r.github.io/docs/codeartifact/delete_repository.html} for full documentation.
}
\keyword{internal}
