% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_describe_package_group}
\alias{codeartifact_describe_package_group}
\title{Returns a PackageGroupDescription object that contains information about
the requested package group}
\usage{
codeartifact_describe_package_group(domain, domainOwner = NULL, packageGroup)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the package group.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{packageGroup}{[required] The pattern of the requested package group.}
}
\description{
Returns a \href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_PackageGroupDescription.html}{PackageGroupDescription} object that contains information about the requested package group.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_describe_package_group/} for full documentation.
}
\keyword{internal}
