% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_describe_repository}
\alias{codeartifact_describe_repository}
\title{Returns a RepositoryDescription object that contains detailed
information about the requested repository}
\usage{
codeartifact_describe_repository(domain, domainOwner = NULL, repository)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository to describe.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{repository}{[required] A string that specifies the name of the requested repository.}
}
\description{
Returns a \code{RepositoryDescription} object that contains detailed information about the requested repository.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_describe_repository/} for full documentation.
}
\keyword{internal}
