% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_disassociate_external_connection}
\alias{codeartifact_disassociate_external_connection}
\title{Removes an existing external connection from a repository}
\usage{
codeartifact_disassociate_external_connection(
  domain,
  domainOwner = NULL,
  repository,
  externalConnection
)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository from which to remove
the external repository.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{repository}{[required] The name of the repository from which the external connection will be
removed.}

\item{externalConnection}{[required] The name of the external connection to be removed from the repository.}
}
\description{
Removes an existing external connection from a repository.

See \url{https://paws-r.github.io/docs/codeartifact/disassociate_external_connection.html} for full documentation.
}
\keyword{internal}
