% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_get_domain_permissions_policy}
\alias{codeartifact_get_domain_permissions_policy}
\title{Returns the resource policy attached to the specified domain}
\usage{
codeartifact_get_domain_permissions_policy(domain, domainOwner = NULL)
}
\arguments{
\item{domain}{[required] The name of the domain to which the resource policy is attached.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}
}
\description{
Returns the resource policy attached to the specified domain.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_get_domain_permissions_policy/} for full documentation.
}
\keyword{internal}
