% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_get_repository_endpoint}
\alias{codeartifact_get_repository_endpoint}
\title{Returns the endpoint of a repository for a specific package format}
\usage{
codeartifact_get_repository_endpoint(
  domain,
  domainOwner = NULL,
  repository,
  format,
  endpointType = NULL
)
}
\arguments{
\item{domain}{[required] The name of the domain that contains the repository.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain that contains the repository. It does not include dashes or
spaces.}

\item{repository}{[required] The name of the repository.}

\item{format}{[required] Returns which endpoint of a repository to return. A repository has one
endpoint for each package format.}

\item{endpointType}{A string that specifies the type of endpoint.}
}
\description{
Returns the endpoint of a repository for a specific package format. A repository has one endpoint for each package format:

See \url{https://www.paws-r-sdk.com/docs/codeartifact_get_repository_endpoint/} for full documentation.
}
\keyword{internal}
