% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_get_repository_permissions_policy}
\alias{codeartifact_get_repository_permissions_policy}
\title{Returns the resource policy that is set on a repository}
\usage{
codeartifact_get_repository_permissions_policy(
  domain,
  domainOwner = NULL,
  repository
)
}
\arguments{
\item{domain}{[required] The name of the domain containing the repository whose associated
resource policy is to be retrieved.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{repository}{[required] The name of the repository whose associated resource policy is to be
retrieved.}
}
\description{
Returns the resource policy that is set on a repository.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_get_repository_permissions_policy/} for full documentation.
}
\keyword{internal}
