% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_domains}
\alias{codeartifact_list_domains}
\title{Returns a list of DomainSummary objects for all domains owned by the
Amazon Web Services account that makes this call}
\usage{
codeartifact_list_domains(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns a list of \href{https://docs.aws.amazon.com/codeartifact/latest/APIReference/API_PackageVersionDescription.html}{DomainSummary} objects for all domains owned by the Amazon Web Services account that makes this call. Each returned \code{DomainSummary} object contains information about a domain.

See \url{https://paws-r.github.io/docs/codeartifact/list_domains.html} for full documentation.
}
\keyword{internal}
