% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_sub_package_groups}
\alias{codeartifact_list_sub_package_groups}
\title{Returns a list of direct children of the specified package group}
\usage{
codeartifact_list_sub_package_groups(
  domain,
  domainOwner = NULL,
  packageGroup,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{domain}{[required] The name of the domain which contains the package group from which to
list sub package groups.}

\item{domainOwner}{The 12-digit account number of the Amazon Web Services account that owns
the domain. It does not include dashes or spaces.}

\item{packageGroup}{[required] The pattern of the package group from which to list sub package groups.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns a list of direct children of the specified package group.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_list_sub_package_groups/} for full documentation.
}
\keyword{internal}
