% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_tag_resource}
\alias{codeartifact_tag_resource}
\title{Adds or updates tags for a resource in CodeArtifact}
\usage{
codeartifact_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to add or
update tags for.}

\item{tags}{[required] The tags you want to modify or add to the resource.}
}
\description{
Adds or updates tags for a resource in CodeArtifact.

See \url{https://www.paws-r-sdk.com/docs/codeartifact_tag_resource/} for full documentation.
}
\keyword{internal}
