% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_untag_resource}
\alias{codeartifact_untag_resource}
\title{Removes tags from a resource in CodeArtifact}
\usage{
codeartifact_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
tags from.}

\item{tagKeys}{[required] The tag key for each tag that you want to remove from the resource.}
}
\description{
Removes tags from a resource in CodeArtifact.

See \url{https://paws-r.github.io/docs/codeartifact/untag_resource.html} for full documentation.
}
\keyword{internal}
