% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_command_executions}
\alias{codebuild_batch_get_command_executions}
\title{Gets information about the command executions}
\usage{
codebuild_batch_get_command_executions(sandboxId, commandExecutionIds)
}
\arguments{
\item{sandboxId}{[required] A \code{sandboxId} or \code{sandboxArn}.}

\item{commandExecutionIds}{[required] A comma separated list of \code{commandExecutionIds}.}
}
\description{
Gets information about the command executions.

See \url{https://www.paws-r-sdk.com/docs/codebuild_batch_get_command_executions/} for full documentation.
}
\keyword{internal}
