% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_fleets}
\alias{codebuild_batch_get_fleets}
\title{Gets information about one or more compute fleets}
\usage{
codebuild_batch_get_fleets(names)
}
\arguments{
\item{names}{[required] The names or ARNs of the compute fleets.}
}
\description{
Gets information about one or more compute fleets.

See \url{https://www.paws-r-sdk.com/docs/codebuild_batch_get_fleets/} for full documentation.
}
\keyword{internal}
