% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_projects}
\alias{codebuild_batch_get_projects}
\title{Gets information about one or more build projects}
\usage{
codebuild_batch_get_projects(names)
}
\arguments{
\item{names}{[required] The names of the build projects.}
}
\description{
Gets information about one or more build projects.
}
\section{Request syntax}{
\preformatted{svc$batch_get_projects(
  names = list(
    "string"
  )
)
}
}

\keyword{internal}
