% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_sandboxes}
\alias{codebuild_batch_get_sandboxes}
\title{Gets information about the sandbox status}
\usage{
codebuild_batch_get_sandboxes(ids)
}
\arguments{
\item{ids}{[required] A comma separated list of \code{sandboxIds} or \code{sandboxArns}.}
}
\description{
Gets information about the sandbox status.

See \url{https://www.paws-r-sdk.com/docs/codebuild_batch_get_sandboxes/} for full documentation.
}
\keyword{internal}
