% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_fleet}
\alias{codebuild_delete_fleet}
\title{Deletes a compute fleet}
\usage{
codebuild_delete_fleet(arn)
}
\arguments{
\item{arn}{[required] The ARN of the compute fleet.}
}
\description{
Deletes a compute fleet. When you delete a compute fleet, its builds are not deleted.

See \url{https://www.paws-r-sdk.com/docs/codebuild_delete_fleet/} for full documentation.
}
\keyword{internal}
