% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_report_group}
\alias{codebuild_delete_report_group}
\title{Deletes a report group}
\usage{
codebuild_delete_report_group(arn, deleteReports)
}
\arguments{
\item{arn}{[required] The ARN of the report group to delete.}

\item{deleteReports}{If \code{true}, deletes any reports that belong to a report group before
deleting the report group.

If \code{false}, you must delete any reports in the report group. Use
\href{https://docs.aws.amazon.com/codebuild/latest/APIReference/API_ListReportsForReportGroup.html}{ListReportsForReportGroup}
to get the reports in a report group. Use
\href{https://docs.aws.amazon.com/codebuild/latest/APIReference/API_DeleteReport.html}{DeleteReport}
to delete the reports. If you call \code{DeleteReportGroup} for a report
group that contains one or more reports, an exception is thrown.}
}
\description{
Deletes a report group. Before you delete a report group, you must
delete its reports.
}
\section{Request syntax}{
\preformatted{svc$delete_report_group(
  arn = "string",
  deleteReports = TRUE|FALSE
)
}
}

\keyword{internal}
