% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_source_credentials}
\alias{codebuild_delete_source_credentials}
\title{Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source
credentials}
\usage{
codebuild_delete_source_credentials(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source
credentials.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_source_credentials(
  arn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
