% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_get_resource_policy}
\alias{codebuild_get_resource_policy}
\title{Gets a resource policy that is identified by its resource ARN}
\usage{
codebuild_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that is associated with the resource policy.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  policy = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a resource policy that is identified by its resource ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_resource_policy(
  resourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
