% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_curated_environment_images}
\alias{codebuild_list_curated_environment_images}
\title{Gets information about Docker images that are managed by AWS CodeBuild}
\usage{
codebuild_list_curated_environment_images()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  platforms = list(
    list(
      platform = "DEBIAN"|"AMAZON_LINUX"|"UBUNTU"|"WINDOWS_SERVER",
      languages = list(
        list(
          language = "JAVA"|"PYTHON"|"NODE_JS"|"RUBY"|"GOLANG"|"DOCKER"|"ANDROID"|"DOTNET"|"BASE"|"PHP",
          images = list(
            list(
              name = "string",
              description = "string",
              versions = list(
                "string"
              )
            )
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about Docker images that are managed by AWS CodeBuild.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_curated_environment_images()
}\if{html}{\out{</div>}}
}

\keyword{internal}
