% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_sandboxes_for_project}
\alias{codebuild_list_sandboxes_for_project}
\title{Gets a list of sandboxes for a given project}
\usage{
codebuild_list_sandboxes_for_project(
  projectName,
  maxResults = NULL,
  sortOrder = NULL,
  nextToken = NULL
)
}
\arguments{
\item{projectName}{[required] The CodeBuild project name.}

\item{maxResults}{The maximum number of sandbox records to be retrieved.}

\item{sortOrder}{The order in which sandbox records should be retrieved.}

\item{nextToken}{The next token, if any, to get paginated results. You will get this
value from previous execution of list sandboxes.}
}
\description{
Gets a list of sandboxes for a given project.

See \url{https://www.paws-r-sdk.com/docs/codebuild_list_sandboxes_for_project/} for full documentation.
}
\keyword{internal}
