% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_retry_build_batch}
\alias{codebuild_retry_build_batch}
\title{Restarts a failed batch build}
\usage{
codebuild_retry_build_batch(id, idempotencyToken, retryType)
}
\arguments{
\item{id}{Specifies the identifier of the batch build to restart.}

\item{idempotencyToken}{A unique, case sensitive identifier you provide to ensure the
idempotency of the \code{RetryBuildBatch} request. The token is included in
the \code{RetryBuildBatch} request and is valid for five minutes. If you
repeat the \code{RetryBuildBatch} request with the same token, but change a
parameter, AWS CodeBuild returns a parameter mismatch error.}

\item{retryType}{Specifies the type of retry to perform.}
}
\description{
Restarts a failed batch build. Only batch builds that have failed can be
retried.
}
\section{Request syntax}{
\preformatted{svc$retry_build_batch(
  id = "string",
  idempotencyToken = "string",
  retryType = "RETRY_ALL_BUILDS"|"RETRY_FAILED_BUILDS"
)
}
}

\keyword{internal}
