% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_start_command_execution}
\alias{codebuild_start_command_execution}
\title{Starts a command execution}
\usage{
codebuild_start_command_execution(sandboxId, command, type = NULL)
}
\arguments{
\item{sandboxId}{[required] A \code{sandboxId} or \code{sandboxArn}.}

\item{command}{[required] The command that needs to be executed.}

\item{type}{The command type.}
}
\description{
Starts a command execution.

See \url{https://www.paws-r-sdk.com/docs/codebuild_start_command_execution/} for full documentation.
}
\keyword{internal}
