% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_start_sandbox}
\alias{codebuild_start_sandbox}
\title{Starts a sandbox}
\usage{
codebuild_start_sandbox(projectName = NULL, idempotencyToken = NULL)
}
\arguments{
\item{projectName}{The CodeBuild project name.}

\item{idempotencyToken}{A unique client token.}
}
\description{
Starts a sandbox.

See \url{https://www.paws-r-sdk.com/docs/codebuild_start_sandbox/} for full documentation.
}
\keyword{internal}
