% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_start_sandbox_connection}
\alias{codebuild_start_sandbox_connection}
\title{Starts a sandbox connection}
\usage{
codebuild_start_sandbox_connection(sandboxId)
}
\arguments{
\item{sandboxId}{[required] A \code{sandboxId} or \code{sandboxArn}.}
}
\description{
Starts a sandbox connection.

See \url{https://www.paws-r-sdk.com/docs/codebuild_start_sandbox_connection/} for full documentation.
}
\keyword{internal}
