% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_stop_sandbox}
\alias{codebuild_stop_sandbox}
\title{Stops a sandbox}
\usage{
codebuild_stop_sandbox(id)
}
\arguments{
\item{id}{[required] Information about the requested sandbox ID.}
}
\description{
Stops a sandbox.

See \url{https://www.paws-r-sdk.com/docs/codebuild_stop_sandbox/} for full documentation.
}
\keyword{internal}
