% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_update_project_visibility}
\alias{codebuild_update_project_visibility}
\title{Changes the public visibility for a project}
\usage{
codebuild_update_project_visibility(
  projectArn,
  projectVisibility,
  resourceAccessRole = NULL
)
}
\arguments{
\item{projectArn}{[required] The Amazon Resource Name (ARN) of the build project.}

\item{projectVisibility}{[required]}

\item{resourceAccessRole}{The ARN of the IAM role that enables CodeBuild to access the CloudWatch
Logs and Amazon S3 artifacts for the project's builds.}
}
\description{
Changes the public visibility for a project. The project's build results, logs, and artifacts are available to the general public. For more information, see \href{https://docs.aws.amazon.com/codebuild/latest/userguide/public-builds.html}{Public build projects} in the \emph{CodeBuild User Guide}.

See \url{https://paws-r.github.io/docs/codebuild/update_project_visibility.html} for full documentation.
}
\keyword{internal}
