% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_create_project}
\alias{codecatalyst_create_project}
\title{Creates a project in a specified space}
\usage{
codecatalyst_create_project(spaceName, displayName, description = NULL)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{displayName}{[required] The friendly name of the project that will be displayed to users.}

\item{description}{The description of the project. This description will be displayed to
all users of the project. We recommend providing a brief description of
the project and its intended purpose.}
}
\description{
Creates a project in a specified space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_create_project/} for full documentation.
}
\keyword{internal}
