% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_create_source_repository}
\alias{codecatalyst_create_source_repository}
\title{Creates an empty Git-based source repository in a specified project}
\usage{
codecatalyst_create_source_repository(
  spaceName,
  projectName,
  name,
  description = NULL
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{name}{[required] The name of the source repository. For more information about name
requirements, see \href{https://docs.aws.amazon.com/codecatalyst/latest/userguide/source-quotas.html}{Quotas for source repositories}.}

\item{description}{The description of the source repository.}
}
\description{
Creates an empty Git-based source repository in a specified project. The repository is created with an initial empty commit with a default branch named \code{main}.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_create_source_repository/} for full documentation.
}
\keyword{internal}
