% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_create_source_repository_branch}
\alias{codecatalyst_create_source_repository_branch}
\title{Creates a branch in a specified source repository in Amazon CodeCatalyst}
\usage{
codecatalyst_create_source_repository_branch(
  spaceName,
  projectName,
  sourceRepositoryName,
  name,
  headCommitId = NULL
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{sourceRepositoryName}{[required] The name of the repository where you want to create a branch.}

\item{name}{[required] The name for the branch you're creating.}

\item{headCommitId}{The commit ID in an existing branch from which you want to create the
new branch.}
}
\description{
Creates a branch in a specified source repository in Amazon CodeCatalyst.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_create_source_repository_branch/} for full documentation.
}
\keyword{internal}
