% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_delete_project}
\alias{codecatalyst_delete_project}
\title{Deletes a project in a space}
\usage{
codecatalyst_delete_project(spaceName, name)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{name}{[required] The name of the project in the space. To retrieve a list of project
names, use \code{\link[=codecatalyst_list_projects]{list_projects}}.}
}
\description{
Deletes a project in a space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_delete_project/} for full documentation.
}
\keyword{internal}
