% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_delete_source_repository}
\alias{codecatalyst_delete_source_repository}
\title{Deletes a source repository in Amazon CodeCatalyst}
\usage{
codecatalyst_delete_source_repository(spaceName, projectName, name)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{name}{[required] The name of the source repository.}
}
\description{
Deletes a source repository in Amazon CodeCatalyst. You cannot use this API to delete a linked repository. It can only be used to delete a Amazon CodeCatalyst source repository.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_delete_source_repository/} for full documentation.
}
\keyword{internal}
