% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_delete_space}
\alias{codecatalyst_delete_space}
\title{Deletes a space}
\usage{
codecatalyst_delete_space(name)
}
\arguments{
\item{name}{[required] The name of the space. To retrieve a list of space names, use
\code{\link[=codecatalyst_list_spaces]{list_spaces}}.}
}
\description{
Deletes a space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_delete_space/} for full documentation.
}
\keyword{internal}
