% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_dev_environment}
\alias{codecatalyst_get_dev_environment}
\title{Returns information about a Dev Environment for a source repository in a
project}
\usage{
codecatalyst_get_dev_environment(spaceName, projectName, id)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{id}{[required] The system-generated unique ID of the Dev Environment for which you want
to view information. To retrieve a list of Dev Environment IDs, use
\code{\link[=codecatalyst_list_dev_environments]{list_dev_environments}}.}
}
\description{
Returns information about a Dev Environment for a source repository in a project. Dev Environments are specific to the user who creates them.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_dev_environment/} for full documentation.
}
\keyword{internal}
