% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_project}
\alias{codecatalyst_get_project}
\title{Returns information about a project}
\usage{
codecatalyst_get_project(spaceName, name)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{name}{[required] The name of the project in the space.}
}
\description{
Returns information about a project.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_project/} for full documentation.
}
\keyword{internal}
