% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_source_repository}
\alias{codecatalyst_get_source_repository}
\title{Returns information about a source repository}
\usage{
codecatalyst_get_source_repository(spaceName, projectName, name)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{name}{[required] The name of the source repository.}
}
\description{
Returns information about a source repository.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_source_repository/} for full documentation.
}
\keyword{internal}
