% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_source_repository_clone_urls}
\alias{codecatalyst_get_source_repository_clone_urls}
\title{Returns information about the URLs that can be used with a Git client to
clone a source repository}
\usage{
codecatalyst_get_source_repository_clone_urls(
  spaceName,
  projectName,
  sourceRepositoryName
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{sourceRepositoryName}{[required] The name of the source repository.}
}
\description{
Returns information about the URLs that can be used with a Git client to clone a source repository.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_source_repository_clone_urls/} for full documentation.
}
\keyword{internal}
