% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_space}
\alias{codecatalyst_get_space}
\title{Returns information about an space}
\usage{
codecatalyst_get_space(name)
}
\arguments{
\item{name}{[required] The name of the space.}
}
\description{
Returns information about an space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_space/} for full documentation.
}
\keyword{internal}
