% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_subscription}
\alias{codecatalyst_get_subscription}
\title{Returns information about the Amazon Web Services account used for
billing purposes and the billing plan for the space}
\usage{
codecatalyst_get_subscription(spaceName)
}
\arguments{
\item{spaceName}{[required] The name of the space.}
}
\description{
Returns information about the Amazon Web Services account used for billing purposes and the billing plan for the space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_subscription/} for full documentation.
}
\keyword{internal}
