% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_user_details}
\alias{codecatalyst_get_user_details}
\title{Returns information about a user}
\usage{
codecatalyst_get_user_details(id = NULL, userName = NULL)
}
\arguments{
\item{id}{The system-generated unique ID of the user.}

\item{userName}{The name of the user as displayed in Amazon CodeCatalyst.}
}
\description{
Returns information about a user.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_user_details/} for full documentation.
}
\keyword{internal}
