% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_workflow}
\alias{codecatalyst_get_workflow}
\title{Returns information about a workflow}
\usage{
codecatalyst_get_workflow(spaceName, id, projectName)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{id}{[required] The ID of the workflow. To rerieve a list of workflow IDs, use
\code{\link[=codecatalyst_list_workflows]{list_workflows}}.}

\item{projectName}{[required] The name of the project in the space.}
}
\description{
Returns information about a workflow.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_workflow/} for full documentation.
}
\keyword{internal}
