% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_get_workflow_run}
\alias{codecatalyst_get_workflow_run}
\title{Returns information about a specified run of a workflow}
\usage{
codecatalyst_get_workflow_run(spaceName, id, projectName)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{id}{[required] The ID of the workflow run. To retrieve a list of workflow run IDs, use
\code{\link[=codecatalyst_list_workflow_runs]{list_workflow_runs}}.}

\item{projectName}{[required] The name of the project in the space.}
}
\description{
Returns information about a specified run of a workflow.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_get_workflow_run/} for full documentation.
}
\keyword{internal}
