% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_list_dev_environment_sessions}
\alias{codecatalyst_list_dev_environment_sessions}
\title{Retrieves a list of active sessions for a Dev Environment in a project}
\usage{
codecatalyst_list_dev_environment_sessions(
  spaceName,
  projectName,
  devEnvironmentId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{devEnvironmentId}{[required] The system-generated unique ID of the Dev Environment.}

\item{nextToken}{A token returned from a call to this API to indicate the next batch of
results to return, if any.}

\item{maxResults}{The maximum number of results to show in a single call to this API. If
the number of results is larger than the number you specified, the
response will include a \code{NextToken} element, which you can use to obtain
additional results.}
}
\description{
Retrieves a list of active sessions for a Dev Environment in a project.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_list_dev_environment_sessions/} for full documentation.
}
\keyword{internal}
