% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_list_spaces}
\alias{codecatalyst_list_spaces}
\title{Retrieves a list of spaces}
\usage{
codecatalyst_list_spaces(nextToken = NULL)
}
\arguments{
\item{nextToken}{A token returned from a call to this API to indicate the next batch of
results to return, if any.}
}
\description{
Retrieves a list of spaces.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_list_spaces/} for full documentation.
}
\keyword{internal}
