% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_start_dev_environment_session}
\alias{codecatalyst_start_dev_environment_session}
\title{Starts a session for a specified Dev Environment}
\usage{
codecatalyst_start_dev_environment_session(
  spaceName,
  projectName,
  id,
  sessionConfiguration
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{id}{[required] The system-generated unique ID of the Dev Environment.}

\item{sessionConfiguration}{[required]}
}
\description{
Starts a session for a specified Dev Environment.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_start_dev_environment_session/} for full documentation.
}
\keyword{internal}
