% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_stop_dev_environment}
\alias{codecatalyst_stop_dev_environment}
\title{Pauses a specified Dev Environment and places it in a non-running state}
\usage{
codecatalyst_stop_dev_environment(spaceName, projectName, id)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{id}{[required] The system-generated unique ID of the Dev Environment.}
}
\description{
Pauses a specified Dev Environment and places it in a non-running state. Stopped Dev Environments do not consume compute minutes.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_stop_dev_environment/} for full documentation.
}
\keyword{internal}
