% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_stop_dev_environment_session}
\alias{codecatalyst_stop_dev_environment_session}
\title{Stops a session for a specified Dev Environment}
\usage{
codecatalyst_stop_dev_environment_session(
  spaceName,
  projectName,
  id,
  sessionId
)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{projectName}{[required] The name of the project in the space.}

\item{id}{[required] The system-generated unique ID of the Dev Environment. To obtain this
ID, use \code{\link[=codecatalyst_list_dev_environments]{list_dev_environments}}.}

\item{sessionId}{[required] The system-generated unique ID of the Dev Environment session. This ID
is returned by
\code{\link[=codecatalyst_start_dev_environment_session]{start_dev_environment_session}}.}
}
\description{
Stops a session for a specified Dev Environment.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_stop_dev_environment_session/} for full documentation.
}
\keyword{internal}
