% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_update_project}
\alias{codecatalyst_update_project}
\title{Changes one or more values for a project}
\usage{
codecatalyst_update_project(spaceName, name, description = NULL)
}
\arguments{
\item{spaceName}{[required] The name of the space.}

\item{name}{[required] The name of the project.}

\item{description}{The description of the project.}
}
\description{
Changes one or more values for a project.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_update_project/} for full documentation.
}
\keyword{internal}
