% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_update_space}
\alias{codecatalyst_update_space}
\title{Changes one or more values for a space}
\usage{
codecatalyst_update_space(name, description = NULL)
}
\arguments{
\item{name}{[required] The name of the space.}

\item{description}{The description of the space.}
}
\description{
Changes one or more values for a space.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_update_space/} for full documentation.
}
\keyword{internal}
