% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecatalyst_operations.R
\name{codecatalyst_verify_session}
\alias{codecatalyst_verify_session}
\title{Verifies whether the calling user has a valid Amazon CodeCatalyst login
and session}
\usage{
codecatalyst_verify_session()
}
\description{
Verifies whether the calling user has a valid Amazon CodeCatalyst login and session. If successful, this returns the ID of the user in Amazon CodeCatalyst.

See \url{https://www.paws-r-sdk.com/docs/codecatalyst_verify_session/} for full documentation.
}
\keyword{internal}
