% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_batch_associate_approval_rule_template_with_repositories}
\alias{codecommit_batch_associate_approval_rule_template_with_repositories}
\title{Creates an association between an approval rule template and one or more
specified repositories}
\usage{
codecommit_batch_associate_approval_rule_template_with_repositories(
  approvalRuleTemplateName, repositoryNames)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the template you want to associate with one or more
repositories.}

\item{repositoryNames}{[required] The names of the repositories you want to associate with the template.

The length constraint limit is for each string in the array. The array
itself can be empty.}
}
\value{
A list with the following syntax:\preformatted{list(
  associatedRepositoryNames = list(
    "string"
  ),
  errors = list(
    list(
      repositoryName = "string",
      errorCode = "string",
      errorMessage = "string"
    )
  )
)
}
}
\description{
Creates an association between an approval rule template and one or more
specified repositories.
}
\section{Request syntax}{
\preformatted{svc$batch_associate_approval_rule_template_with_repositories(
  approvalRuleTemplateName = "string",
  repositoryNames = list(
    "string"
  )
)
}
}

\keyword{internal}
