% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_batch_disassociate_approval_rule_template_from_repositories}
\alias{codecommit_batch_disassociate_approval_rule_template_from_repositories}
\title{Removes the association between an approval rule template and one or
more specified repositories}
\usage{
codecommit_batch_disassociate_approval_rule_template_from_repositories(
  approvalRuleTemplateName,
  repositoryNames
)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the template that you want to disassociate from one or more
repositories.}

\item{repositoryNames}{[required] The repository names that you want to disassociate from the approval
rule template.

The length constraint limit is for each string in the array. The array
itself can be empty.}
}
\description{
Removes the association between an approval rule template and one or more specified repositories.

See \url{https://www.paws-r-sdk.com/docs/codecommit_batch_disassociate_approval_rule_template_from_repositories/} for full documentation.
}
\keyword{internal}
