% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_create_branch}
\alias{codecommit_create_branch}
\title{Creates a branch in a repository and points the branch to a commit}
\usage{
codecommit_create_branch(repositoryName, branchName, commitId)
}
\arguments{
\item{repositoryName}{[required] The name of the repository in which you want to create the new branch.}

\item{branchName}{[required] The name of the new branch to create.}

\item{commitId}{[required] The ID of the commit to point the new branch to.}
}
\value{
An empty list.
}
\description{
Creates a branch in a repository and points the branch to a commit.

Calling the create branch operation does not set a repository's default
branch. To do this, call the update default branch operation.
}
\section{Request syntax}{
\preformatted{svc$create_branch(
  repositoryName = "string",
  branchName = "string",
  commitId = "string"
)
}
}

\keyword{internal}
