% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_branch}
\alias{codecommit_delete_branch}
\title{Deletes a branch from a repository, unless that branch is the default
branch for the repository}
\usage{
codecommit_delete_branch(repositoryName, branchName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branch to be deleted.}

\item{branchName}{[required] The name of the branch to delete.}
}
\description{
Deletes a branch from a repository, unless that branch is the default branch for the repository.

See \url{https://paws-r.github.io/docs/codecommit/delete_branch.html} for full documentation.
}
\keyword{internal}
