% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_file}
\alias{codecommit_delete_file}
\title{Deletes a specified file from a specified branch}
\usage{
codecommit_delete_file(repositoryName, branchName, filePath,
  parentCommitId, keepEmptyFolders, commitMessage, name, email)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the file to delete.}

\item{branchName}{[required] The name of the branch where the commit will be made deleting the file.}

\item{filePath}{[required] The fully-qualified path to the file that will be deleted, including the
full name and extension of that file. For example, /examples/file.md is
a fully qualified path to a file named file.md in a folder named
examples.}

\item{parentCommitId}{[required] The ID of the commit that is the tip of the branch where you want to
create the commit that will delete the file. This must be the HEAD
commit for the branch. The commit that deletes the file will be created
from this commit ID.}

\item{keepEmptyFolders}{Specifies whether to delete the folder or directory that contains the
file you want to delete if that file is the only object in the folder or
directory. By default, empty folders will be deleted. This includes
empty folders that are part of the directory structure. For example, if
the path to a file is dir1/dir2/dir3/dir4, and dir2 and dir3 are empty,
deleting the last file in dir4 will also delete the empty folders dir4,
dir3, and dir2.}

\item{commitMessage}{The commit message you want to include as part of deleting the file.
Commit messages are limited to 256 KB. If no message is specified, a
default message will be used.}

\item{name}{The name of the author of the commit that deletes the file. If no name
is specified, the user's ARN will be used as the author name and
committer name.}

\item{email}{The email address for the commit that deletes the file. If no email
address is specified, the email address will be left blank.}
}
\description{
Deletes a specified file from a specified branch. A commit is created on
the branch that contains the revision. The file will still exist in the
commits prior to the commit that contains the deletion.
}
\section{Request syntax}{
\preformatted{svc$delete_file(
  repositoryName = "string",
  branchName = "string",
  filePath = "string",
  parentCommitId = "string",
  keepEmptyFolders = TRUE|FALSE,
  commitMessage = "string",
  name = "string",
  email = "string"
)
}
}

\keyword{internal}
