% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_approval_rule_template}
\alias{codecommit_get_approval_rule_template}
\title{Returns information about a specified approval rule template}
\usage{
codecommit_get_approval_rule_template(approvalRuleTemplateName)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the approval rule template for which you want to get
information.}
}
\value{
A list with the following syntax:\preformatted{list(
  approvalRuleTemplate = list(
    approvalRuleTemplateId = "string",
    approvalRuleTemplateName = "string",
    approvalRuleTemplateDescription = "string",
    approvalRuleTemplateContent = "string",
    ruleContentSha256 = "string",
    lastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    creationDate = as.POSIXct(
      "2015-01-01"
    ),
    lastModifiedUser = "string"
  )
)
}
}
\description{
Returns information about a specified approval rule template.
}
\section{Request syntax}{
\preformatted{svc$get_approval_rule_template(
  approvalRuleTemplateName = "string"
)
}
}

\keyword{internal}
