% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_merge_options}
\alias{codecommit_get_merge_options}
\title{Returns information about the merge options available for merging two
specified branches}
\usage{
codecommit_get_merge_options(repositoryName, sourceCommitSpecifier,
  destinationCommitSpecifier, conflictDetailLevel,
  conflictResolutionStrategy)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the commits about which you
want to get merge options.}

\item{sourceCommitSpecifier}{[required] The branch, tag, HEAD, or other fully qualified reference used to
identify a commit. For example, a branch name or a full commit ID.}

\item{destinationCommitSpecifier}{[required] The branch, tag, HEAD, or other fully qualified reference used to
identify a commit. For example, a branch name or a full commit ID.}

\item{conflictDetailLevel}{The level of conflict detail to use. If unspecified, the default
FILE\_LEVEL is used, which will return a not mergeable result if the
same file has differences in both branches. If LINE\_LEVEL is specified,
a conflict will be considered not mergeable if the same file in both
branches has differences on the same line.}

\item{conflictResolutionStrategy}{Specifies which branch to use when resolving conflicts, or whether to
attempt automatically merging two versions of a file. The default is
NONE, which requires any conflicts to be resolved manually before the
merge operation will be successful.}
}
\description{
Returns information about the merge options available for merging two
specified branches. For details about why a particular merge option is
not available, use GetMergeConflicts or DescribeMergeConflicts.
}
\section{Request syntax}{
\preformatted{svc$get_merge_options(
  repositoryName = "string",
  sourceCommitSpecifier = "string",
  destinationCommitSpecifier = "string",
  conflictDetailLevel = "FILE_LEVEL"|"LINE_LEVEL",
  conflictResolutionStrategy = "NONE"|"ACCEPT_SOURCE"|"ACCEPT_DESTINATION"|"AUTOMERGE"
)
}
}

\keyword{internal}
