% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_repository_triggers}
\alias{codecommit_get_repository_triggers}
\title{Gets information about triggers configured for a repository}
\usage{
codecommit_get_repository_triggers(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository for which the trigger is configured.}
}
\value{
A list with the following syntax:\preformatted{list(
  configurationId = "string",
  triggers = list(
    list(
      name = "string",
      destinationArn = "string",
      customData = "string",
      branches = list(
        "string"
      ),
      events = list(
        "all"|"updateReference"|"createReference"|"deleteReference"
      )
    )
  )
)
}
}
\description{
Gets information about triggers configured for a repository.
}
\section{Request syntax}{
\preformatted{svc$get_repository_triggers(
  repositoryName = "string"
)
}
}

\keyword{internal}
