% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_list_associated_approval_rule_templates_for_repository}
\alias{codecommit_list_associated_approval_rule_templates_for_repository}
\title{Lists all approval rule templates that are associated with a specified
repository}
\usage{
codecommit_list_associated_approval_rule_templates_for_repository(
  repositoryName,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository for which you want to list all associated
approval rule templates.}

\item{nextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}

\item{maxResults}{A non-zero, non-negative integer used to limit the number of returned
results.}
}
\description{
Lists all approval rule templates that are associated with a specified repository.

See \url{https://paws-r.github.io/docs/codecommit/list_associated_approval_rule_templates_for_repository.html} for full documentation.
}
\keyword{internal}
