% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_list_branches}
\alias{codecommit_list_branches}
\title{Gets information about one or more branches in a repository}
\usage{
codecommit_list_branches(repositoryName, nextToken)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branches.}

\item{nextToken}{An enumeration token that allows the operation to batch the results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  branches = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about one or more branches in a repository.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_branches(
  repositoryName = "string",
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
