% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_post_comment_for_compared_commit}
\alias{codecommit_post_comment_for_compared_commit}
\title{Posts a comment on the comparison between two commits}
\usage{
codecommit_post_comment_for_compared_commit(
  repositoryName,
  beforeCommitId = NULL,
  afterCommitId,
  location = NULL,
  content,
  clientRequestToken = NULL
)
}
\arguments{
\item{repositoryName}{[required] The name of the repository where you want to post a comment on the
comparison between commits.}

\item{beforeCommitId}{To establish the directionality of the comparison, the full commit ID of
the before commit. Required for commenting on any commit unless that
commit is the initial commit.}

\item{afterCommitId}{[required] To establish the directionality of the comparison, the full commit ID of
the after commit.}

\item{location}{The location of the comparison where you want to comment.}

\item{content}{[required] The content of the comment you want to make.}

\item{clientRequestToken}{A unique, client-generated idempotency token that, when provided in a
request, ensures the request cannot be repeated with a changed
parameter. If a request is received with the same parameters and a token
is included, the request returns information about the initial request
that used that token.}
}
\description{
Posts a comment on the comparison between two commits.

See \url{https://www.paws-r-sdk.com/docs/codecommit_post_comment_for_compared_commit/} for full documentation.
}
\keyword{internal}
