% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_put_comment_reaction}
\alias{codecommit_put_comment_reaction}
\title{Adds or updates a reaction to a specified comment for the user whose
identity is used to make the request}
\usage{
codecommit_put_comment_reaction(commentId, reactionValue)
}
\arguments{
\item{commentId}{[required] The ID of the comment to which you want to add or update a reaction.}

\item{reactionValue}{[required] The emoji reaction you want to add or update. To remove a reaction,
provide a value of blank or null. You can also provide the value of
none. For information about emoji reaction values supported in AWS
CodeCommit, see the \href{https://docs.aws.amazon.com/codecommit/latest/userguide/how-to-commit-comment.html#emoji-reaction-table}{AWS CodeCommit User Guide}.}
}
\description{
Adds or updates a reaction to a specified comment for the user whose identity is used to make the request. You can only add or update a reaction for yourself. You cannot add, modify, or delete a reaction for another user.

See \url{https://paws-r.github.io/docs/codecommit/put_comment_reaction.html} for full documentation.
}
\keyword{internal}
