% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_tag_resource}
\alias{codecommit_tag_resource}
\title{Adds or updates tags for a resource in AWS CodeCommit}
\usage{
codecommit_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which you want to add
or update tags.}

\item{tags}{[required] The key-value pair to use when tagging this repository.}
}
\description{
Adds or updates tags for a resource in AWS CodeCommit. For a list of valid resources in AWS CodeCommit, see \href{https://docs.aws.amazon.com/codecommit/latest/userguide/#arn-formats}{CodeCommit Resources and Operations} in the \emph{AWS CodeCommit User Guide}.

See \url{https://www.paws-r-sdk.com/docs/codecommit_tag_resource/} for full documentation.
}
\keyword{internal}
