% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_test_repository_triggers}
\alias{codecommit_test_repository_triggers}
\title{Tests the functionality of repository triggers by sending information to
the trigger target}
\usage{
codecommit_test_repository_triggers(repositoryName, triggers)
}
\arguments{
\item{repositoryName}{[required] The name of the repository in which to test the triggers.}

\item{triggers}{[required] The list of triggers to test.}
}
\value{
A list with the following syntax:\preformatted{list(
  successfulExecutions = list(
    "string"
  ),
  failedExecutions = list(
    list(
      trigger = "string",
      failureMessage = "string"
    )
  )
)
}
}
\description{
Tests the functionality of repository triggers by sending information to
the trigger target. If real data is available in the repository, the
test sends data from the last commit. If no data is available, sample
data is generated.
}
\section{Request syntax}{
\preformatted{svc$test_repository_triggers(
  repositoryName = "string",
  triggers = list(
    list(
      name = "string",
      destinationArn = "string",
      customData = "string",
      branches = list(
        "string"
      ),
      events = list(
        "all"|"updateReference"|"createReference"|"deleteReference"
      )
    )
  )
)
}
}

\keyword{internal}
