% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_approval_rule_template_name}
\alias{codecommit_update_approval_rule_template_name}
\title{Updates the name of a specified approval rule template}
\usage{
codecommit_update_approval_rule_template_name(
  oldApprovalRuleTemplateName,
  newApprovalRuleTemplateName
)
}
\arguments{
\item{oldApprovalRuleTemplateName}{[required] The current name of the approval rule template.}

\item{newApprovalRuleTemplateName}{[required] The new name you want to apply to the approval rule template.}
}
\description{
Updates the name of a specified approval rule template.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_approval_rule_template_name/} for full documentation.
}
\keyword{internal}
