% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_default_branch}
\alias{codecommit_update_default_branch}
\title{Sets or changes the default branch name for the specified repository}
\usage{
codecommit_update_default_branch(repositoryName, defaultBranchName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to set or change the default branch for.}

\item{defaultBranchName}{[required] The name of the branch to set as the default.}
}
\value{
An empty list.
}
\description{
Sets or changes the default branch name for the specified repository.

If you use this operation to change the default branch name to the
current default branch name, a success message is returned even though
the default branch did not change.
}
\section{Request syntax}{
\preformatted{svc$update_default_branch(
  repositoryName = "string",
  defaultBranchName = "string"
)
}
}

\keyword{internal}
