% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_pull_request_approval_state}
\alias{codecommit_update_pull_request_approval_state}
\title{Updates the state of a user's approval on a pull request}
\usage{
codecommit_update_pull_request_approval_state(
  pullRequestId,
  revisionId,
  approvalState
)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request.}

\item{revisionId}{[required] The system-generated ID of the revision.}

\item{approvalState}{[required] The approval state to associate with the user on the pull request.}
}
\description{
Updates the state of a user's approval on a pull request. The user is derived from the signed-in account when the request is made.

See \url{https://paws-r.github.io/docs/codecommit/update_pull_request_approval_state.html} for full documentation.
}
\keyword{internal}
