% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_pull_request_description}
\alias{codecommit_update_pull_request_description}
\title{Replaces the contents of the description of a pull request}
\usage{
codecommit_update_pull_request_description(pullRequestId, description)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use
\code{\link[=codecommit_list_pull_requests]{list_pull_requests}}.}

\item{description}{[required] The updated content of the description for the pull request. This
content replaces the existing description.}
}
\description{
Replaces the contents of the description of a pull request.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_pull_request_description/} for full documentation.
}
\keyword{internal}
