% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_pull_request_status}
\alias{codecommit_update_pull_request_status}
\title{Updates the status of a pull request}
\usage{
codecommit_update_pull_request_status(pullRequestId, pullRequestStatus)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use
\code{\link[=codecommit_list_pull_requests]{list_pull_requests}}.}

\item{pullRequestStatus}{[required] The status of the pull request. The only valid operations are to update
the status from \code{OPEN} to \code{OPEN}, \code{OPEN} to \code{CLOSED} or from \code{CLOSED} to
\code{CLOSED}.}
}
\description{
Updates the status of a pull request.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_pull_request_status/} for full documentation.
}
\keyword{internal}
