% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_pull_request_title}
\alias{codecommit_update_pull_request_title}
\title{Replaces the title of a pull request}
\usage{
codecommit_update_pull_request_title(pullRequestId, title)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use
\code{\link[=codecommit_list_pull_requests]{list_pull_requests}}.}

\item{title}{[required] The updated title of the pull request. This replaces the existing title.}
}
\description{
Replaces the title of a pull request.

See \url{https://paws-r.github.io/docs/codecommit/update_pull_request_title.html} for full documentation.
}
\keyword{internal}
