% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_create_connection}
\alias{codeconnections_create_connection}
\title{Creates a connection that can then be given to other Amazon Web Services
services like CodePipeline so that it can access third-party code
repositories}
\usage{
codeconnections_create_connection(
  ProviderType = NULL,
  ConnectionName,
  Tags = NULL,
  HostArn = NULL
)
}
\arguments{
\item{ProviderType}{The name of the external provider where your third-party code repository
is configured.}

\item{ConnectionName}{[required] The name of the connection to be created.}

\item{Tags}{The key-value pair to use when tagging the resource.}

\item{HostArn}{The Amazon Resource Name (ARN) of the host associated with the
connection to be created.}
}
\description{
Creates a connection that can then be given to other Amazon Web Services services like CodePipeline so that it can access third-party code repositories. The connection is in pending status until the third-party connection handshake is completed from the console.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_create_connection/} for full documentation.
}
\keyword{internal}
