% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_get_host}
\alias{codeconnections_get_host}
\title{Returns the host ARN and details such as status, provider type,
endpoint, and, if applicable, the VPC configuration}
\usage{
codeconnections_get_host(HostArn)
}
\arguments{
\item{HostArn}{[required] The Amazon Resource Name (ARN) of the requested host.}
}
\description{
Returns the host ARN and details such as status, provider type, endpoint, and, if applicable, the VPC configuration.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_get_host/} for full documentation.
}
\keyword{internal}
