% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_get_repository_sync_status}
\alias{codeconnections_get_repository_sync_status}
\title{Returns details about the sync status for a repository}
\usage{
codeconnections_get_repository_sync_status(Branch, RepositoryLinkId, SyncType)
}
\arguments{
\item{Branch}{[required] The branch of the repository link for the requested repository sync
status.}

\item{RepositoryLinkId}{[required] The repository link ID for the requested repository sync status.}

\item{SyncType}{[required] The sync type of the requested sync status.}
}
\description{
Returns details about the sync status for a repository. A repository sync uses Git sync to push and pull changes from your remote repository.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_get_repository_sync_status/} for full documentation.
}
\keyword{internal}
