% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_get_resource_sync_status}
\alias{codeconnections_get_resource_sync_status}
\title{Returns the status of the sync with the Git repository for a specific
Amazon Web Services resource}
\usage{
codeconnections_get_resource_sync_status(ResourceName, SyncType)
}
\arguments{
\item{ResourceName}{[required] The name of the Amazon Web Services resource for the sync status with
the Git repository.}

\item{SyncType}{[required] The sync type for the sync status with the Git repository.}
}
\description{
Returns the status of the sync with the Git repository for a specific Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_get_resource_sync_status/} for full documentation.
}
\keyword{internal}
