% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_list_tags_for_resource}
\alias{codeconnections_list_tags_for_resource}
\title{Gets the set of key-value pairs (metadata) that are used to manage the
resource}
\usage{
codeconnections_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to get
information about tags, if any.}
}
\description{
Gets the set of key-value pairs (metadata) that are used to manage the resource.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
