% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconnections_operations.R
\name{codeconnections_tag_resource}
\alias{codeconnections_tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
codeconnections_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which you want to add
or update tags.}

\item{Tags}{[required] The tags you want to modify or add to the resource.}
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata that can be used to manage a resource.

See \url{https://www.paws-r-sdk.com/docs/codeconnections_tag_resource/} for full documentation.
}
\keyword{internal}
