% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_add_tags_to_on_premises_instances}
\alias{codedeploy_add_tags_to_on_premises_instances}
\title{Adds tags to on-premises instances}
\usage{
codedeploy_add_tags_to_on_premises_instances(tags, instanceNames)
}
\arguments{
\item{tags}{[required] The tag key-value pairs to add to the on-premises instances.

Keys and values are both required. Keys cannot be null or empty strings.
Value-only tags are not allowed.}

\item{instanceNames}{[required] The names of the on-premises instances to which to add tags.}
}
\description{
Adds tags to on-premises instances.

See \url{https://paws-r.github.io/docs/codedeploy/add_tags_to_on_premises_instances.html} for full documentation.
}
\keyword{internal}
