% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_applications}
\alias{codedeploy_batch_get_applications}
\title{Gets information about one or more applications}
\usage{
codedeploy_batch_get_applications(applicationNames)
}
\arguments{
\item{applicationNames}{[required] A list of application names separated by spaces. The maximum number of
application names you can specify is 100.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  applicationsInfo = list(
    list(
      applicationId = "string",
      applicationName = "string",
      createTime = as.POSIXct(
        "2015-01-01"
      ),
      linkedToGitHub = TRUE|FALSE,
      gitHubAccountName = "string",
      computePlatform = "Server"|"Lambda"|"ECS"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about one or more applications. The maximum number of
applications that can be returned is 100.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$batch_get_applications(
  applicationNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
