% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_deployment_instances}
\alias{codedeploy_batch_get_deployment_instances}
\title{This method works, but is deprecated}
\usage{
codedeploy_batch_get_deployment_instances(deploymentId, instanceIds)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{instanceIds}{[required] The unique IDs of instances used in the deployment. The maximum number
of instance IDs you can specify is 25.}
}
\description{
This method works, but is deprecated. Use \code{\link[=codedeploy_batch_get_deployment_targets]{batch_get_deployment_targets}} instead.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_batch_get_deployment_instances/} for full documentation.
}
\keyword{internal}
