% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_create_application}
\alias{codedeploy_create_application}
\title{Creates an application}
\usage{
codedeploy_create_application(
  applicationName,
  computePlatform = NULL,
  tags = NULL
)
}
\arguments{
\item{applicationName}{[required] The name of the application. This name must be unique with the
applicable IAM user or AWS account.}

\item{computePlatform}{The destination platform type for the deployment (\code{Lambda}, \code{Server}, or
\code{ECS}).}

\item{tags}{The metadata that you apply to CodeDeploy applications to help you
organize and categorize them. Each tag consists of a key and an optional
value, both of which you define.}
}
\description{
Creates an application.

See \url{https://paws-r.github.io/docs/codedeploy/create_application.html} for full documentation.
}
\keyword{internal}
