% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_application}
\alias{codedeploy_delete_application}
\title{Deletes an application}
\usage{
codedeploy_delete_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}
}
\value{
An empty list.
}
\description{
Deletes an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_application(
  applicationName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
