% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_deployment_config}
\alias{codedeploy_delete_deployment_config}
\title{Deletes a deployment configuration}
\usage{
codedeploy_delete_deployment_config(deploymentConfigName)
}
\arguments{
\item{deploymentConfigName}{[required] The name of a deployment configuration associated with the IAM user or
AWS account.}
}
\value{
An empty list.
}
\description{
Deletes a deployment configuration.

A deployment configuration cannot be deleted if it is currently in use.
Predefined configurations cannot be deleted.
}
\section{Request syntax}{
\preformatted{svc$delete_deployment_config(
  deploymentConfigName = "string"
)
}
}

\keyword{internal}
