% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_deployment_group}
\alias{codedeploy_delete_deployment_group}
\title{Deletes a deployment group}
\usage{
codedeploy_delete_deployment_group(applicationName, deploymentGroupName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}

\item{deploymentGroupName}{[required] The name of a deployment group for the specified application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  hooksNotCleanedUp = list(
    list(
      name = "string",
      hook = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a deployment group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_deployment_group(
  applicationName = "string",
  deploymentGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
