% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_application_revision}
\alias{codedeploy_get_application_revision}
\title{Gets information about an application revision}
\usage{
codedeploy_get_application_revision(applicationName, revision)
}
\arguments{
\item{applicationName}{[required] The name of the application that corresponds to the revision.}

\item{revision}{[required] Information about the application revision to get, including type and
location.}
}
\description{
Gets information about an application revision.

See \url{https://paws-r.github.io/docs/codedeploy/get_application_revision.html} for full documentation.
}
\keyword{internal}
