% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment_target}
\alias{codedeploy_get_deployment_target}
\title{Returns information about a deployment target}
\usage{
codedeploy_get_deployment_target(deploymentId, targetId)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{targetId}{[required] The unique ID of a deployment target.}
}
\description{
Returns information about a deployment target.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_get_deployment_target/} for full documentation.
}
\keyword{internal}
