% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_on_premises_instance}
\alias{codedeploy_get_on_premises_instance}
\title{Gets information about an on-premises instance}
\usage{
codedeploy_get_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance about which to get information.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  instanceInfo = list(
    instanceName = "string",
    iamSessionArn = "string",
    iamUserArn = "string",
    instanceArn = "string",
    registerTime = as.POSIXct(
      "2015-01-01"
    ),
    deregisterTime = as.POSIXct(
      "2015-01-01"
    ),
    tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about an on-premises instance.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_on_premises_instance(
  instanceName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
