% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_applications}
\alias{codedeploy_list_applications}
\title{Lists the applications registered with the IAM user or AWS account}
\usage{
codedeploy_list_applications(nextToken)
}
\arguments{
\item{nextToken}{An identifier returned from the previous list applications call. It can
be used to return the next set of applications in the list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  applications = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the applications registered with the IAM user or AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_applications(
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
