% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_deployment_configs}
\alias{codedeploy_list_deployment_configs}
\title{Lists the deployment configurations with the IAM user or AWS account}
\usage{
codedeploy_list_deployment_configs(nextToken)
}
\arguments{
\item{nextToken}{An identifier returned from the previous
\code{\link[=codedeploy_list_deployment_configs]{list_deployment_configs}} call. It
can be used to return the next set of deployment configurations in the
list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  deploymentConfigsList = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the deployment configurations with the IAM user or AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_deployment_configs(
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
