% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_register_on_premises_instance}
\alias{codedeploy_register_on_premises_instance}
\title{Registers an on-premises instance}
\usage{
codedeploy_register_on_premises_instance(instanceName, iamSessionArn,
  iamUserArn)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance to register.}

\item{iamSessionArn}{The ARN of the IAM session to associate with the on-premises instance.}

\item{iamUserArn}{The ARN of the IAM user to associate with the on-premises instance.}
}
\value{
An empty list.
}
\description{
Registers an on-premises instance.

Only one IAM ARN (an IAM session ARN or IAM user ARN) is supported in
the request. You cannot use both.
}
\section{Request syntax}{
\preformatted{svc$register_on_premises_instance(
  instanceName = "string",
  iamSessionArn = "string",
  iamUserArn = "string"
)
}
}

\keyword{internal}
