% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_stop_deployment}
\alias{codedeploy_stop_deployment}
\title{Attempts to stop an ongoing deployment}
\usage{
codedeploy_stop_deployment(deploymentId, autoRollbackEnabled)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{autoRollbackEnabled}{Indicates, when a deployment is stopped, whether instances that have
been updated should be rolled back to the previous version of the
application revision.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  status = "Pending"|"Succeeded",
  statusMessage = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Attempts to stop an ongoing deployment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_deployment(
  deploymentId = "string",
  autoRollbackEnabled = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
