% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_tag_resource}
\alias{codedeploy_tag_resource}
\title{Associates the list of tags in the input Tags parameter with the
resource identified by the ResourceArn input parameter}
\usage{
codedeploy_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of a resource, such as a CodeDeploy application or deployment
group.}

\item{Tags}{[required] A list of tags that \code{\link[=codedeploy_tag_resource]{tag_resource}} associates
with a resource. The resource is identified by the \code{ResourceArn} input
parameter.}
}
\description{
Associates the list of tags in the input \code{Tags} parameter with the resource identified by the \code{ResourceArn} input parameter.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_tag_resource/} for full documentation.
}
\keyword{internal}
