% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_add_notification_channels}
\alias{codeguruprofiler_add_notification_channels}
\title{Add up to 2 anomaly notifications channels for a profiling group}
\usage{
codeguruprofiler_add_notification_channels(channels, profilingGroupName)
}
\arguments{
\item{channels}{[required] One or 2 channels to report to when anomalies are detected.}

\item{profilingGroupName}{[required] The name of the profiling group that we are setting up notifications
for.}
}
\description{
Add up to 2 anomaly notifications channels for a profiling group.

See \url{https://paws-r.github.io/docs/codeguruprofiler/add_notification_channels.html} for full documentation.
}
\keyword{internal}
