% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_delete_profiling_group}
\alias{codeguruprofiler_delete_profiling_group}
\title{Deletes a profiling group}
\usage{
codeguruprofiler_delete_profiling_group(profilingGroupName)
}
\arguments{
\item{profilingGroupName}{[required] The name of the profiling group to delete.}
}
\description{
Deletes a profiling group.

See \url{https://paws-r.github.io/docs/codeguruprofiler/delete_profiling_group.html} for full documentation.
}
\keyword{internal}
