% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_describe_profiling_group}
\alias{codeguruprofiler_describe_profiling_group}
\title{Returns a ProfilingGroupDescription object that contains information
about the requested profiling group}
\usage{
codeguruprofiler_describe_profiling_group(profilingGroupName)
}
\arguments{
\item{profilingGroupName}{[required] The name of the profiling group to get information about.}
}
\description{
Returns a \href{https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_ProfilingGroupDescription.html}{\code{ProfilingGroupDescription}} object that contains information about the requested profiling group.

See \url{https://www.paws-r-sdk.com/docs/codeguruprofiler_describe_profiling_group/} for full documentation.
}
\keyword{internal}
