% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_get_policy}
\alias{codeguruprofiler_get_policy}
\title{Returns the JSON-formatted resource-based policy on a profiling group}
\usage{
codeguruprofiler_get_policy(profilingGroupName)
}
\arguments{
\item{profilingGroupName}{[required] The name of the profiling group.}
}
\description{
Returns the JSON-formatted resource-based policy on a profiling group.

See \url{https://www.paws-r-sdk.com/docs/codeguruprofiler_get_policy/} for full documentation.
}
\keyword{internal}
