% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_remove_notification_channel}
\alias{codeguruprofiler_remove_notification_channel}
\title{Remove one anomaly notifications channel for a profiling group}
\usage{
codeguruprofiler_remove_notification_channel(channelId, profilingGroupName)
}
\arguments{
\item{channelId}{[required] The id of the channel that we want to stop receiving notifications.}

\item{profilingGroupName}{[required] The name of the profiling group we want to change notification
configuration for.}
}
\description{
Remove one anomaly notifications channel for a profiling group.

See \url{https://paws-r.github.io/docs/codeguruprofiler/remove_notification_channel.html} for full documentation.
}
\keyword{internal}
