% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_submit_feedback}
\alias{codeguruprofiler_submit_feedback}
\title{Sends feedback to CodeGuru Profiler about whether the anomaly detected
by the analysis is useful or not}
\usage{
codeguruprofiler_submit_feedback(
  anomalyInstanceId,
  comment = NULL,
  profilingGroupName,
  type
)
}
\arguments{
\item{anomalyInstanceId}{[required] The universally unique identifier (UUID) of the
\href{https://docs.aws.amazon.com/codeguru/latest/profiler-api/API_AnomalyInstance.html}{\code{AnomalyInstance}}
object that is included in the analysis data.}

\item{comment}{Optional feedback about this anomaly.}

\item{profilingGroupName}{[required] The name of the profiling group that is associated with the analysis
data.}

\item{type}{[required] The feedback tpye. Thee are two valid values, \code{Positive} and \code{Negative}.}
}
\description{
Sends feedback to CodeGuru Profiler about whether the anomaly detected by the analysis is useful or not.

See \url{https://paws-r.github.io/docs/codeguruprofiler/submit_feedback.html} for full documentation.
}
\keyword{internal}
