% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_tag_resource}
\alias{codeguruprofiler_tag_resource}
\title{Use to assign one or more tags to a resource}
\usage{
codeguruprofiler_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that the tags are added
to.}

\item{tags}{[required] The list of tags that are added to the specified resource.}
}
\description{
Use to assign one or more tags to a resource.

See \url{https://paws-r.github.io/docs/codeguruprofiler/tag_resource.html} for full documentation.
}
\keyword{internal}
