% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_untag_resource}
\alias{codeguruprofiler_untag_resource}
\title{Use to remove one or more tags from a resource}
\usage{
codeguruprofiler_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that contains the tags to
remove.}

\item{tagKeys}{[required] A list of tag keys. Existing tags of resources with keys in this list
are removed from the specified resource.}
}
\description{
Use to remove one or more tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/codeguruprofiler_untag_resource/} for full documentation.
}
\keyword{internal}
