% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeguruprofiler_operations.R
\name{codeguruprofiler_update_profiling_group}
\alias{codeguruprofiler_update_profiling_group}
\title{Updates a profiling group}
\usage{
codeguruprofiler_update_profiling_group(
  agentOrchestrationConfig,
  profilingGroupName
)
}
\arguments{
\item{agentOrchestrationConfig}{[required] Specifies whether profiling is enabled or disabled for a profiling
group.}

\item{profilingGroupName}{[required] The name of the profiling group to update.}
}
\description{
Updates a profiling group.

See \url{https://paws-r.github.io/docs/codeguruprofiler/update_profiling_group.html} for full documentation.
}
\keyword{internal}
