% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_disassociate_repository}
\alias{codegurureviewer_disassociate_repository}
\title{Removes the association between Amazon CodeGuru Reviewer and a
repository}
\usage{
codegurureviewer_disassociate_repository(AssociationArn)
}
\arguments{
\item{AssociationArn}{[required] The Amazon Resource Name (ARN) of the
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_RepositoryAssociation.html}{\code{RepositoryAssociation}}
object. You can retrieve this ARN by calling
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_ListRepositoryAssociations.html}{\code{list_repository_associations}}
.}
}
\description{
Removes the association between Amazon CodeGuru Reviewer and a repository.

See \url{https://paws-r.github.io/docs/codegurureviewer/disassociate_repository.html} for full documentation.
}
\keyword{internal}
