% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_list_tags_for_resource}
\alias{codegurureviewer_list_tags_for_resource}
\title{Returns the list of tags associated with an associated repository
resource}
\usage{
codegurureviewer_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_RepositoryAssociation.html}{\code{RepositoryAssociation}}
object. You can retrieve this ARN by calling
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_ListRepositoryAssociations.html}{\code{list_repository_associations}}
.}
}
\description{
Returns the list of tags associated with an associated repository resource.

See \url{https://paws-r.github.io/docs/codegurureviewer/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
