% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_put_recommendation_feedback}
\alias{codegurureviewer_put_recommendation_feedback}
\title{Stores customer feedback for a CodeGuru Reviewer recommendation}
\usage{
codegurureviewer_put_recommendation_feedback(
  CodeReviewArn,
  RecommendationId,
  Reactions
)
}
\arguments{
\item{CodeReviewArn}{[required] The Amazon Resource Name (ARN) of the
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_CodeReview.html}{CodeReview}
object.}

\item{RecommendationId}{[required] The recommendation ID that can be used to track the provided
recommendations and then to collect the feedback.}

\item{Reactions}{[required] List for storing reactions. Reactions are utf-8 text code for emojis. If
you send an empty list it clears all your feedback.}
}
\description{
Stores customer feedback for a CodeGuru Reviewer recommendation. When this API is called again with different reactions the previous feedback is overwritten.

See \url{https://www.paws-r-sdk.com/docs/codegurureviewer_put_recommendation_feedback/} for full documentation.
}
\keyword{internal}
