% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_get_account_configuration}
\alias{codegurusecurity_get_account_configuration}
\title{Use to get account level configuration}
\usage{
codegurusecurity_get_account_configuration()
}
\description{
Use to get account level configuration.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_get_account_configuration/} for full documentation.
}
\keyword{internal}
