% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_get_scan}
\alias{codegurusecurity_get_scan}
\title{Returns details about a scan, including whether or not a scan has
completed}
\usage{
codegurusecurity_get_scan(scanName, runId = NULL)
}
\arguments{
\item{scanName}{[required] The name of the scan you want to view details about.}

\item{runId}{UUID that identifies the individual scan run you want to view details
about. You retrieve this when you call the
\code{\link[=codegurusecurity_create_scan]{create_scan}} operation. Defaults to the
latest scan run if missing.}
}
\description{
Returns details about a scan, including whether or not a scan has completed.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_get_scan/} for full documentation.
}
\keyword{internal}
