% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_list_tags_for_resource}
\alias{codegurusecurity_list_tags_for_resource}
\title{Returns a list of all tags associated with a scan}
\usage{
codegurusecurity_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the \code{ScanName} object. You can retrieve this ARN by calling
\code{\link[=codegurusecurity_create_scan]{create_scan}},
\code{\link[=codegurusecurity_list_scans]{list_scans}}, or
\code{\link[=codegurusecurity_get_scan]{get_scan}}.}
}
\description{
Returns a list of all tags associated with a scan.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
