% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_tag_resource}
\alias{codegurusecurity_tag_resource}
\title{Use to add one or more tags to an existing scan}
\usage{
codegurusecurity_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the \code{ScanName} object. You can retrieve this ARN by calling
\code{\link[=codegurusecurity_create_scan]{create_scan}},
\code{\link[=codegurusecurity_list_scans]{list_scans}}, or
\code{\link[=codegurusecurity_get_scan]{get_scan}}.}

\item{tags}{[required] An array of key-value pairs used to tag an existing scan. A tag is a
custom attribute label with two parts:
\itemize{
\item A tag key. For example, \code{CostCenter}, \code{Environment}, or \code{Secret}.
Tag keys are case sensitive.
\item An optional tag value field. For example, \code{111122223333},
\code{Production}, or a team name. Omitting the tag value is the same as
using an empty string. Tag values are case sensitive.
}}
}
\description{
Use to add one or more tags to an existing scan.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_tag_resource/} for full documentation.
}
\keyword{internal}
