% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_untag_resource}
\alias{codegurusecurity_untag_resource}
\title{Use to remove one or more tags from an existing scan}
\usage{
codegurusecurity_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the \code{ScanName} object. You can retrieve this ARN by calling
\code{\link[=codegurusecurity_create_scan]{create_scan}},
\code{\link[=codegurusecurity_list_scans]{list_scans}}, or
\code{\link[=codegurusecurity_get_scan]{get_scan}}.}

\item{tagKeys}{[required] A list of keys for each tag you want to remove from a scan.}
}
\description{
Use to remove one or more tags from an existing scan.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_untag_resource/} for full documentation.
}
\keyword{internal}
