% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_update_account_configuration}
\alias{codegurusecurity_update_account_configuration}
\title{Use to update account-level configuration with an encryption key}
\usage{
codegurusecurity_update_account_configuration(encryptionConfig)
}
\arguments{
\item{encryptionConfig}{[required] The KMS key ARN you want to use for encryption. Defaults to service-side
encryption if missing.}
}
\description{
Use to update account-level configuration with an encryption key.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_update_account_configuration/} for full documentation.
}
\keyword{internal}
