% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_create_pipeline}
\alias{codepipeline_create_pipeline}
\title{Creates a pipeline}
\usage{
codepipeline_create_pipeline(pipeline, tags = NULL)
}
\arguments{
\item{pipeline}{[required] Represents the structure of actions and stages to be performed in the
pipeline.}

\item{tags}{The tags for the pipeline.}
}
\description{
Creates a pipeline.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_create_pipeline/} for full documentation.
}
\keyword{internal}
