% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_delete_custom_action_type}
\alias{codepipeline_delete_custom_action_type}
\title{Marks a custom action as deleted}
\usage{
codepipeline_delete_custom_action_type(category, provider, version)
}
\arguments{
\item{category}{[required] The category of the custom action that you want to delete, such as
source or deploy.}

\item{provider}{[required] The provider of the service used in the custom action, such as AWS
CodeDeploy.}

\item{version}{[required] The version of the custom action to delete.}
}
\description{
Marks a custom action as deleted. \code{\link[=codepipeline_poll_for_jobs]{poll_for_jobs}} for the custom action fails after the action is marked for deletion. Used for custom actions only.

See \url{https://paws-r.github.io/docs/codepipeline/delete_custom_action_type.html} for full documentation.
}
\keyword{internal}
