% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_delete_pipeline}
\alias{codepipeline_delete_pipeline}
\title{Deletes the specified pipeline}
\usage{
codepipeline_delete_pipeline(name)
}
\arguments{
\item{name}{[required] The name of the pipeline to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes the specified pipeline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_pipeline(
  name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
