% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_deregister_webhook_with_third_party}
\alias{codepipeline_deregister_webhook_with_third_party}
\title{Removes the connection between the webhook that was created by
CodePipeline and the external tool with events to be detected}
\usage{
codepipeline_deregister_webhook_with_third_party(webhookName)
}
\arguments{
\item{webhookName}{The name of the webhook you want to deregister.}
}
\value{
An empty list.
}
\description{
Removes the connection between the webhook that was created by
CodePipeline and the external tool with events to be detected. Currently
supported only for webhooks that target an action type of GitHub.
}
\section{Request syntax}{
\preformatted{svc$deregister_webhook_with_third_party(
  webhookName = "string"
)
}
}

\keyword{internal}
