% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_get_action_type}
\alias{codepipeline_get_action_type}
\title{Returns information about an action type created for an external
provider, where the action is to be used by customers of the external
provider}
\usage{
codepipeline_get_action_type(category, owner, provider, version)
}
\arguments{
\item{category}{[required] Defines what kind of action can be taken in the stage. The following are
the valid values:
\itemize{
\item \code{Source}
\item \code{Build}
\item \code{Test}
\item \code{Deploy}
\item \code{Approval}
\item \code{Invoke}
}}

\item{owner}{[required] The creator of an action type that was created with any supported
integration model. There are two valid values: \code{AWS} and \code{ThirdParty}.}

\item{provider}{[required] The provider of the action type being called. The provider name is
specified when the action type is created.}

\item{version}{[required] A string that describes the action type version.}
}
\value{
A list with the following syntax:\preformatted{list(
  actionType = list(
    description = "string",
    executor = list(
      configuration = list(
        lambdaExecutorConfiguration = list(
          lambdaFunctionArn = "string"
        ),
        jobWorkerExecutorConfiguration = list(
          pollingAccounts = list(
            "string"
          ),
          pollingServicePrincipals = list(
            "string"
          )
        )
      ),
      type = "JobWorker"|"Lambda",
      policyStatementsTemplate = "string",
      jobTimeout = 123
    ),
    id = list(
      category = "Source"|"Build"|"Deploy"|"Test"|"Invoke"|"Approval",
      owner = "string",
      provider = "string",
      version = "string"
    ),
    inputArtifactDetails = list(
      minimumCount = 123,
      maximumCount = 123
    ),
    outputArtifactDetails = list(
      minimumCount = 123,
      maximumCount = 123
    ),
    permissions = list(
      allowedAccounts = list(
        "string"
      )
    ),
    properties = list(
      list(
        name = "string",
        optional = TRUE|FALSE,
        key = TRUE|FALSE,
        noEcho = TRUE|FALSE,
        queryable = TRUE|FALSE,
        description = "string"
      )
    ),
    urls = list(
      configurationUrl = "string",
      entityUrlTemplate = "string",
      executionUrlTemplate = "string",
      revisionUrlTemplate = "string"
    )
  )
)
}
}
\description{
Returns information about an action type created for an external
provider, where the action is to be used by customers of the external
provider. The action can be created with any supported integration
model.
}
\section{Request syntax}{
\preformatted{svc$get_action_type(
  category = "Source"|"Build"|"Deploy"|"Test"|"Invoke"|"Approval",
  owner = "string",
  provider = "string",
  version = "string"
)
}
}

\keyword{internal}
