% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_get_job_details}
\alias{codepipeline_get_job_details}
\title{Returns information about a job}
\usage{
codepipeline_get_job_details(jobId)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID for the job.}
}
\description{
Returns information about a job. Used for custom actions only.

See \url{https://paws-r.github.io/docs/codepipeline/get_job_details.html} for full documentation.
}
\keyword{internal}
