% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_action_executions}
\alias{codepipeline_list_action_executions}
\title{Lists the action executions that have occurred in a pipeline}
\usage{
codepipeline_list_action_executions(pipelineName, filter, maxResults,
  nextToken)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline for which you want to list action execution
history.}

\item{filter}{Input information used to filter action execution history.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned nextToken
value. The action execution history is limited to the most recent 12
months, based on action execution start times. Default value is 100.}

\item{nextToken}{The token that was returned from the previous ListActionExecutions call,
which can be used to return the next set of action executions in the
list.}
}
\description{
Lists the action executions that have occurred in a pipeline.
}
\section{Request syntax}{
\preformatted{svc$list_action_executions(
  pipelineName = "string",
  filter = list(
    pipelineExecutionId = "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
