% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_deploy_action_execution_targets}
\alias{codepipeline_list_deploy_action_execution_targets}
\title{Lists the targets for the deploy action}
\usage{
codepipeline_list_deploy_action_execution_targets(
  pipelineName = NULL,
  actionExecutionId,
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{pipelineName}{The name of the pipeline with the deploy action.}

\item{actionExecutionId}{[required] The execution ID for the deploy action.}

\item{filters}{Filters the targets for a specified deploy action.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{nextToken}{An identifier that was returned from the previous list action types
call, which can be used to return the next set of action types in the
list.}
}
\description{
Lists the targets for the deploy action.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_list_deploy_action_execution_targets/} for full documentation.
}
\keyword{internal}
