% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_rule_types}
\alias{codepipeline_list_rule_types}
\title{Lists the rules for the condition}
\usage{
codepipeline_list_rule_types(ruleOwnerFilter = NULL, regionFilter = NULL)
}
\arguments{
\item{ruleOwnerFilter}{The rule owner to filter on.}

\item{regionFilter}{The rule Region to filter on.}
}
\description{
Lists the rules for the condition.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_list_rule_types/} for full documentation.
}
\keyword{internal}
