% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_put_job_failure_result}
\alias{codepipeline_put_job_failure_result}
\title{Represents the failure of a job as returned to the pipeline by a job
worker}
\usage{
codepipeline_put_job_failure_result(jobId, failureDetails)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID of the job that failed. This is the same
ID returned from \code{\link[=codepipeline_poll_for_jobs]{poll_for_jobs}}.}

\item{failureDetails}{[required] The details about the failure of a job.}
}
\description{
Represents the failure of a job as returned to the pipeline by a job worker. Used for custom actions only.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_put_job_failure_result/} for full documentation.
}
\keyword{internal}
