% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_put_third_party_job_failure_result}
\alias{codepipeline_put_third_party_job_failure_result}
\title{Represents the failure of a third party job as returned to the pipeline
by a job worker}
\usage{
codepipeline_put_third_party_job_failure_result(
  jobId,
  clientToken,
  failureDetails
)
}
\arguments{
\item{jobId}{[required] The ID of the job that failed. This is the same ID returned from
\code{\link[=codepipeline_poll_for_third_party_jobs]{poll_for_third_party_jobs}}.}

\item{clientToken}{[required] The clientToken portion of the clientId and clientToken pair used to
verify that the calling entity is allowed access to the job and its
details.}

\item{failureDetails}{[required] Represents information about failure details.}
}
\description{
Represents the failure of a third party job as returned to the pipeline by a job worker. Used for partner actions only.

See \url{https://paws-r.github.io/docs/codepipeline/put_third_party_job_failure_result.html} for full documentation.
}
\keyword{internal}
