% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_register_webhook_with_third_party}
\alias{codepipeline_register_webhook_with_third_party}
\title{Configures a connection between the webhook that was created and the
external tool with events to be detected}
\usage{
codepipeline_register_webhook_with_third_party(webhookName)
}
\arguments{
\item{webhookName}{The name of an existing webhook created with PutWebhook to register with
a supported third party.}
}
\value{
An empty list.
}
\description{
Configures a connection between the webhook that was created and the
external tool with events to be detected.
}
\section{Request syntax}{
\preformatted{svc$register_webhook_with_third_party(
  webhookName = "string"
)
}
}

\keyword{internal}
