% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_rollback_stage}
\alias{codepipeline_rollback_stage}
\title{Rolls back a stage execution}
\usage{
codepipeline_rollback_stage(pipelineName, stageName, targetPipelineExecutionId)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline for which the stage will be rolled back.}

\item{stageName}{[required] The name of the stage in the pipeline to be rolled back.}

\item{targetPipelineExecutionId}{[required] The pipeline execution ID for the stage to be rolled back to.}
}
\description{
Rolls back a stage execution.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_rollback_stage/} for full documentation.
}
\keyword{internal}
