% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_tag_resource}
\alias{codepipeline_tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
codepipeline_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource you want to add tags to.}

\item{tags}{[required] The tags you want to modify or add to the resource.}
}
\value{
An empty list.
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata
that can be used to manage a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
