% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_update_action_type}
\alias{codepipeline_update_action_type}
\title{Updates an action type that was created with any supported integration
model, where the action type is to be used by customers of the action
type provider}
\usage{
codepipeline_update_action_type(actionType)
}
\arguments{
\item{actionType}{[required] The action type definition for the action type to be updated.}
}
\value{
An empty list.
}
\description{
Updates an action type that was created with any supported integration
model, where the action type is to be used by customers of the action
type provider. Use a JSON file with the action definition and
\code{\link[=codepipeline_update_action_type]{update_action_type}} to provide the
full structure.
}
\section{Request syntax}{
\preformatted{svc$update_action_type(
  actionType = list(
    description = "string",
    executor = list(
      configuration = list(
        lambdaExecutorConfiguration = list(
          lambdaFunctionArn = "string"
        ),
        jobWorkerExecutorConfiguration = list(
          pollingAccounts = list(
            "string"
          ),
          pollingServicePrincipals = list(
            "string"
          )
        )
      ),
      type = "JobWorker"|"Lambda",
      policyStatementsTemplate = "string",
      jobTimeout = 123
    ),
    id = list(
      category = "Source"|"Build"|"Deploy"|"Test"|"Invoke"|"Approval",
      owner = "string",
      provider = "string",
      version = "string"
    ),
    inputArtifactDetails = list(
      minimumCount = 123,
      maximumCount = 123
    ),
    outputArtifactDetails = list(
      minimumCount = 123,
      maximumCount = 123
    ),
    permissions = list(
      allowedAccounts = list(
        "string"
      )
    ),
    properties = list(
      list(
        name = "string",
        optional = TRUE|FALSE,
        key = TRUE|FALSE,
        noEcho = TRUE|FALSE,
        queryable = TRUE|FALSE,
        description = "string"
      )
    ),
    urls = list(
      configurationUrl = "string",
      entityUrlTemplate = "string",
      executionUrlTemplate = "string",
      revisionUrlTemplate = "string"
    )
  )
)
}
}

\keyword{internal}
