% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_update_pipeline}
\alias{codepipeline_update_pipeline}
\title{Updates a specified pipeline with edits or changes to its structure}
\usage{
codepipeline_update_pipeline(pipeline)
}
\arguments{
\item{pipeline}{[required] The name of the pipeline to be updated.}
}
\description{
Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure and \code{\link[=codepipeline_update_pipeline]{update_pipeline}} to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.

See \url{https://paws-r.github.io/docs/codepipeline/update_pipeline.html} for full documentation.
}
\keyword{internal}
