% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_describe_project}
\alias{codestar_describe_project}
\title{Describes a project and its resources}
\usage{
codestar_describe_project(id)
}
\arguments{
\item{id}{[required] The ID of the project.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  name = "string",
  id = "string",
  arn = "string",
  description = "string",
  clientRequestToken = "string",
  createdTimeStamp = as.POSIXct(
    "2015-01-01"
  ),
  stackId = "string",
  projectTemplateId = "string",
  status = list(
    state = "string",
    reason = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a project and its resources.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_project(
  id = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
