% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_list_tags_for_project}
\alias{codestar_list_tags_for_project}
\title{Gets the tags for a project}
\usage{
codestar_list_tags_for_project(id, nextToken, maxResults)
}
\arguments{
\item{id}{[required] The ID of the project to get tags for.}

\item{nextToken}{Reserved for future use.}

\item{maxResults}{Reserved for future use.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the tags for a project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_project(
  id = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
