% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_list_team_members}
\alias{codestar_list_team_members}
\title{Lists all team members associated with a project}
\usage{
codestar_list_team_members(projectId, nextToken, maxResults)
}
\arguments{
\item{projectId}{[required] The ID of the project for which you want to list team members.}

\item{nextToken}{The continuation token for the next set of results, if the results
cannot be returned in one response.}

\item{maxResults}{The maximum number of team members you want returned in a response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  teamMembers = list(
    list(
      userArn = "string",
      projectRole = "string",
      remoteAccessAllowed = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all team members associated with a project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_team_members(
  projectId = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
