% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_list_user_profiles}
\alias{codestar_list_user_profiles}
\title{Lists all the user profiles configured for your AWS account in AWS
CodeStar}
\usage{
codestar_list_user_profiles(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The continuation token for the next set of results, if the results
cannot be returned in one response.}

\item{maxResults}{The maximum number of results to return in a response.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  userProfiles = list(
    list(
      userArn = "string",
      displayName = "string",
      emailAddress = "string",
      sshPublicKey = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the user profiles configured for your AWS account in AWS
CodeStar.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_user_profiles(
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
