% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_tag_project}
\alias{codestar_tag_project}
\title{Adds tags to a project}
\usage{
codestar_tag_project(id, tags)
}
\arguments{
\item{id}{[required] The ID of the project you want to add a tag to.}

\item{tags}{[required] The tags you want to add to the project.}
}
\description{
Adds tags to a project.

See \url{https://paws-r.github.io/docs/codestar/tag_project.html} for full documentation.
}
\keyword{internal}
