% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_untag_project}
\alias{codestar_untag_project}
\title{Removes tags from a project}
\usage{
codestar_untag_project(id, tags)
}
\arguments{
\item{id}{[required] The ID of the project to remove tags from.}

\item{tags}{[required] The tags to remove from the project.}
}
\value{
An empty list.
}
\description{
Removes tags from a project.
}
\section{Request syntax}{
\preformatted{svc$untag_project(
  id = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
