% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_update_project}
\alias{codestar_update_project}
\title{Updates a project in AWS CodeStar}
\usage{
codestar_update_project(id, name, description)
}
\arguments{
\item{id}{[required] The ID of the project you want to update.}

\item{name}{The name of the project you want to update.}

\item{description}{The description of the project, if any.}
}
\value{
An empty list.
}
\description{
Updates a project in AWS CodeStar.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_project(
  id = "string",
  name = "string",
  description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
