% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_delete_connection}
\alias{codestarconnections_delete_connection}
\title{The connection to be deleted}
\usage{
codestarconnections_delete_connection(ConnectionArn)
}
\arguments{
\item{ConnectionArn}{[required] The Amazon Resource Name (ARN) of the connection to be deleted.

The ARN is never reused if the connection is deleted.}
}
\description{
The connection to be deleted.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_delete_connection/} for full documentation.
}
\keyword{internal}
