% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_delete_host}
\alias{codestarconnections_delete_host}
\title{The host to be deleted}
\usage{
codestarconnections_delete_host(HostArn)
}
\arguments{
\item{HostArn}{[required] The Amazon Resource Name (ARN) of the host to be deleted.}
}
\description{
The host to be deleted. Before you delete a host, all connections associated to the host must be deleted.

See \url{https://paws-r.github.io/docs/codestarconnections/delete_host.html} for full documentation.
}
\keyword{internal}
