% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_delete_sync_configuration}
\alias{codestarconnections_delete_sync_configuration}
\title{Deletes the sync configuration for a specified repository and connection}
\usage{
codestarconnections_delete_sync_configuration(SyncType, ResourceName)
}
\arguments{
\item{SyncType}{[required] The type of sync configuration to be deleted.}

\item{ResourceName}{[required] The name of the Amazon Web Services resource associated with the sync
configuration to be deleted.}
}
\description{
Deletes the sync configuration for a specified repository and connection.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_delete_sync_configuration/} for full documentation.
}
\keyword{internal}
