% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_get_sync_configuration}
\alias{codestarconnections_get_sync_configuration}
\title{Returns details about a sync configuration, including the sync type and
resource name}
\usage{
codestarconnections_get_sync_configuration(SyncType, ResourceName)
}
\arguments{
\item{SyncType}{[required] The sync type for the sync configuration for which you want to retrieve
information.}

\item{ResourceName}{[required] The name of the Amazon Web Services resource for the sync configuration
for which you want to retrieve information.}
}
\description{
Returns details about a sync configuration, including the sync type and resource name. A sync configuration allows the configuration to sync (push and pull) changes from the remote repository for a specified branch in a Git repository.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_get_sync_configuration/} for full documentation.
}
\keyword{internal}
