% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_list_repository_links}
\alias{codestarconnections_list_repository_links}
\title{Lists the repository links created for connections in your account}
\usage{
codestarconnections_list_repository_links(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{A non-zero, non-negative integer used to limit the number of returned
results.}

\item{NextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}
}
\description{
Lists the repository links created for connections in your account.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_list_repository_links/} for full documentation.
}
\keyword{internal}
