% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_list_repository_sync_definitions}
\alias{codestarconnections_list_repository_sync_definitions}
\title{Lists the repository sync definitions for repository links in your
account}
\usage{
codestarconnections_list_repository_sync_definitions(
  RepositoryLinkId,
  SyncType
)
}
\arguments{
\item{RepositoryLinkId}{[required] The ID of the repository link for the sync definition for which you want
to retrieve information.}

\item{SyncType}{[required] The sync type of the repository link for the the sync definition for
which you want to retrieve information.}
}
\description{
Lists the repository sync definitions for repository links in your account.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_list_repository_sync_definitions/} for full documentation.
}
\keyword{internal}
