% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_list_sync_configurations}
\alias{codestarconnections_list_sync_configurations}
\title{Returns a list of sync configurations for a specified repository}
\usage{
codestarconnections_list_sync_configurations(
  MaxResults = NULL,
  NextToken = NULL,
  RepositoryLinkId,
  SyncType
)
}
\arguments{
\item{MaxResults}{A non-zero, non-negative integer used to limit the number of returned
results.}

\item{NextToken}{An enumeration token that allows the operation to batch the results of
the operation.}

\item{RepositoryLinkId}{[required] The ID of the repository link for the requested list of sync
configurations.}

\item{SyncType}{[required] The sync type for the requested list of sync configurations.}
}
\description{
Returns a list of sync configurations for a specified repository.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_list_sync_configurations/} for full documentation.
}
\keyword{internal}
