% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_list_tags_for_resource}
\alias{codestarconnections_list_tags_for_resource}
\title{Gets the set of key-value pairs (metadata) that are used to manage the
resource}
\usage{
codestarconnections_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource for which you want to get
information about tags, if any.}
}
\description{
Gets the set of key-value pairs (metadata) that are used to manage the resource.

See \url{https://paws-r.github.io/docs/codestarconnections/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
