% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_untag_resource}
\alias{codestarconnections_untag_resource}
\title{Removes tags from an Amazon Web Services resource}
\usage{
codestarconnections_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove tags from.}

\item{TagKeys}{[required] The list of keys for the tags to be removed from the resource.}
}
\description{
Removes tags from an Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_untag_resource/} for full documentation.
}
\keyword{internal}
