% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_update_repository_link}
\alias{codestarconnections_update_repository_link}
\title{Updates the association between your connection and a specified external
Git repository}
\usage{
codestarconnections_update_repository_link(
  ConnectionArn = NULL,
  EncryptionKeyArn = NULL,
  RepositoryLinkId
)
}
\arguments{
\item{ConnectionArn}{The Amazon Resource Name (ARN) of the connection for the repository link
to be updated. The updated connection ARN must have the same
providerType (such as GitHub) as the original connection ARN for the
repo link.}

\item{EncryptionKeyArn}{The Amazon Resource Name (ARN) of the encryption key for the repository
link to be updated.}

\item{RepositoryLinkId}{[required] The ID of the repository link to be updated.}
}
\description{
Updates the association between your connection and a specified external Git repository. A repository link allows Git sync to monitor and sync changes to files in a specified Git repository.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_update_repository_link/} for full documentation.
}
\keyword{internal}
