% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_delete_notification_rule}
\alias{codestarnotifications_delete_notification_rule}
\title{Deletes a notification rule for a resource}
\usage{
codestarnotifications_delete_notification_rule(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the notification rule you want to
delete.}
}
\description{
Deletes a notification rule for a resource.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_delete_notification_rule/} for full documentation.
}
\keyword{internal}
