% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_describe_notification_rule}
\alias{codestarnotifications_describe_notification_rule}
\title{Returns information about a specified notification rule}
\usage{
codestarnotifications_describe_notification_rule(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the notification rule.}
}
\description{
Returns information about a specified notification rule.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_describe_notification_rule/} for full documentation.
}
\keyword{internal}
