% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_list_event_types}
\alias{codestarnotifications_list_event_types}
\title{Returns information about the event types available for configuring
notifications}
\usage{
codestarnotifications_list_event_types(
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The filters to use to return information by service or resource type.}

\item{NextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}

\item{MaxResults}{A non-negative integer used to limit the number of returned results. The
default number is 50. The maximum number of results that can be returned
is 100.}
}
\description{
Returns information about the event types available for configuring notifications.

See \url{https://paws-r.github.io/docs/codestarnotifications/list_event_types.html} for full documentation.
}
\keyword{internal}
