% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_list_notification_rules}
\alias{codestarnotifications_list_notification_rules}
\title{Returns a list of the notification rules for an Amazon Web Services
account}
\usage{
codestarnotifications_list_notification_rules(
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The filters to use to return information by service or resource type.
For valid values, see ListNotificationRulesFilter.

A filter with the same name can appear more than once when used with OR
statements. Filters with different names should be applied with AND
statements.}

\item{NextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}

\item{MaxResults}{A non-negative integer used to limit the number of returned results. The
maximum number of results that can be returned is 100.}
}
\description{
Returns a list of the notification rules for an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_list_notification_rules/} for full documentation.
}
\keyword{internal}
