% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_list_tags_for_resource}
\alias{codestarnotifications_list_tags_for_resource}
\title{Returns a list of the tags associated with a notification rule}
\usage{
codestarnotifications_list_tags_for_resource(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) for the notification rule.}
}
\description{
Returns a list of the tags associated with a notification rule.

See \url{https://paws-r.github.io/docs/codestarnotifications/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
