% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_list_targets}
\alias{codestarnotifications_list_targets}
\title{Returns a list of the notification rule targets for an AWS account}
\usage{
codestarnotifications_list_targets(
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The filters to use to return information by service or resource type.
Valid filters include target type, target address, and target status.

A filter with the same name can appear more than once when used with OR
statements. Filters with different names should be applied with AND
statements.}

\item{NextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}

\item{MaxResults}{A non-negative integer used to limit the number of returned results. The
maximum number of results that can be returned is 100.}
}
\description{
Returns a list of the notification rule targets for an AWS account.

See \url{https://paws-r.github.io/docs/codestarnotifications/list_targets.html} for full documentation.
}
\keyword{internal}
