% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_subscribe}
\alias{codestarnotifications_subscribe}
\title{Creates an association between a notification rule and an Amazon Q
Developer in chat applications topic or Amazon Q Developer in chat
applications client so that the associated target can receive
notifications when the events described in the rule are triggered}
\usage{
codestarnotifications_subscribe(Arn, Target, ClientRequestToken = NULL)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the notification rule for which you
want to create the association.}

\item{Target}{[required]}

\item{ClientRequestToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}
}
\description{
Creates an association between a notification rule and an Amazon Q Developer in chat applications topic or Amazon Q Developer in chat applications client so that the associated target can receive notifications when the events described in the rule are triggered.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_subscribe/} for full documentation.
}
\keyword{internal}
