% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_tag_resource}
\alias{codestarnotifications_tag_resource}
\title{Associates a set of provided tags with a notification rule}
\usage{
codestarnotifications_tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the notification rule to tag.}

\item{Tags}{[required] The list of tags to associate with the resource. Tag key names cannot
start with "aws".}
}
\description{
Associates a set of provided tags with a notification rule.

See \url{https://paws-r.github.io/docs/codestarnotifications/tag_resource.html} for full documentation.
}
\keyword{internal}
