% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_untag_resource}
\alias{codestarnotifications_untag_resource}
\title{Removes the association between one or more provided tags and a
notification rule}
\usage{
codestarnotifications_untag_resource(Arn, TagKeys)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the notification rule from which to
remove the tags.}

\item{TagKeys}{[required] The key names of the tags to remove.}
}
\description{
Removes the association between one or more provided tags and a notification rule.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_untag_resource/} for full documentation.
}
\keyword{internal}
