% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_add_notification_channel}
\alias{devopsguru_add_notification_channel}
\title{Adds a notification channel to DevOps Guru}
\usage{
devopsguru_add_notification_channel(Config)
}
\arguments{
\item{Config}{[required] A \code{NotificationChannelConfig} object that specifies what type of
notification channel to add. The one supported notification channel is
Amazon Simple Notification Service (Amazon SNS).}
}
\description{
Adds a notification channel to DevOps Guru. A notification channel is used to notify you about important DevOps Guru events, such as when an insight is generated.

See \url{https://paws-r.github.io/docs/devopsguru/add_notification_channel.html} for full documentation.
}
\keyword{internal}
