% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_delete_insight}
\alias{devopsguru_delete_insight}
\title{Deletes the insight along with the associated anomalies, events and
recommendations}
\usage{
devopsguru_delete_insight(Id)
}
\arguments{
\item{Id}{[required] The ID of the insight.}
}
\description{
Deletes the insight along with the associated anomalies, events and recommendations.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_delete_insight/} for full documentation.
}
\keyword{internal}
