% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_event_sources_config}
\alias{devopsguru_describe_event_sources_config}
\title{Returns the integration status of services that are integrated with
DevOps Guru as Consumer via EventBridge}
\usage{
devopsguru_describe_event_sources_config()
}
\description{
Returns the integration status of services that are integrated with DevOps Guru as Consumer via EventBridge. The one service that can be integrated with DevOps Guru is Amazon CodeGuru Profiler, which can produce proactive recommendations which can be stored and viewed in DevOps Guru.

See \url{https://paws-r.github.io/docs/devopsguru/describe_event_sources_config.html} for full documentation.
}
\keyword{internal}
