% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_feedback}
\alias{devopsguru_describe_feedback}
\title{Returns the most recent feedback submitted in the current Amazon Web
Services account and Region}
\usage{
devopsguru_describe_feedback(InsightId = NULL)
}
\arguments{
\item{InsightId}{The ID of the insight for which the feedback was provided.}
}
\description{
Returns the most recent feedback submitted in the current Amazon Web Services account and Region.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_describe_feedback/} for full documentation.
}
\keyword{internal}
