% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_insight}
\alias{devopsguru_describe_insight}
\title{Returns details about an insight that you specify using its ID}
\usage{
devopsguru_describe_insight(Id, AccountId = NULL)
}
\arguments{
\item{Id}{[required] The ID of the insight.}

\item{AccountId}{The ID of the member account in the organization.}
}
\description{
Returns details about an insight that you specify using its ID.

See \url{https://paws-r.github.io/docs/devopsguru/describe_insight.html} for full documentation.
}
\keyword{internal}
