% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_organization_health}
\alias{devopsguru_describe_organization_health}
\title{Returns active insights, predictive insights, and resource hours
analyzed in last hour}
\usage{
devopsguru_describe_organization_health(
  AccountIds = NULL,
  OrganizationalUnitIds = NULL
)
}
\arguments{
\item{AccountIds}{The ID of the Amazon Web Services account.}

\item{OrganizationalUnitIds}{The ID of the organizational unit.}
}
\description{
Returns active insights, predictive insights, and resource hours analyzed in last hour.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_describe_organization_health/} for full documentation.
}
\keyword{internal}
