% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_describe_organization_overview}
\alias{devopsguru_describe_organization_overview}
\title{Returns an overview of your organization's history based on the
specified time range}
\usage{
devopsguru_describe_organization_overview(
  FromTime,
  ToTime = NULL,
  AccountIds = NULL,
  OrganizationalUnitIds = NULL
)
}
\arguments{
\item{FromTime}{[required] The start of the time range passed in. The start time granularity is at
the day level. The floor of the start time is used. Returned information
occurred after this day.}

\item{ToTime}{The end of the time range passed in. The start time granularity is at
the day level. The floor of the start time is used. Returned information
occurred before this day. If this is not specified, then the current day
is used.}

\item{AccountIds}{The ID of the Amazon Web Services account.}

\item{OrganizationalUnitIds}{The ID of the organizational unit.}
}
\description{
Returns an overview of your organization's history based on the specified time range. The overview includes the total reactive and proactive insights.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_describe_organization_overview/} for full documentation.
}
\keyword{internal}
