% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_anomalous_log_groups}
\alias{devopsguru_list_anomalous_log_groups}
\title{Returns the list of log groups that contain log anomalies}
\usage{
devopsguru_list_anomalous_log_groups(
  InsightId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InsightId}{[required] The ID of the insight containing the log groups.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns the list of log groups that contain log anomalies.

See \url{https://paws-r.github.io/docs/devopsguru/list_anomalous_log_groups.html} for full documentation.
}
\keyword{internal}
