% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_events}
\alias{devopsguru_list_events}
\title{Returns a list of the events emitted by the resources that are evaluated
by DevOps Guru}
\usage{
devopsguru_list_events(
  Filters,
  MaxResults = NULL,
  NextToken = NULL,
  AccountId = NULL
)
}
\arguments{
\item{Filters}{[required] A \code{ListEventsFilters} object used to specify which events to return.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}

\item{AccountId}{The ID of the Amazon Web Services account.}
}
\description{
Returns a list of the events emitted by the resources that are evaluated by DevOps Guru. You can use filters to specify which events are returned.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_list_events/} for full documentation.
}
\keyword{internal}
