% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_notification_channels}
\alias{devopsguru_list_notification_channels}
\title{Returns a list of notification channels configured for DevOps Guru}
\usage{
devopsguru_list_notification_channels(NextToken = NULL)
}
\arguments{
\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns a list of notification channels configured for DevOps Guru. Each notification channel is used to notify you when DevOps Guru generates an insight that contains information about how to improve your operations. The one supported notification channel is Amazon Simple Notification Service (Amazon SNS).

See \url{https://paws-r.github.io/docs/devopsguru/list_notification_channels.html} for full documentation.
}
\keyword{internal}
