% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_organization_insights}
\alias{devopsguru_list_organization_insights}
\title{Returns a list of insights associated with the account or OU Id}
\usage{
devopsguru_list_organization_insights(
  StatusFilter,
  MaxResults = NULL,
  AccountIds = NULL,
  OrganizationalUnitIds = NULL,
  NextToken = NULL
)
}
\arguments{
\item{StatusFilter}{[required]}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{AccountIds}{The ID of the Amazon Web Services account.}

\item{OrganizationalUnitIds}{The ID of the organizational unit.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Returns a list of insights associated with the account or OU Id.

See \url{https://paws-r.github.io/docs/devopsguru/list_organization_insights.html} for full documentation.
}
\keyword{internal}
