% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_list_recommendations}
\alias{devopsguru_list_recommendations}
\title{Returns a list of a specified insight's recommendations}
\usage{
devopsguru_list_recommendations(
  InsightId,
  NextToken = NULL,
  Locale = NULL,
  AccountId = NULL
)
}
\arguments{
\item{InsightId}{[required] The ID of the requested insight.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}

\item{Locale}{A locale that specifies the language to use for recommendations.}

\item{AccountId}{The ID of the Amazon Web Services account.}
}
\description{
Returns a list of a specified insight's recommendations. Each recommendation includes a list of related metrics and a list of related events.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_list_recommendations/} for full documentation.
}
\keyword{internal}
