% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_put_feedback}
\alias{devopsguru_put_feedback}
\title{Collects customer feedback about the specified insight}
\usage{
devopsguru_put_feedback(InsightFeedback = NULL)
}
\arguments{
\item{InsightFeedback}{The feedback from customers is about the recommendations in this
insight.}
}
\description{
Collects customer feedback about the specified insight.

See \url{https://paws-r.github.io/docs/devopsguru/put_feedback.html} for full documentation.
}
\keyword{internal}
