% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_remove_notification_channel}
\alias{devopsguru_remove_notification_channel}
\title{Removes a notification channel from DevOps Guru}
\usage{
devopsguru_remove_notification_channel(Id)
}
\arguments{
\item{Id}{[required] The ID of the notification channel to be removed.}
}
\description{
Removes a notification channel from DevOps Guru. A notification channel is used to notify you when DevOps Guru generates an insight that contains information about how to improve your operations.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_remove_notification_channel/} for full documentation.
}
\keyword{internal}
