% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_search_organization_insights}
\alias{devopsguru_search_organization_insights}
\title{Returns a list of insights in your organization}
\usage{
devopsguru_search_organization_insights(
  AccountIds,
  StartTimeRange,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Type
)
}
\arguments{
\item{AccountIds}{[required] The ID of the Amazon Web Services account.}

\item{StartTimeRange}{[required]}

\item{Filters}{A \code{SearchOrganizationInsightsFilters} object that is used to set the
severity and status filters on your insight search.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}

\item{Type}{[required] The type of insights you are searching for (\code{REACTIVE} or \code{PROACTIVE}).}
}
\description{
Returns a list of insights in your organization. You can specify which insights are returned by their start time, one or more statuses (\code{ONGOING}, \code{CLOSED}, and \code{CLOSED}), one or more severities (\code{LOW}, \code{MEDIUM}, and \code{HIGH}), and type (\code{REACTIVE} or \code{PROACTIVE}).

See \url{https://www.paws-r-sdk.com/docs/devopsguru_search_organization_insights/} for full documentation.
}
\keyword{internal}
