% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_start_cost_estimation}
\alias{devopsguru_start_cost_estimation}
\title{Starts the creation of an estimate of the monthly cost to analyze your
Amazon Web Services resources}
\usage{
devopsguru_start_cost_estimation(ResourceCollection, ClientToken = NULL)
}
\arguments{
\item{ResourceCollection}{[required] The collection of Amazon Web Services resources used to create a monthly
DevOps Guru cost estimate.}

\item{ClientToken}{The idempotency token used to identify each cost estimate request.}
}
\description{
Starts the creation of an estimate of the monthly cost to analyze your Amazon Web Services resources.

See \url{https://paws-r.github.io/docs/devopsguru/start_cost_estimation.html} for full documentation.
}
\keyword{internal}
