% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_update_event_sources_config}
\alias{devopsguru_update_event_sources_config}
\title{Enables or disables integration with a service that can be integrated
with DevOps Guru}
\usage{
devopsguru_update_event_sources_config(EventSources = NULL)
}
\arguments{
\item{EventSources}{Configuration information about the integration of DevOps Guru as the
Consumer via EventBridge with another AWS Service.}
}
\description{
Enables or disables integration with a service that can be integrated with DevOps Guru. The one service that can be integrated with DevOps Guru is Amazon CodeGuru Profiler, which can produce proactive recommendations which can be stored and viewed in DevOps Guru.

See \url{https://paws-r.github.io/docs/devopsguru/update_event_sources_config.html} for full documentation.
}
\keyword{internal}
