% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_update_resource_collection}
\alias{devopsguru_update_resource_collection}
\title{Updates the collection of resources that DevOps Guru analyzes}
\usage{
devopsguru_update_resource_collection(Action, ResourceCollection)
}
\arguments{
\item{Action}{[required] Specifies if the resource collection in the request is added or deleted
to the resource collection.}

\item{ResourceCollection}{[required]}
}
\description{
Updates the collection of resources that DevOps Guru analyzes. The two types of Amazon Web Services resource collections supported are Amazon Web Services CloudFormation stacks and Amazon Web Services resources that contain the same Amazon Web Services tag. DevOps Guru can be configured to analyze the Amazon Web Services resources that are defined in the stacks or that are tagged using the same tag \emph{key}. You can specify up to 500 Amazon Web Services CloudFormation stacks. This method also creates the IAM role required for you to use DevOps Guru.

See \url{https://www.paws-r-sdk.com/docs/devopsguru_update_resource_collection/} for full documentation.
}
\keyword{internal}
