% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devopsguru_operations.R
\name{devopsguru_update_service_integration}
\alias{devopsguru_update_service_integration}
\title{Enables or disables integration with a service that can be integrated
with DevOps Guru}
\usage{
devopsguru_update_service_integration(ServiceIntegration)
}
\arguments{
\item{ServiceIntegration}{[required] An \code{IntegratedServiceConfig} object used to specify the integrated
service you want to update, and whether you want to update it to enabled
or disabled.}
}
\description{
Enables or disables integration with a service that can be integrated with DevOps Guru. The one service that can be integrated with DevOps Guru is Amazon Web Services Systems Manager, which can be used to create an OpsItem for each generated insight.

See \url{https://paws-r.github.io/docs/devopsguru/update_service_integration.html} for full documentation.
}
\keyword{internal}
