% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_associate_source_network_stack}
\alias{drs_associate_source_network_stack}
\title{Associate a Source Network to an existing CloudFormation Stack and
modify launch templates to use this network}
\usage{
drs_associate_source_network_stack(cfnStackName, sourceNetworkID)
}
\arguments{
\item{cfnStackName}{[required] CloudFormation template to associate with a Source Network.}

\item{sourceNetworkID}{[required] The Source Network ID to associate with CloudFormation template.}
}
\description{
Associate a Source Network to an existing CloudFormation Stack and modify launch templates to use this network. Can be used for reverting to previously deployed CloudFormation stacks.

See \url{https://www.paws-r-sdk.com/docs/drs_associate_source_network_stack/} for full documentation.
}
\keyword{internal}
