% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_cancel_recovery_plan_execution}
\alias{drs_cancel_recovery_plan_execution}
\title{Cancels an in-progress Recovery Plan execution}
\usage{
drs_cancel_recovery_plan_execution(recoveryPlanExecutionArn)
}
\arguments{
\item{recoveryPlanExecutionArn}{[required] The ARN of the Recovery Plan execution to cancel.}
}
\description{
Cancels an in-progress Recovery Plan execution. Remaining steps are skipped.

See \url{https://www.paws-r-sdk.com/docs/drs_cancel_recovery_plan_execution/} for full documentation.
}
\keyword{internal}
