% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_create_recovery_plan}
\alias{drs_create_recovery_plan}
\title{Creates a Recovery Plan to orchestrate multi-server disaster recovery}
\usage{
drs_create_recovery_plan(
  name,
  description = NULL,
  clientToken = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of a Recovery Plan.}

\item{description}{The description of a Recovery Plan.}

\item{clientToken}{A unique string provided to ensure request idempotency.}

\item{tags}{The tags to apply to the Recovery Plan.}
}
\description{
Creates a Recovery Plan to orchestrate multi-server disaster recovery.

See \url{https://www.paws-r-sdk.com/docs/drs_create_recovery_plan/} for full documentation.
}
\keyword{internal}
