% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_create_recovery_plan_step}
\alias{drs_create_recovery_plan_step}
\title{Creates a step in a Recovery Plan}
\usage{
drs_create_recovery_plan_step(
  recoveryPlanArn,
  stepName,
  stepOrder = NULL,
  configuration,
  clientToken = NULL
)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan to add the step to.}

\item{stepName}{[required] The name of a Recovery Plan Step.}

\item{stepOrder}{The order of a step within a Recovery Plan (1-based).}

\item{configuration}{[required] Type-specific configuration for a recovery plan step. Exactly one member must be set.}

\item{clientToken}{A unique string provided to ensure request idempotency.}
}
\description{
Creates a step in a Recovery Plan. A step is either \code{SERVER} type (servers to recover in parallel) or \code{WAIT} type (timed pause between steps).

See \url{https://www.paws-r-sdk.com/docs/drs_create_recovery_plan_step/} for full documentation.
}
\keyword{internal}
