% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_create_source_network}
\alias{drs_create_source_network}
\title{Create a new Source Network resource for a provided VPC ID}
\usage{
drs_create_source_network(originAccountID, originRegion, tags = NULL, vpcID)
}
\arguments{
\item{originAccountID}{[required] Account containing the VPC to protect.}

\item{originRegion}{[required] Region containing the VPC to protect.}

\item{tags}{A set of tags to be associated with the Source Network resource.}

\item{vpcID}{[required] Which VPC ID to protect.}
}
\description{
Create a new Source Network resource for a provided VPC ID.

See \url{https://www.paws-r-sdk.com/docs/drs_create_source_network/} for full documentation.
}
\keyword{internal}
