% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_job}
\alias{drs_delete_job}
\title{Deletes a single Job by ID}
\usage{
drs_delete_job(jobID)
}
\arguments{
\item{jobID}{[required] The ID of the Job to be deleted.}
}
\description{
Deletes a single Job by ID.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_job/} for full documentation.
}
\keyword{internal}
