% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_launch_action}
\alias{drs_delete_launch_action}
\title{Deletes a resource launch action}
\usage{
drs_delete_launch_action(actionId, resourceId)
}
\arguments{
\item{actionId}{[required]}

\item{resourceId}{[required]}
}
\description{
Deletes a resource launch action.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_launch_action/} for full documentation.
}
\keyword{internal}
