% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_recovery_instance}
\alias{drs_delete_recovery_instance}
\title{Deletes a single Recovery Instance by ID}
\usage{
drs_delete_recovery_instance(recoveryInstanceID)
}
\arguments{
\item{recoveryInstanceID}{[required] The ID of the Recovery Instance to be deleted.}
}
\description{
Deletes a single Recovery Instance by ID. This deletes the Recovery Instance resource from Elastic Disaster Recovery. The Recovery Instance must be disconnected first in order to delete it.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_recovery_instance/} for full documentation.
}
\keyword{internal}
