% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_recovery_plan}
\alias{drs_delete_recovery_plan}
\title{Deletes a Recovery Plan}
\usage{
drs_delete_recovery_plan(recoveryPlanArn)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan to delete.}
}
\description{
Deletes a Recovery Plan. Cannot delete a plan that has an execution in a non-terminal status (\code{CREATED}, \code{IN_PROGRESS}).

See \url{https://www.paws-r-sdk.com/docs/drs_delete_recovery_plan/} for full documentation.
}
\keyword{internal}
