% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_recovery_plan_execution}
\alias{drs_delete_recovery_plan_execution}
\title{Deletes a Recovery Plan execution record}
\usage{
drs_delete_recovery_plan_execution(recoveryPlanExecutionArn)
}
\arguments{
\item{recoveryPlanExecutionArn}{[required] The ARN of the Recovery Plan execution to delete.}
}
\description{
Deletes a Recovery Plan execution record. Must be in a terminal status.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_recovery_plan_execution/} for full documentation.
}
\keyword{internal}
