% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_recovery_plan_step}
\alias{drs_delete_recovery_plan_step}
\title{Deletes a step from a Recovery Plan}
\usage{
drs_delete_recovery_plan_step(recoveryPlanStepArn)
}
\arguments{
\item{recoveryPlanStepArn}{[required] The ARN of the Recovery Plan step to delete.}
}
\description{
Deletes a step from a Recovery Plan.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_recovery_plan_step/} for full documentation.
}
\keyword{internal}
