% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_source_network}
\alias{drs_delete_source_network}
\title{Delete Source Network resource}
\usage{
drs_delete_source_network(sourceNetworkID)
}
\arguments{
\item{sourceNetworkID}{[required] ID of the Source Network to delete.}
}
\description{
Delete Source Network resource.

See \url{https://www.paws-r-sdk.com/docs/drs_delete_source_network/} for full documentation.
}
\keyword{internal}
