% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_delete_source_server}
\alias{drs_delete_source_server}
\title{Deletes a single Source Server by ID}
\usage{
drs_delete_source_server(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Server to be deleted.}
}
\description{
Deletes a single Source Server by ID. The Source Server must be disconnected first.

See \url{https://paws-r.github.io/docs/drs/delete_source_server.html} for full documentation.
}
\keyword{internal}
