% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_describe_jobs}
\alias{drs_describe_jobs}
\title{Returns a list of Jobs}
\usage{
drs_describe_jobs(filters = NULL, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{filters}{A set of filters by which to return Jobs.}

\item{maxResults}{Maximum number of Jobs to retrieve.}

\item{nextToken}{The token of the next Job to retrieve.}
}
\description{
Returns a list of Jobs. Use the JobsID and fromDate and toDate filters to limit which jobs are returned. The response is sorted by creationDataTime - latest date first. Jobs are created by the StartRecovery, TerminateRecoveryInstances and StartFailbackLaunch APIs. Jobs are also created by DiagnosticLaunch and TerminateDiagnosticInstances, which are APIs available only to \emph{Support} and only used in response to relevant support tickets.

See \url{https://www.paws-r-sdk.com/docs/drs_describe_jobs/} for full documentation.
}
\keyword{internal}
