% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_describe_recovery_instances}
\alias{drs_describe_recovery_instances}
\title{Lists all Recovery Instances or multiple Recovery Instances by ID}
\usage{
drs_describe_recovery_instances(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{A set of filters by which to return Recovery Instances.}

\item{maxResults}{Maximum number of Recovery Instances to retrieve.}

\item{nextToken}{The token of the next Recovery Instance to retrieve.}
}
\description{
Lists all Recovery Instances or multiple Recovery Instances by ID.

See \url{https://www.paws-r-sdk.com/docs/drs_describe_recovery_instances/} for full documentation.
}
\keyword{internal}
