% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_describe_replication_configuration_templates}
\alias{drs_describe_replication_configuration_templates}
\title{Lists all ReplicationConfigurationTemplates, filtered by Source Server
IDs}
\usage{
drs_describe_replication_configuration_templates(
  maxResults = NULL,
  nextToken = NULL,
  replicationConfigurationTemplateIDs = NULL
)
}
\arguments{
\item{maxResults}{Maximum number of Replication Configuration Templates to retrieve.}

\item{nextToken}{The token of the next Replication Configuration Template to retrieve.}

\item{replicationConfigurationTemplateIDs}{The IDs of the Replication Configuration Templates to retrieve. An empty
list means all Replication Configuration Templates.}
}
\description{
Lists all ReplicationConfigurationTemplates, filtered by Source Server IDs.

See \url{https://www.paws-r-sdk.com/docs/drs_describe_replication_configuration_templates/} for full documentation.
}
\keyword{internal}
