% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_describe_source_servers}
\alias{drs_describe_source_servers}
\title{Lists all Source Servers or multiple Source Servers filtered by ID}
\usage{
drs_describe_source_servers(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{A set of filters by which to return Source Servers.}

\item{maxResults}{Maximum number of Source Servers to retrieve.}

\item{nextToken}{The token of the next Source Server to retrieve.}
}
\description{
Lists all Source Servers or multiple Source Servers filtered by ID.

See \url{https://www.paws-r-sdk.com/docs/drs_describe_source_servers/} for full documentation.
}
\keyword{internal}
