% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_export_source_network_cfn_template}
\alias{drs_export_source_network_cfn_template}
\title{Export the Source Network CloudFormation template to an S3 bucket}
\usage{
drs_export_source_network_cfn_template(sourceNetworkID)
}
\arguments{
\item{sourceNetworkID}{[required] The Source Network ID to export its CloudFormation template to an S3
bucket.}
}
\description{
Export the Source Network CloudFormation template to an S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/drs_export_source_network_cfn_template/} for full documentation.
}
\keyword{internal}
