% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_failback_replication_configuration}
\alias{drs_get_failback_replication_configuration}
\title{Lists all Failback ReplicationConfigurations, filtered by Recovery
Instance ID}
\usage{
drs_get_failback_replication_configuration(recoveryInstanceID)
}
\arguments{
\item{recoveryInstanceID}{[required] The ID of the Recovery Instance whose failback replication configuration
should be returned.}
}
\description{
Lists all Failback ReplicationConfigurations, filtered by Recovery Instance ID.

See \url{https://paws-r.github.io/docs/drs/get_failback_replication_configuration.html} for full documentation.
}
\keyword{internal}
