% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_launch_configuration}
\alias{drs_get_launch_configuration}
\title{Gets a LaunchConfiguration, filtered by Source Server IDs}
\usage{
drs_get_launch_configuration(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Server that we want to retrieve a Launch
Configuration for.}
}
\description{
Gets a LaunchConfiguration, filtered by Source Server IDs.

See \url{https://paws-r.github.io/docs/drs/get_launch_configuration.html} for full documentation.
}
\keyword{internal}
