% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_recovery_plan}
\alias{drs_get_recovery_plan}
\title{Gets a Recovery Plan by ARN}
\usage{
drs_get_recovery_plan(recoveryPlanArn)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan to retrieve.}
}
\description{
Gets a Recovery Plan by ARN.

See \url{https://www.paws-r-sdk.com/docs/drs_get_recovery_plan/} for full documentation.
}
\keyword{internal}
