% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_recovery_plan_execution}
\alias{drs_get_recovery_plan_execution}
\title{Gets the details of a Recovery Plan execution}
\usage{
drs_get_recovery_plan_execution(recoveryPlanExecutionArn)
}
\arguments{
\item{recoveryPlanExecutionArn}{[required] The ARN of the Recovery Plan execution.}
}
\description{
Gets the details of a Recovery Plan execution.

See \url{https://www.paws-r-sdk.com/docs/drs_get_recovery_plan_execution/} for full documentation.
}
\keyword{internal}
