% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_recovery_plan_execution_step}
\alias{drs_get_recovery_plan_execution_step}
\title{Gets the details of a step within a Recovery Plan execution}
\usage{
drs_get_recovery_plan_execution_step(recoveryPlanExecutionStepArn)
}
\arguments{
\item{recoveryPlanExecutionStepArn}{[required] The ARN of the execution step.}
}
\description{
Gets the details of a step within a Recovery Plan execution.

See \url{https://www.paws-r-sdk.com/docs/drs_get_recovery_plan_execution_step/} for full documentation.
}
\keyword{internal}
