% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_recovery_plan_step}
\alias{drs_get_recovery_plan_step}
\title{Gets a Recovery Plan step by ARN}
\usage{
drs_get_recovery_plan_step(recoveryPlanStepArn)
}
\arguments{
\item{recoveryPlanStepArn}{[required] The ARN of the Recovery Plan step to retrieve.}
}
\description{
Gets a Recovery Plan step by ARN.

See \url{https://www.paws-r-sdk.com/docs/drs_get_recovery_plan_step/} for full documentation.
}
\keyword{internal}
