% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_get_replication_configuration}
\alias{drs_get_replication_configuration}
\title{Gets a ReplicationConfiguration, filtered by Source Server ID}
\usage{
drs_get_replication_configuration(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Serve for this Replication Configuration.r}
}
\description{
Gets a ReplicationConfiguration, filtered by Source Server ID.

See \url{https://www.paws-r-sdk.com/docs/drs_get_replication_configuration/} for full documentation.
}
\keyword{internal}
