% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_recovery_plan_execution_steps}
\alias{drs_list_recovery_plan_execution_steps}
\title{Lists all steps within a Recovery Plan execution}
\usage{
drs_list_recovery_plan_execution_steps(
  recoveryPlanExecutionArn,
  filter = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{recoveryPlanExecutionArn}{[required] The ARN of the Recovery Plan execution.}

\item{filter}{Filters for listing execution steps.}

\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists all steps within a Recovery Plan execution.

See \url{https://www.paws-r-sdk.com/docs/drs_list_recovery_plan_execution_steps/} for full documentation.
}
\keyword{internal}
