% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_recovery_plan_executions}
\alias{drs_list_recovery_plan_executions}
\title{Lists executions of Recovery Plans, optionally filtered by plan or
status}
\usage{
drs_list_recovery_plan_executions(
  recoveryPlanArn = NULL,
  status = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{recoveryPlanArn}{Filter executions by Recovery Plan ARN.}

\item{status}{Filter executions by status.}

\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists executions of Recovery Plans, optionally filtered by plan or status.

See \url{https://www.paws-r-sdk.com/docs/drs_list_recovery_plan_executions/} for full documentation.
}
\keyword{internal}
