% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_recovery_plan_steps}
\alias{drs_list_recovery_plan_steps}
\title{Lists all steps in a Recovery Plan}
\usage{
drs_list_recovery_plan_steps(
  recoveryPlanArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan.}

\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists all steps in a Recovery Plan.

See \url{https://www.paws-r-sdk.com/docs/drs_list_recovery_plan_steps/} for full documentation.
}
\keyword{internal}
