% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_recovery_plans}
\alias{drs_list_recovery_plans}
\title{Lists all Recovery Plans in the account}
\usage{
drs_list_recovery_plans(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of results to return.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists all Recovery Plans in the account.

See \url{https://www.paws-r-sdk.com/docs/drs_list_recovery_plans/} for full documentation.
}
\keyword{internal}
