% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_staging_accounts}
\alias{drs_list_staging_accounts}
\title{Returns an array of staging accounts for existing extended source
servers}
\usage{
drs_list_staging_accounts(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of staging Accounts to retrieve.}

\item{nextToken}{The token of the next staging Account to retrieve.}
}
\description{
Returns an array of staging accounts for existing extended source servers.

See \url{https://paws-r.github.io/docs/drs/list_staging_accounts.html} for full documentation.
}
\keyword{internal}
