% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_reorder_recovery_plan_steps}
\alias{drs_reorder_recovery_plan_steps}
\title{Reorders steps in a Recovery Plan}
\usage{
drs_reorder_recovery_plan_steps(recoveryPlanArn, orderedStepArns)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan.}

\item{orderedStepArns}{[required] Ordered list of all step ARNs representing the desired sequence.}
}
\description{
Reorders steps in a Recovery Plan. Accepts a complete ordered list of step ARNs.

See \url{https://www.paws-r-sdk.com/docs/drs_reorder_recovery_plan_steps/} for full documentation.
}
\keyword{internal}
