% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_retry_recovery_plan_execution_step}
\alias{drs_retry_recovery_plan_execution_step}
\title{Retries a failed SERVER type execution step}
\usage{
drs_retry_recovery_plan_execution_step(recoveryPlanExecutionStepArn)
}
\arguments{
\item{recoveryPlanExecutionStepArn}{[required] The ARN of the execution step to retry.}
}
\description{
Retries a failed \code{SERVER} type execution step.

See \url{https://www.paws-r-sdk.com/docs/drs_retry_recovery_plan_execution_step/} for full documentation.
}
\keyword{internal}
