% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_reverse_replication}
\alias{drs_reverse_replication}
\title{Start replication to origin / target region - applies only to protected
instances that originated in EC2}
\usage{
drs_reverse_replication(recoveryInstanceID)
}
\arguments{
\item{recoveryInstanceID}{[required] The ID of the Recovery Instance that we want to reverse the replication
for.}
}
\description{
Start replication to origin / target region - applies only to protected instances that originated in EC2. For recovery instances on target region - starts replication back to origin region. For failback instances on origin region - starts replication to target region to re-protect them.

See \url{https://www.paws-r-sdk.com/docs/drs_reverse_replication/} for full documentation.
}
\keyword{internal}
