% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_failback_launch}
\alias{drs_start_failback_launch}
\title{Initiates a Job for launching the machine that is being failed back to
from the specified Recovery Instance}
\usage{
drs_start_failback_launch(recoveryInstanceIDs, tags = NULL)
}
\arguments{
\item{recoveryInstanceIDs}{[required] The IDs of the Recovery Instance whose failback launch we want to
request.}

\item{tags}{The tags to be associated with the failback launch Job.}
}
\description{
Initiates a Job for launching the machine that is being failed back to from the specified Recovery Instance. This will run conversion on the failback client and will reboot your machine, thus completing the failback process.

See \url{https://paws-r.github.io/docs/drs/start_failback_launch.html} for full documentation.
}
\keyword{internal}
