% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_recovery}
\alias{drs_start_recovery}
\title{Launches Recovery Instances for the specified Source Servers}
\usage{
drs_start_recovery(isDrill = NULL, sourceServers, tags = NULL)
}
\arguments{
\item{isDrill}{Whether this Source Server Recovery operation is a drill or not.}

\item{sourceServers}{[required] The Source Servers that we want to start a Recovery Job for.}

\item{tags}{The tags to be associated with the Recovery Job.}
}
\description{
Launches Recovery Instances for the specified Source Servers. For each Source Server you may choose a point in time snapshot to launch from, or use an on demand snapshot.

See \url{https://www.paws-r-sdk.com/docs/drs_start_recovery/} for full documentation.
}
\keyword{internal}
