% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_recovery_plan_execution}
\alias{drs_start_recovery_plan_execution}
\title{Starts executing a Recovery Plan in DRILL or RECOVERY mode}
\usage{
drs_start_recovery_plan_execution(
  recoveryPlanArn,
  mode,
  clientToken = NULL,
  sourceServers = NULL,
  tags = NULL
)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan to execute.}

\item{mode}{[required] The execution mode (\code{DRILL} or \code{RECOVERY}).}

\item{clientToken}{A unique string provided to ensure request idempotency.}

\item{sourceServers}{Optional list of source servers with specific recovery snapshots. If not provided, the latest snapshot is used for each server.}

\item{tags}{The tags to apply to the Recovery Plan execution.}
}
\description{
Starts executing a Recovery Plan in \code{DRILL} or \code{RECOVERY} mode. A plan cannot have more than one execution in a non-terminal status at a time.

See \url{https://www.paws-r-sdk.com/docs/drs_start_recovery_plan_execution/} for full documentation.
}
\keyword{internal}
