% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_replication}
\alias{drs_start_replication}
\title{Starts replication for a stopped Source Server}
\usage{
drs_start_replication(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Server to start replication for.}
}
\description{
Starts replication for a stopped Source Server. This action would make the Source Server protected again and restart billing for it.

See \url{https://www.paws-r-sdk.com/docs/drs_start_replication/} for full documentation.
}
\keyword{internal}
