% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_source_network_recovery}
\alias{drs_start_source_network_recovery}
\title{Deploy VPC for the specified Source Network and modify launch templates
to use this network}
\usage{
drs_start_source_network_recovery(
  deployAsNew = NULL,
  sourceNetworks,
  tags = NULL
)
}
\arguments{
\item{deployAsNew}{Don't update existing CloudFormation Stack, recover the network using a
new stack.}

\item{sourceNetworks}{[required] The Source Networks that we want to start a Recovery Job for.}

\item{tags}{The tags to be associated with the Source Network recovery Job.}
}
\description{
Deploy VPC for the specified Source Network and modify launch templates to use this network. The VPC will be deployed using a dedicated CloudFormation stack.

See \url{https://www.paws-r-sdk.com/docs/drs_start_source_network_recovery/} for full documentation.
}
\keyword{internal}
