% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_start_source_network_replication}
\alias{drs_start_source_network_replication}
\title{Starts replication for a Source Network}
\usage{
drs_start_source_network_replication(sourceNetworkID)
}
\arguments{
\item{sourceNetworkID}{[required] ID of the Source Network to replicate.}
}
\description{
Starts replication for a Source Network. This action would make the Source Network protected.

See \url{https://www.paws-r-sdk.com/docs/drs_start_source_network_replication/} for full documentation.
}
\keyword{internal}
