% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_stop_replication}
\alias{drs_stop_replication}
\title{Stops replication for a Source Server}
\usage{
drs_stop_replication(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Server to stop replication for.}
}
\description{
Stops replication for a Source Server. This action would make the Source Server unprotected, delete its existing snapshots and stop billing for it.

See \url{https://www.paws-r-sdk.com/docs/drs_stop_replication/} for full documentation.
}
\keyword{internal}
