% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_tag_resource}
\alias{drs_tag_resource}
\title{Adds or overwrites only the specified tags for the specified Elastic
Disaster Recovery resource or resources}
\usage{
drs_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] ARN of the resource for which tags are to be added or updated.}

\item{tags}{[required] Array of tags to be added or updated.}
}
\description{
Adds or overwrites only the specified tags for the specified Elastic Disaster Recovery resource or resources. When you specify an existing tag key, the value is overwritten with the new value. Each resource can have a maximum of 50 tags. Each tag consists of a key and optional value.

See \url{https://www.paws-r-sdk.com/docs/drs_tag_resource/} for full documentation.
}
\keyword{internal}
