% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_terminate_recovery_instances}
\alias{drs_terminate_recovery_instances}
\title{Initiates a Job for terminating the EC2 resources associated with the
specified Recovery Instances, and then will delete the Recovery
Instances from the Elastic Disaster Recovery service}
\usage{
drs_terminate_recovery_instances(recoveryInstanceIDs)
}
\arguments{
\item{recoveryInstanceIDs}{[required] The IDs of the Recovery Instances that should be terminated.}
}
\description{
Initiates a Job for terminating the EC2 resources associated with the specified Recovery Instances, and then will delete the Recovery Instances from the Elastic Disaster Recovery service.

See \url{https://www.paws-r-sdk.com/docs/drs_terminate_recovery_instances/} for full documentation.
}
\keyword{internal}
