% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_untag_resource}
\alias{drs_untag_resource}
\title{Deletes the specified set of tags from the specified set of Elastic
Disaster Recovery resources}
\usage{
drs_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] ARN of the resource for which tags are to be removed.}

\item{tagKeys}{[required] Array of tags to be removed.}
}
\description{
Deletes the specified set of tags from the specified set of Elastic Disaster Recovery resources.

See \url{https://www.paws-r-sdk.com/docs/drs_untag_resource/} for full documentation.
}
\keyword{internal}
