% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_update_failback_replication_configuration}
\alias{drs_update_failback_replication_configuration}
\title{Allows you to update the failback replication configuration of a
Recovery Instance by ID}
\usage{
drs_update_failback_replication_configuration(
  recoveryInstanceID,
  name = NULL,
  bandwidthThrottling = NULL,
  usePrivateIP = NULL,
  internetProtocol = NULL
)
}
\arguments{
\item{recoveryInstanceID}{[required] The ID of the Recovery Instance.}

\item{name}{The name of the Failback Replication Configuration.}

\item{bandwidthThrottling}{Configure bandwidth throttling for the outbound data transfer rate of
the Recovery Instance in Mbps.}

\item{usePrivateIP}{Whether to use Private IP for the failback replication of the Recovery
Instance.}

\item{internetProtocol}{Which version of the Internet Protocol to use for replication of data.
(IPv4 or IPv6)}
}
\description{
Allows you to update the failback replication configuration of a Recovery Instance by ID.

See \url{https://www.paws-r-sdk.com/docs/drs_update_failback_replication_configuration/} for full documentation.
}
\keyword{internal}
