% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_update_recovery_plan}
\alias{drs_update_recovery_plan}
\title{Updates a Recovery Plan's name or description}
\usage{
drs_update_recovery_plan(recoveryPlanArn, name = NULL, description = NULL)
}
\arguments{
\item{recoveryPlanArn}{[required] The ARN of the Recovery Plan to update.}

\item{name}{The name of a Recovery Plan.}

\item{description}{The description of a Recovery Plan.}
}
\description{
Updates a Recovery Plan's name or description.

See \url{https://www.paws-r-sdk.com/docs/drs_update_recovery_plan/} for full documentation.
}
\keyword{internal}
