% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_update_recovery_plan_step}
\alias{drs_update_recovery_plan_step}
\title{Updates a Recovery Plan step's name or configuration}
\usage{
drs_update_recovery_plan_step(
  recoveryPlanStepArn,
  stepName = NULL,
  configuration = NULL
)
}
\arguments{
\item{recoveryPlanStepArn}{[required] The ARN of the Recovery Plan step to update.}

\item{stepName}{The name of a Recovery Plan Step.}

\item{configuration}{Type-specific configuration for a recovery plan step. Exactly one member must be set.}
}
\description{
Updates a Recovery Plan step's name or configuration. Step type is immutable.

See \url{https://www.paws-r-sdk.com/docs/drs_update_recovery_plan_step/} for full documentation.
}
\keyword{internal}
