% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_create_experiment_template}
\alias{fis_create_experiment_template}
\title{Creates an experiment template}
\usage{
fis_create_experiment_template(
  clientToken,
  description,
  stopConditions,
  targets = NULL,
  actions,
  roleArn,
  tags = NULL,
  logConfiguration = NULL
)
}
\arguments{
\item{clientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{description}{[required] A description for the experiment template.}

\item{stopConditions}{[required] The stop conditions.}

\item{targets}{The targets for the experiment.}

\item{actions}{[required] The actions for the experiment.}

\item{roleArn}{[required] The Amazon Resource Name (ARN) of an IAM role that grants the FIS
service permission to perform service actions on your behalf.}

\item{tags}{The tags to apply to the experiment template.}

\item{logConfiguration}{The configuration for experiment logging.}
}
\description{
Creates an experiment template.

See \url{https://paws-r.github.io/docs/fis/create_experiment_template.html} for full documentation.
}
\keyword{internal}
