% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_delete_experiment_template}
\alias{fis_delete_experiment_template}
\title{Deletes the specified experiment template}
\usage{
fis_delete_experiment_template(id)
}
\arguments{
\item{id}{[required] The ID of the experiment template.}
}
\description{
Deletes the specified experiment template.

See \url{https://paws-r.github.io/docs/fis/delete_experiment_template.html} for full documentation.
}
\keyword{internal}
