% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_delete_target_account_configuration}
\alias{fis_delete_target_account_configuration}
\title{Deletes the specified target account configuration of the experiment
template}
\usage{
fis_delete_target_account_configuration(experimentTemplateId, accountId)
}
\arguments{
\item{experimentTemplateId}{[required] The ID of the experiment template.}

\item{accountId}{[required] The AWS account ID of the target account.}
}
\description{
Deletes the specified target account configuration of the experiment template.

See \url{https://www.paws-r-sdk.com/docs/fis_delete_target_account_configuration/} for full documentation.
}
\keyword{internal}
