% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_action}
\alias{fis_get_action}
\title{Gets information about the specified FIS action}
\usage{
fis_get_action(id)
}
\arguments{
\item{id}{[required] The ID of the action.}
}
\description{
Gets information about the specified FIS action.

See \url{https://paws-r.github.io/docs/fis/get_action.html} for full documentation.
}
\keyword{internal}
