% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_experiment}
\alias{fis_get_experiment}
\title{Gets information about the specified experiment}
\usage{
fis_get_experiment(id)
}
\arguments{
\item{id}{[required] The ID of the experiment.}
}
\description{
Gets information about the specified experiment.

See \url{https://paws-r.github.io/docs/fis/get_experiment.html} for full documentation.
}
\keyword{internal}
