% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_experiment_target_account_configuration}
\alias{fis_get_experiment_target_account_configuration}
\title{Gets information about the specified target account configuration of the
experiment}
\usage{
fis_get_experiment_target_account_configuration(experimentId, accountId)
}
\arguments{
\item{experimentId}{[required] The ID of the experiment.}

\item{accountId}{[required] The AWS account ID of the target account.}
}
\description{
Gets information about the specified target account configuration of the experiment.

See \url{https://www.paws-r-sdk.com/docs/fis_get_experiment_target_account_configuration/} for full documentation.
}
\keyword{internal}
