% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_experiment_template}
\alias{fis_get_experiment_template}
\title{Gets information about the specified experiment template}
\usage{
fis_get_experiment_template(id)
}
\arguments{
\item{id}{[required] The ID of the experiment template.}
}
\description{
Gets information about the specified experiment template.

See \url{https://www.paws-r-sdk.com/docs/fis_get_experiment_template/} for full documentation.
}
\keyword{internal}
