% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_safety_lever}
\alias{fis_get_safety_lever}
\title{Gets information about the specified safety lever}
\usage{
fis_get_safety_lever(id)
}
\arguments{
\item{id}{[required] The ID of the safety lever.}
}
\description{
Gets information about the specified safety lever.

See \url{https://www.paws-r-sdk.com/docs/fis_get_safety_lever/} for full documentation.
}
\keyword{internal}
