% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_get_target_resource_type}
\alias{fis_get_target_resource_type}
\title{Gets information about the specified resource type}
\usage{
fis_get_target_resource_type(resourceType)
}
\arguments{
\item{resourceType}{[required] The resource type.}
}
\description{
Gets information about the specified resource type.

See \url{https://www.paws-r-sdk.com/docs/fis_get_target_resource_type/} for full documentation.
}
\keyword{internal}
