% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_actions}
\alias{fis_list_actions}
\title{Lists the available FIS actions}
\usage{
fis_list_actions(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists the available FIS actions.

See \url{https://paws-r.github.io/docs/fis/list_actions.html} for full documentation.
}
\keyword{internal}
