% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_experiment_resolved_targets}
\alias{fis_list_experiment_resolved_targets}
\title{Lists the resolved targets information of the specified experiment}
\usage{
fis_list_experiment_resolved_targets(
  experimentId,
  maxResults = NULL,
  nextToken = NULL,
  targetName = NULL
)
}
\arguments{
\item{experimentId}{[required] The ID of the experiment.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{nextToken}{The token for the next page of results.}

\item{targetName}{The name of the target.}
}
\description{
Lists the resolved targets information of the specified experiment.

See \url{https://www.paws-r-sdk.com/docs/fis_list_experiment_resolved_targets/} for full documentation.
}
\keyword{internal}
