% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_experiment_templates}
\alias{fis_list_experiment_templates}
\title{Lists your experiment templates}
\usage{
fis_list_experiment_templates(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists your experiment templates.

See \url{https://paws-r.github.io/docs/fis/list_experiment_templates.html} for full documentation.
}
\keyword{internal}
