% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_target_account_configurations}
\alias{fis_list_target_account_configurations}
\title{Lists the target account configurations of the specified experiment
template}
\usage{
fis_list_target_account_configurations(
  experimentTemplateId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{experimentTemplateId}{[required] The ID of the experiment template.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned nextToken
value.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists the target account configurations of the specified experiment template.

See \url{https://www.paws-r-sdk.com/docs/fis_list_target_account_configurations/} for full documentation.
}
\keyword{internal}
