% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_target_resource_types}
\alias{fis_list_target_resource_types}
\title{Lists the target resource types}
\usage{
fis_list_target_resource_types(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{nextToken}{The token for the next page of results.}
}
\description{
Lists the target resource types.

See \url{https://www.paws-r-sdk.com/docs/fis_list_target_resource_types/} for full documentation.
}
\keyword{internal}
