% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_start_experiment}
\alias{fis_start_experiment}
\title{Starts running an experiment from the specified experiment template}
\usage{
fis_start_experiment(clientToken, experimentTemplateId, tags = NULL)
}
\arguments{
\item{clientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{experimentTemplateId}{[required] The ID of the experiment template.}

\item{tags}{The tags to apply to the experiment.}
}
\description{
Starts running an experiment from the specified experiment template.

See \url{https://paws-r.github.io/docs/fis/start_experiment.html} for full documentation.
}
\keyword{internal}
