% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_stop_experiment}
\alias{fis_stop_experiment}
\title{Stops the specified experiment}
\usage{
fis_stop_experiment(id)
}
\arguments{
\item{id}{[required] The ID of the experiment.}
}
\description{
Stops the specified experiment.

See \url{https://www.paws-r-sdk.com/docs/fis_stop_experiment/} for full documentation.
}
\keyword{internal}
