% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_tag_resource}
\alias{fis_tag_resource}
\title{Applies the specified tags to the specified resource}
\usage{
fis_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags for the resource.}
}
\description{
Applies the specified tags to the specified resource.

See \url{https://paws-r.github.io/docs/fis/tag_resource.html} for full documentation.
}
\keyword{internal}
