% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_untag_resource}
\alias{fis_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
fis_untag_resource(resourceArn, tagKeys = NULL)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{The tag keys to remove.}
}
\description{
Removes the specified tags from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/fis_untag_resource/} for full documentation.
}
\keyword{internal}
