% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_update_experiment_template}
\alias{fis_update_experiment_template}
\title{Updates the specified experiment template}
\usage{
fis_update_experiment_template(
  id,
  description = NULL,
  stopConditions = NULL,
  targets = NULL,
  actions = NULL,
  roleArn = NULL,
  logConfiguration = NULL,
  experimentOptions = NULL
)
}
\arguments{
\item{id}{[required] The ID of the experiment template.}

\item{description}{A description for the template.}

\item{stopConditions}{The stop conditions for the experiment.}

\item{targets}{The targets for the experiment.}

\item{actions}{The actions for the experiment.}

\item{roleArn}{The Amazon Resource Name (ARN) of an IAM role that grants the FIS
service permission to perform service actions on your behalf.}

\item{logConfiguration}{The configuration for experiment logging.}

\item{experimentOptions}{The experiment options for the experiment template.}
}
\description{
Updates the specified experiment template.

See \url{https://www.paws-r-sdk.com/docs/fis_update_experiment_template/} for full documentation.
}
\keyword{internal}
