% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_update_safety_lever_state}
\alias{fis_update_safety_lever_state}
\title{Updates the specified safety lever state}
\usage{
fis_update_safety_lever_state(id, state)
}
\arguments{
\item{id}{[required] The ID of the safety lever.}

\item{state}{[required] The state of the safety lever.}
}
\description{
Updates the specified safety lever state.

See \url{https://www.paws-r-sdk.com/docs/fis_update_safety_lever_state/} for full documentation.
}
\keyword{internal}
