% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_update_target_account_configuration}
\alias{fis_update_target_account_configuration}
\title{Updates the target account configuration for the specified experiment
template}
\usage{
fis_update_target_account_configuration(
  experimentTemplateId,
  accountId,
  roleArn = NULL,
  description = NULL
)
}
\arguments{
\item{experimentTemplateId}{[required] The ID of the experiment template.}

\item{accountId}{[required] The Amazon Web Services account ID of the target account.}

\item{roleArn}{The Amazon Resource Name (ARN) of an IAM role for the target account.}

\item{description}{The description of the target account.}
}
\description{
Updates the target account configuration for the specified experiment template.

See \url{https://www.paws-r-sdk.com/docs/fis_update_target_account_configuration/} for full documentation.
}
\keyword{internal}
