% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_associate_profiles}
\alias{wellarchitected_associate_profiles}
\title{Associate a profile with a workload}
\usage{
wellarchitected_associate_profiles(WorkloadId, ProfileArns)
}
\arguments{
\item{WorkloadId}{[required]}

\item{ProfileArns}{[required] The list of profile ARNs to associate with the workload.}
}
\description{
Associate a profile with a workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_associate_profiles/} for full documentation.
}
\keyword{internal}
