% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_agent_goal}
\alias{wellarchitected_create_agent_goal}
\title{Creates an optimization goal associated with a profile}
\usage{
wellarchitected_create_agent_goal(
  clientToken = NULL,
  profileArn,
  pillars,
  title,
  description = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile to associate the goal with.}

\item{pillars}{[required] The Well-Architected Tool Framework pillars to associate with this goal.}

\item{title}{[required] The title of the goal.}

\item{description}{A description of the goal.}
}
\description{
Creates an optimization goal associated with a profile. Goals define specific targets and objectives for the optimization process.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_agent_goal/} for full documentation.
}
\keyword{internal}
