% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_review_template}
\alias{wellarchitected_create_review_template}
\title{Create a review template}
\usage{
wellarchitected_create_review_template(
  TemplateName,
  Description,
  Lenses,
  Notes = NULL,
  Tags = NULL,
  ClientRequestToken
)
}
\arguments{
\item{TemplateName}{[required] Name of the review template.}

\item{Description}{[required] The review template description.}

\item{Lenses}{[required] Lenses applied to the review template.}

\item{Notes}{The notes associated with the workload.

For a review template, these are the notes that will be associated with the workload when the template is applied.}

\item{Tags}{The tags assigned to the review template.}

\item{ClientRequestToken}{[required] A unique case-sensitive string used to ensure that this request is idempotent (executes only once).

You should not reuse the same token for other requests. If you retry a request with the same client request token and the same parameters after the original request has completed successfully, the result of the original request is returned.

This token is listed as required, however, if you do not specify it, the Amazon Web Services SDKs automatically generate one for you. If you are not using the Amazon Web Services SDK or the CLI, you must provide this token or the request will fail.}
}
\description{
Create a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_review_template/} for full documentation.
}
\keyword{internal}
