% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_template_share}
\alias{wellarchitected_create_template_share}
\title{Create a review template share}
\usage{
wellarchitected_create_template_share(
  TemplateArn,
  SharedWith,
  ClientRequestToken
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{SharedWith}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a review template share.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_template_share/} for full documentation.
}
\keyword{internal}
