% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_workload_share}
\alias{wellarchitected_create_workload_share}
\title{Create a workload share}
\usage{
wellarchitected_create_workload_share(
  WorkloadId,
  SharedWith,
  PermissionType,
  ClientRequestToken
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{SharedWith}{[required]}

\item{PermissionType}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a workload share.

See \url{https://paws-r.github.io/docs/wellarchitected/create_workload_share.html} for full documentation.
}
\keyword{internal}
