% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_delete_agent_goal}
\alias{wellarchitected_delete_agent_goal}
\title{Deletes an optimization goal from a profile}
\usage{
wellarchitected_delete_agent_goal(profileArn, id)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile containing the goal.}

\item{id}{[required] The unique identifier of the goal to delete.}
}
\description{
Deletes an optimization goal from a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_delete_agent_goal/} for full documentation.
}
\keyword{internal}
