% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_delete_review_template}
\alias{wellarchitected_delete_review_template}
\title{Delete a review template}
\usage{
wellarchitected_delete_review_template(TemplateArn, ClientRequestToken)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{ClientRequestToken}{[required]}
}
\description{
Delete a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_delete_review_template/} for full documentation.
}
\keyword{internal}
