% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_disassociate_lenses}
\alias{wellarchitected_disassociate_lenses}
\title{Disassociate a lens from a workload}
\usage{
wellarchitected_disassociate_lenses(WorkloadId, LensAliases)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{LensAliases}{[required] List of lens aliases to associate or disassociate with a workload. Up to 10 lenses can be specified.

Identify a lens using its LensSummary$LensAlias.}
}
\description{
Disassociate a lens from a workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_disassociate_lenses/} for full documentation.
}
\keyword{internal}
