% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_disassociate_profiles}
\alias{wellarchitected_disassociate_profiles}
\title{Disassociate a profile from a workload}
\usage{
wellarchitected_disassociate_profiles(WorkloadId, ProfileArns)
}
\arguments{
\item{WorkloadId}{[required]}

\item{ProfileArns}{[required] The list of profile ARNs to disassociate from the workload.}
}
\description{
Disassociate a profile from a workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_disassociate_profiles/} for full documentation.
}
\keyword{internal}
