% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_export_lens}
\alias{wellarchitected_export_lens}
\title{Export an existing lens}
\usage{
wellarchitected_export_lens(LensAlias, LensVersion = NULL)
}
\arguments{
\item{LensAlias}{[required]}

\item{LensVersion}{The lens version to be exported.}
}
\description{
Export an existing lens.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_export_lens/} for full documentation.
}
\keyword{internal}
