% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_agent_context}
\alias{wellarchitected_get_agent_context}
\title{Retrieves detailed information about a specific context associated with
a profile}
\usage{
wellarchitected_get_agent_context(profileArn, id)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile containing the context.}

\item{id}{[required] The unique identifier of the context to retrieve.}
}
\description{
Retrieves detailed information about a specific context associated with a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_agent_context/} for full documentation.
}
\keyword{internal}
