% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_agent_goal}
\alias{wellarchitected_get_agent_goal}
\title{Retrieves detailed information about a specific optimization goal}
\usage{
wellarchitected_get_agent_goal(profileArn, id)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile containing the goal.}

\item{id}{[required] The unique identifier of the goal to retrieve.}
}
\description{
Retrieves detailed information about a specific optimization goal.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_agent_goal/} for full documentation.
}
\keyword{internal}
