% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_agent_profile}
\alias{wellarchitected_get_agent_profile}
\title{Retrieves detailed information about an optimization profile, including
its configuration and metadata}
\usage{
wellarchitected_get_agent_profile(profileArn)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the optimization profile to retrieve.}
}
\description{
Retrieves detailed information about an optimization profile, including its configuration and metadata.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_agent_profile/} for full documentation.
}
\keyword{internal}
