% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_answer}
\alias{wellarchitected_get_answer}
\title{Get the answer to a specific question in a workload review}
\usage{
wellarchitected_get_answer(
  WorkloadId,
  LensAlias,
  QuestionId,
  MilestoneNumber = NULL
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{QuestionId}{[required] The ID of the question.}

\item{MilestoneNumber}{The milestone number.

A workload can have a maximum of 100 milestones.}
}
\description{
Get the answer to a specific question in a workload review.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_answer/} for full documentation.
}
\keyword{internal}
