% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_global_settings}
\alias{wellarchitected_get_global_settings}
\title{Global settings for all workloads}
\usage{
wellarchitected_get_global_settings()
}
\description{
Global settings for all workloads.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_global_settings/} for full documentation.
}
\keyword{internal}
