% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens}
\alias{wellarchitected_get_lens}
\title{Get an existing lens}
\usage{
wellarchitected_get_lens(LensAlias, LensVersion = NULL)
}
\arguments{
\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{LensVersion}{The lens version to be retrieved.}
}
\description{
Get an existing lens.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_lens/} for full documentation.
}
\keyword{internal}
