% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens_review_report}
\alias{wellarchitected_get_lens_review_report}
\title{Get lens review report}
\usage{
wellarchitected_get_lens_review_report(
  WorkloadId,
  LensAlias,
  MilestoneNumber = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{MilestoneNumber}{}
}
\description{
Get lens review report.

See \url{https://paws-r.github.io/docs/wellarchitected/get_lens_review_report.html} for full documentation.
}
\keyword{internal}
