% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens_version_difference}
\alias{wellarchitected_get_lens_version_difference}
\title{Get lens version differences}
\usage{
wellarchitected_get_lens_version_difference(
  LensAlias,
  BaseLensVersion = NULL,
  TargetLensVersion = NULL
)
}
\arguments{
\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{BaseLensVersion}{The base version of the lens.}

\item{TargetLensVersion}{The lens version to target a difference for.}
}
\description{
Get lens version differences.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_lens_version_difference/} for full documentation.
}
\keyword{internal}
