% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_milestone}
\alias{wellarchitected_get_milestone}
\title{Get a milestone for an existing workload}
\usage{
wellarchitected_get_milestone(WorkloadId, MilestoneNumber)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{MilestoneNumber}{[required] The milestone number.

A workload can have a maximum of 100 milestones.}
}
\description{
Get a milestone for an existing workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_milestone/} for full documentation.
}
\keyword{internal}
