% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_profile}
\alias{wellarchitected_get_profile}
\title{Get profile information}
\usage{
wellarchitected_get_profile(ProfileArn, ProfileVersion = NULL)
}
\arguments{
\item{ProfileArn}{[required] The profile ARN.}

\item{ProfileVersion}{The profile version.}
}
\description{
Get profile information.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_profile/} for full documentation.
}
\keyword{internal}
