% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_review_template}
\alias{wellarchitected_get_review_template}
\title{Get review template}
\usage{
wellarchitected_get_review_template(TemplateArn)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}
}
\description{
Get review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_review_template/} for full documentation.
}
\keyword{internal}
