% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_agent_contexts}
\alias{wellarchitected_list_agent_contexts}
\title{Lists contexts associated with a profile}
\usage{
wellarchitected_list_agent_contexts(
  profileArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile to list contexts for.}

\item{maxResults}{The maximum number of results to return for this request.}

\item{nextToken}{The token to use to retrieve the next set of results.}
}
\description{
Lists contexts associated with a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_agent_contexts/} for full documentation.
}
\keyword{internal}
