% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_agent_goals}
\alias{wellarchitected_list_agent_goals}
\title{Lists optimization goals associated with a specified profile}
\usage{
wellarchitected_list_agent_goals(
  profileArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the optimization profile to list goals for.}

\item{maxResults}{The maximum number of goals to return in a single response.}

\item{nextToken}{A pagination token returned from a previous call to continue retrieving results.}
}
\description{
Lists optimization goals associated with a specified profile. Goals define specific targets and objectives for the optimization process.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_agent_goals/} for full documentation.
}
\keyword{internal}
