% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_agent_profiles}
\alias{wellarchitected_list_agent_profiles}
\title{Lists optimization profiles in your account}
\usage{
wellarchitected_list_agent_profiles(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of profiles to return in a single call. Default is 100.}

\item{nextToken}{A pagination token returned from a previous call to continue retrieving results.}
}
\description{
Lists optimization profiles in your account. Profiles define the scope and configuration for generating optimization recommendations.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_agent_profiles/} for full documentation.
}
\keyword{internal}
