% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_agent_recommendation_generations}
\alias{wellarchitected_list_agent_recommendation_generations}
\title{Lists recommendation generation processes for a specified profile}
\usage{
wellarchitected_list_agent_recommendation_generations(
  profileArn,
  recommendationType = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{profileArn}{[required] The Amazon Resource Name (ARN) of the optimization profile to list generation processes for.}

\item{recommendationType}{Optional filter by recommendation type.}

\item{maxResults}{The maximum number of generation processes to return in a single response.}

\item{nextToken}{A pagination token returned from a previous call to continue retrieving results.}
}
\description{
Lists recommendation generation processes for a specified profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_agent_recommendation_generations/} for full documentation.
}
\keyword{internal}
