% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_answers}
\alias{wellarchitected_list_answers}
\title{List of answers}
\usage{
wellarchitected_list_answers(
  WorkloadId,
  LensAlias,
  PillarId = NULL,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List of answers.

See \url{https://paws-r.github.io/docs/wellarchitected/list_answers.html} for full documentation.
}
\keyword{internal}
