% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_check_details}
\alias{wellarchitected_list_check_details}
\title{List of Trusted Advisor check details by account related to the workload}
\usage{
wellarchitected_list_check_details(
  WorkloadId,
  NextToken = NULL,
  MaxResults = NULL,
  LensArn,
  PillarId,
  QuestionId,
  ChoiceId
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{NextToken}{The token to use to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{LensArn}{[required] Well-Architected Lens ARN.}

\item{PillarId}{[required] The ID used to identify a pillar, for example, \code{security}.

A pillar is identified by its PillarReviewSummary$PillarId.}

\item{QuestionId}{[required] The ID of the question.}

\item{ChoiceId}{[required] The ID of a choice.}
}
\description{
List of Trusted Advisor check details by account related to the workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_check_details/} for full documentation.
}
\keyword{internal}
