% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_check_summaries}
\alias{wellarchitected_list_check_summaries}
\title{List of Trusted Advisor checks summarized for all accounts related to
the workload}
\usage{
wellarchitected_list_check_summaries(
  WorkloadId,
  NextToken = NULL,
  MaxResults = NULL,
  LensArn,
  PillarId,
  QuestionId,
  ChoiceId
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{NextToken}{}

\item{MaxResults}{}

\item{LensArn}{[required] Well-Architected Lens ARN.}

\item{PillarId}{[required]}

\item{QuestionId}{[required]}

\item{ChoiceId}{[required]}
}
\description{
List of Trusted Advisor checks summarized for all accounts related to the workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_check_summaries/} for full documentation.
}
\keyword{internal}
