% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_review_improvements}
\alias{wellarchitected_list_lens_review_improvements}
\title{List the improvements of a particular lens review}
\usage{
wellarchitected_list_lens_review_improvements(
  WorkloadId,
  LensAlias,
  PillarId = NULL,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  QuestionPriority = NULL
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{PillarId}{The ID used to identify a pillar, for example, \code{security}.

A pillar is identified by its PillarReviewSummary$PillarId.}

\item{MilestoneNumber}{The milestone number.

A workload can have a maximum of 100 milestones.}

\item{NextToken}{The token to use to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{QuestionPriority}{The priority of the question.}
}
\description{
List the improvements of a particular lens review.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_lens_review_improvements/} for full documentation.
}
\keyword{internal}
