% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_reviews}
\alias{wellarchitected_list_lens_reviews}
\title{List lens reviews for a particular workload}
\usage{
wellarchitected_list_lens_reviews(
  WorkloadId,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{MilestoneNumber}{The milestone number.

A workload can have a maximum of 100 milestones.}

\item{NextToken}{The token to use to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List lens reviews for a particular workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_lens_reviews/} for full documentation.
}
\keyword{internal}
