% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_milestones}
\alias{wellarchitected_list_milestones}
\title{List all milestones for an existing workload}
\usage{
wellarchitected_list_milestones(
  WorkloadId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{NextToken}{The token to use to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List all milestones for an existing workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_milestones/} for full documentation.
}
\keyword{internal}
