% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_profiles}
\alias{wellarchitected_list_profiles}
\title{List profiles}
\usage{
wellarchitected_list_profiles(
  ProfileNamePrefix = NULL,
  ProfileOwnerType = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ProfileNamePrefix}{An optional string added to the beginning of each profile name returned
in the results.}

\item{ProfileOwnerType}{Profile owner type.}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List profiles.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_profiles/} for full documentation.
}
\keyword{internal}
