% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_review_template_answers}
\alias{wellarchitected_list_review_template_answers}
\title{List the answers of a review template}
\usage{
wellarchitected_list_review_template_answers(
  TemplateArn,
  LensAlias,
  PillarId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The ARN of the review template.}

\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{PillarId}{The ID used to identify a pillar, for example, \code{security}.

A pillar is identified by its PillarReviewSummary$PillarId.}

\item{NextToken}{The token to use to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List the answers of a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_review_template_answers/} for full documentation.
}
\keyword{internal}
