% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_review_templates}
\alias{wellarchitected_list_review_templates}
\title{List review templates}
\usage{
wellarchitected_list_review_templates(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List review templates.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_review_templates/} for full documentation.
}
\keyword{internal}
