% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_share_invitations}
\alias{wellarchitected_list_share_invitations}
\title{List the workload invitations}
\usage{
wellarchitected_list_share_invitations(
  WorkloadNamePrefix = NULL,
  LensNamePrefix = NULL,
  ShareResourceType = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  ProfileNamePrefix = NULL
)
}
\arguments{
\item{WorkloadNamePrefix}{}

\item{LensNamePrefix}{An optional string added to the beginning of each lens name returned in
the results.}

\item{ShareResourceType}{The type of share invitations to be returned.}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{ProfileNamePrefix}{Profile name prefix.}
}
\description{
List the workload invitations.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_share_invitations/} for full documentation.
}
\keyword{internal}
