% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_template_shares}
\alias{wellarchitected_list_template_shares}
\title{List review template shares}
\usage{
wellarchitected_list_template_shares(
  TemplateArn,
  SharedWithPrefix = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Status = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{SharedWithPrefix}{The Amazon Web Services account ID, organization ID, or organizational
unit (OU) ID with which the profile is shared.}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{Status}{}
}
\description{
List review template shares.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_template_shares/} for full documentation.
}
\keyword{internal}
