% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_tag_resource}
\alias{wellarchitected_tag_resource}
\title{Adds one or more tags to the specified resource}
\usage{
wellarchitected_tag_resource(WorkloadArn, Tags)
}
\arguments{
\item{WorkloadArn}{[required]}

\item{Tags}{[required] The tags for the resource.}
}
\description{
Adds one or more tags to the specified resource.

See \url{https://paws-r.github.io/docs/wellarchitected/tag_resource.html} for full documentation.
}
\keyword{internal}
