% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_agent_profile}
\alias{wellarchitected_update_agent_profile}
\title{Updates an existing optimization profile's configuration, including its
pillars, execution role, and aggregation settings}
\usage{
wellarchitected_update_agent_profile(
  clientToken = NULL,
  profileArn,
  displayName = NULL,
  description = NULL,
  executionRoleArn = NULL,
  aggregationConfiguration = NULL,
  businessOverview = NULL,
  pillars = NULL,
  deletionProtection = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{profileArn}{[required] The Amazon Resource Name (ARN) of the profile to update.}

\item{displayName}{The updated display name of the profile.}

\item{description}{The updated description of the profile.}

\item{executionRoleArn}{The updated ARN of the IAM execution role.}

\item{aggregationConfiguration}{The updated aggregation configuration.}

\item{businessOverview}{The updated business overview for the profile.}

\item{pillars}{The updated Well-Architected Tool Framework pillars for the profile.}

\item{deletionProtection}{Indicates whether deletion protection is enabled for the profile.}
}
\description{
Updates an existing optimization profile's configuration, including its pillars, execution role, and aggregation settings.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_agent_profile/} for full documentation.
}
\keyword{internal}
