% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_agent_recommendation_status}
\alias{wellarchitected_update_agent_recommendation_status}
\title{Updates the status of a recommendation to track its progress through the
implementation lifecycle}
\usage{
wellarchitected_update_agent_recommendation_status(
  recommendationArn,
  status,
  updateReason = NULL
)
}
\arguments{
\item{recommendationArn}{[required] The Amazon Resource Name (ARN) of the recommendation to update.}

\item{status}{[required] The new status to assign to the recommendation.}

\item{updateReason}{A free-text reason explaining this status update.}
}
\description{
Updates the status of a recommendation to track its progress through the implementation lifecycle.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_agent_recommendation_status/} for full documentation.
}
\keyword{internal}
