% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_answer}
\alias{wellarchitected_update_answer}
\title{Update the answer to a specific question in a workload review}
\usage{
wellarchitected_update_answer(
  WorkloadId,
  LensAlias,
  QuestionId,
  SelectedChoices = NULL,
  ChoiceUpdates = NULL,
  Notes = NULL,
  IsApplicable = NULL,
  Reason = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}

\item{SelectedChoices}{}

\item{ChoiceUpdates}{A list of choices to update on a question in your workload. The String
key corresponds to the choice ID to be updated.}

\item{Notes}{}

\item{IsApplicable}{}

\item{Reason}{The reason why a question is not applicable to your workload.}
}
\description{
Update the answer to a specific question in a workload review.

See \url{https://paws-r.github.io/docs/wellarchitected/update_answer.html} for full documentation.
}
\keyword{internal}
