% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_profile}
\alias{wellarchitected_update_profile}
\title{Update a profile}
\usage{
wellarchitected_update_profile(
  ProfileArn,
  ProfileDescription = NULL,
  ProfileQuestions = NULL
)
}
\arguments{
\item{ProfileArn}{[required] The profile ARN.}

\item{ProfileDescription}{The profile description.}

\item{ProfileQuestions}{Profile questions.}
}
\description{
Update a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_profile/} for full documentation.
}
\keyword{internal}
