% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_review_template}
\alias{wellarchitected_update_review_template}
\title{Update a review template}
\usage{
wellarchitected_update_review_template(
  TemplateArn,
  TemplateName = NULL,
  Description = NULL,
  Notes = NULL,
  LensesToAssociate = NULL,
  LensesToDisassociate = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{TemplateName}{The review template name.}

\item{Description}{The review template description.}

\item{Notes}{}

\item{LensesToAssociate}{A list of lens aliases or ARNs to apply to the review template.}

\item{LensesToDisassociate}{A list of lens aliases or ARNs to unapply to the review template. The
\code{wellarchitected} lens cannot be unapplied.}
}
\description{
Update a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_review_template/} for full documentation.
}
\keyword{internal}
