% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_review_template_answer}
\alias{wellarchitected_update_review_template_answer}
\title{Update a review template answer}
\usage{
wellarchitected_update_review_template_answer(
  TemplateArn,
  LensAlias,
  QuestionId,
  SelectedChoices = NULL,
  ChoiceUpdates = NULL,
  Notes = NULL,
  IsApplicable = NULL,
  Reason = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{QuestionId}{[required] The ID of the question.}

\item{SelectedChoices}{List of selected choice IDs in a question answer.

The values entered replace the previously selected choices.}

\item{ChoiceUpdates}{A list of choices to be updated.}

\item{Notes}{The notes associated with the workload.

For a review template, these are the notes that will be associated with the workload when the template is applied.}

\item{IsApplicable}{Defines whether this question is applicable to a lens review.}

\item{Reason}{The update reason.}
}
\description{
Update a review template answer.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_review_template_answer/} for full documentation.
}
\keyword{internal}
