% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_review_template_lens_review}
\alias{wellarchitected_update_review_template_lens_review}
\title{Update a lens review associated with a review template}
\usage{
wellarchitected_update_review_template_lens_review(
  TemplateArn,
  LensAlias,
  LensNotes = NULL,
  PillarNotes = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{LensAlias}{[required]}

\item{LensNotes}{}

\item{PillarNotes}{}
}
\description{
Update a lens review associated with a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_review_template_lens_review/} for full documentation.
}
\keyword{internal}
