% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_share_invitation}
\alias{wellarchitected_update_share_invitation}
\title{Update a workload invitation}
\usage{
wellarchitected_update_share_invitation(
  ShareInvitationId,
  ShareInvitationAction
)
}
\arguments{
\item{ShareInvitationId}{[required] The ID assigned to the share invitation.}

\item{ShareInvitationAction}{[required]}
}
\description{
Update a workload invitation.

See \url{https://paws-r.github.io/docs/wellarchitected/update_share_invitation.html} for full documentation.
}
\keyword{internal}
