% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_workload_share}
\alias{wellarchitected_update_workload_share}
\title{Update a workload share}
\usage{
wellarchitected_update_workload_share(ShareId, WorkloadId, PermissionType)
}
\arguments{
\item{ShareId}{[required] The ID associated with the share.}

\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{PermissionType}{[required] Permission granted on a share request.}
}
\description{
Update a workload share.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_workload_share/} for full documentation.
}
\keyword{internal}
