% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_upgrade_lens_review}
\alias{wellarchitected_upgrade_lens_review}
\title{Upgrade lens review for a particular workload}
\usage{
wellarchitected_upgrade_lens_review(
  WorkloadId,
  LensAlias,
  MilestoneName,
  ClientRequestToken = NULL
)
}
\arguments{
\item{WorkloadId}{[required] The ID assigned to the workload. This ID is unique within an Amazon Web Services Region.}

\item{LensAlias}{[required] The alias of the lens.

For Amazon Web Services official lenses, this is either the lens alias, such as \code{serverless}, or the lens ARN, such as \verb{arn:aws:wellarchitected:us-east-1::lens/serverless}. Note that some operations (such as ExportLens and CreateLensShare) are not permitted on Amazon Web Services official lenses.

For custom lenses, this is the lens ARN, such as \verb{arn:aws:wellarchitected:us-west-2:123456789012:lens/0123456789abcdef01234567890abcdef}.

Each lens is identified by its LensSummary$LensAlias.}

\item{MilestoneName}{[required] The name of the milestone in a workload.

Milestone names must be unique within a workload.}

\item{ClientRequestToken}{A unique case-sensitive string used to ensure that this request is idempotent (executes only once).

You should not reuse the same token for other requests. If you retry a request with the same client request token and the same parameters after the original request has completed successfully, the result of the original request is returned.

This token is listed as required, however, if you do not specify it, the Amazon Web Services SDKs automatically generate one for you. If you are not using the Amazon Web Services SDK or the CLI, you must provide this token or the request will fail.}
}
\description{
Upgrade lens review for a particular workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_upgrade_lens_review/} for full documentation.
}
\keyword{internal}
