% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_create_group}
\alias{xray_create_group}
\title{Creates a group resource with a name and a filter expression}
\usage{
xray_create_group(GroupName, FilterExpression)
}
\arguments{
\item{GroupName}{[required] The case-sensitive name of the new group. Default is a reserved name and
names must be unique.}

\item{FilterExpression}{The filter expression defining criteria by which to group traces.}
}
\description{
Creates a group resource with a name and a filter expression.
}
\section{Request syntax}{
\preformatted{svc$create_group(
  GroupName = "string",
  FilterExpression = "string"
)
}
}

\keyword{internal}
