% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_delete_group}
\alias{xray_delete_group}
\title{Deletes a group resource}
\usage{
xray_delete_group(GroupName, GroupARN)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN of the group that was generated on creation.}
}
\value{
An empty list.
}
\description{
Deletes a group resource.
}
\section{Request syntax}{
\preformatted{svc$delete_group(
  GroupName = "string",
  GroupARN = "string"
)
}
}

\keyword{internal}
