% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_encryption_config}
\alias{xray_get_encryption_config}
\title{Retrieves the current encryption configuration for X-Ray data}
\usage{
xray_get_encryption_config()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EncryptionConfig = list(
    KeyId = "string",
    Status = "UPDATING"|"ACTIVE",
    Type = "NONE"|"KMS"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the current encryption configuration for X-Ray data.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_encryption_config()
}\if{html}{\out{</div>}}
}

\keyword{internal}
