% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_group}
\alias{xray_get_group}
\title{Retrieves group resource details}
\usage{
xray_get_group(GroupName, GroupARN)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN of the group that was generated on creation.}
}
\description{
Retrieves group resource details.
}
\section{Request syntax}{
\preformatted{svc$get_group(
  GroupName = "string",
  GroupARN = "string"
)
}
}

\keyword{internal}
