% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_indexing_rules}
\alias{xray_get_indexing_rules}
\title{Retrieves all indexing rules}
\usage{
xray_get_indexing_rules(NextToken = NULL)
}
\arguments{
\item{NextToken}{Specify the pagination token returned by a previous request to retrieve
the next page of indexes.}
}
\description{
Retrieves all indexing rules.

See \url{https://www.paws-r-sdk.com/docs/xray_get_indexing_rules/} for full documentation.
}
\keyword{internal}
