% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_insight}
\alias{xray_get_insight}
\title{Retrieves the summary information of an insight}
\usage{
xray_get_insight(InsightId)
}
\arguments{
\item{InsightId}{[required] The insight's unique identifier. Use the GetInsightSummaries action to
retrieve an InsightId.}
}
\description{
Retrieves the summary information of an insight. This includes impact to clients and root cause services, the top anomalous services, the category, the state of the insight, and the start and end time of the insight.

See \url{https://www.paws-r-sdk.com/docs/xray_get_insight/} for full documentation.
}
\keyword{internal}
