% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_retrieved_traces_graph}
\alias{xray_get_retrieved_traces_graph}
\title{Retrieves a service graph for traces based on the specified
RetrievalToken from the CloudWatch log group generated by Transaction
Search}
\usage{
xray_get_retrieved_traces_graph(RetrievalToken, NextToken = NULL)
}
\arguments{
\item{RetrievalToken}{[required] Retrieval token.}

\item{NextToken}{Specify the pagination token returned by a previous request to retrieve
the next page of indexes.}
}
\description{
Retrieves a service graph for traces based on the specified \code{RetrievalToken} from the CloudWatch log group generated by Transaction Search. This API does not initiate a retrieval job. You must first execute \code{\link[=xray_start_trace_retrieval]{start_trace_retrieval}} to obtain the required \code{RetrievalToken}.

See \url{https://www.paws-r-sdk.com/docs/xray_get_retrieved_traces_graph/} for full documentation.
}
\keyword{internal}
