% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_sampling_rules}
\alias{xray_get_sampling_rules}
\title{Retrieves all sampling rules}
\usage{
xray_get_sampling_rules(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token.}
}
\description{
Retrieves all sampling rules.

See \url{https://www.paws-r-sdk.com/docs/xray_get_sampling_rules/} for full documentation.
}
\keyword{internal}
