% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_sampling_targets}
\alias{xray_get_sampling_targets}
\title{Requests a sampling quota for rules that the service is using to sample
requests}
\usage{
xray_get_sampling_targets(SamplingStatisticsDocuments)
}
\arguments{
\item{SamplingStatisticsDocuments}{[required] Information about rules that the service is using to sample requests.}
}
\description{
Requests a sampling quota for rules that the service is using to sample requests.

See \url{https://paws-r.github.io/docs/xray/get_sampling_targets.html} for full documentation.
}
\keyword{internal}
