% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_trace_graph}
\alias{xray_get_trace_graph}
\title{Retrieves a service graph for one or more specific trace IDs}
\usage{
xray_get_trace_graph(TraceIds, NextToken = NULL)
}
\arguments{
\item{TraceIds}{[required] Trace IDs of requests for which to generate a service graph.}

\item{NextToken}{Pagination token.}
}
\description{
Retrieves a service graph for one or more specific trace IDs.

See \url{https://paws-r.github.io/docs/xray/get_trace_graph.html} for full documentation.
}
\keyword{internal}
