% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_list_retrieved_traces}
\alias{xray_list_retrieved_traces}
\title{Retrieves a list of traces for a given RetrievalToken from the
CloudWatch log group generated by Transaction Search}
\usage{
xray_list_retrieved_traces(
  RetrievalToken,
  TraceFormat = NULL,
  NextToken = NULL
)
}
\arguments{
\item{RetrievalToken}{[required] Retrieval token.}

\item{TraceFormat}{Format of the requested traces.}

\item{NextToken}{Specify the pagination token returned by a previous request to retrieve
the next page of indexes.}
}
\description{
Retrieves a list of traces for a given \code{RetrievalToken} from the CloudWatch log group generated by Transaction Search. For information on what each trace returns, see \code{\link[=xray_batch_get_traces]{batch_get_traces}}.

See \url{https://www.paws-r-sdk.com/docs/xray_list_retrieved_traces/} for full documentation.
}
\keyword{internal}
