% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_put_telemetry_records}
\alias{xray_put_telemetry_records}
\title{Used by the Amazon Web Services X-Ray daemon to upload telemetry}
\usage{
xray_put_telemetry_records(
  TelemetryRecords,
  EC2InstanceId = NULL,
  Hostname = NULL,
  ResourceARN = NULL
)
}
\arguments{
\item{TelemetryRecords}{[required]}
}
\description{
Used by the Amazon Web Services X-Ray daemon to upload telemetry.

See \url{https://www.paws-r-sdk.com/docs/xray_put_telemetry_records/} for full documentation.
}
\keyword{internal}
