% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_put_trace_segments}
\alias{xray_put_trace_segments}
\title{Uploads segment documents to Amazon Web Services X-Ray}
\usage{
xray_put_trace_segments(TraceSegmentDocuments)
}
\arguments{
\item{TraceSegmentDocuments}{[required] A string containing a JSON document defining one or more segments or
subsegments.}
}
\description{
Uploads segment documents to Amazon Web Services X-Ray. The \href{https://docs.aws.amazon.com/xray/index.html}{X-Ray SDK} generates segment documents and sends them to the X-Ray daemon, which uploads them in batches. A segment document can be a completed segment, an in-progress segment, or an array of subsegments.

See \url{https://www.paws-r-sdk.com/docs/xray_put_trace_segments/} for full documentation.
}
\keyword{internal}
