% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_untag_resource}
\alias{xray_untag_resource}
\title{Removes tags from an Amazon Web Services X-Ray group or sampling rule}
\usage{
xray_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Number (ARN) of an X-Ray group or sampling rule.}

\item{TagKeys}{[required] Keys for one or more tags that you want to remove from an X-Ray group or
sampling rule.}
}
\description{
Removes tags from an Amazon Web Services X-Ray group or sampling rule. You cannot edit or delete system tags (those with an \verb{aws:} prefix).

See \url{https://paws-r.github.io/docs/xray/untag_resource.html} for full documentation.
}
\keyword{internal}
