% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_update_group}
\alias{xray_update_group}
\title{Updates a group resource}
\usage{
xray_update_group(
  GroupName = NULL,
  GroupARN = NULL,
  FilterExpression = NULL,
  InsightsConfiguration = NULL
)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN that was generated upon creation.}

\item{FilterExpression}{The updated filter expression defining criteria by which to group
traces.}

\item{InsightsConfiguration}{The structure containing configurations related to insights.
\itemize{
\item The InsightsEnabled boolean can be set to true to enable insights
for the group or false to disable insights for the group.
\item The NotifcationsEnabled boolean can be set to true to enable
insights notifications for the group. Notifications can only be
enabled on a group with InsightsEnabled set to true.
}}
}
\description{
Updates a group resource.

See \url{https://paws-r.github.io/docs/xray/update_group.html} for full documentation.
}
\keyword{internal}
