% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_update_indexing_rule}
\alias{xray_update_indexing_rule}
\title{Modifies an indexing rule’s configuration}
\usage{
xray_update_indexing_rule(Name, Rule)
}
\arguments{
\item{Name}{[required] Name of the indexing rule to be updated.}

\item{Rule}{[required] Rule configuration to be updated.}
}
\description{
Modifies an indexing rule’s configuration.

See \url{https://www.paws-r-sdk.com/docs/xray_update_indexing_rule/} for full documentation.
}
\keyword{internal}
