% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_update_trace_segment_destination}
\alias{xray_update_trace_segment_destination}
\title{Modifies the destination of data sent to PutTraceSegments}
\usage{
xray_update_trace_segment_destination(Destination = NULL)
}
\arguments{
\item{Destination}{The configured destination of trace segments.}
}
\description{
Modifies the destination of data sent to \code{\link[=xray_put_trace_segments]{put_trace_segments}}. The Transaction Search feature requires the CloudWatchLogs destination. For more information, see \href{https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Transaction-Search.html}{Transaction Search}.

See \url{https://www.paws-r-sdk.com/docs/xray_update_trace_segment_destination/} for full documentation.
}
\keyword{internal}
