% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_delete_environment_membership}
\alias{cloud9_delete_environment_membership}
\title{Deletes an environment member from a development environment}
\usage{
cloud9_delete_environment_membership(environmentId, userArn)
}
\arguments{
\item{environmentId}{[required] The ID of the environment to delete the environment member from.}

\item{userArn}{[required] The Amazon Resource Name (ARN) of the environment member to delete from
the environment.}
}
\description{
Deletes an environment member from a development environment.

See \url{https://www.paws-r-sdk.com/docs/cloud9_delete_environment_membership/} for full documentation.
}
\keyword{internal}
