% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_describe_environment_status}
\alias{cloud9_describe_environment_status}
\title{Gets status information for an Cloud9 development environment}
\usage{
cloud9_describe_environment_status(environmentId)
}
\arguments{
\item{environmentId}{[required] The ID of the environment to get status information about.}
}
\description{
Gets status information for an Cloud9 development environment.

See \url{https://paws-r.github.io/docs/cloud9/describe_environment_status.html} for full documentation.
}
\keyword{internal}
