% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_describe_environments}
\alias{cloud9_describe_environments}
\title{Gets information about Cloud9 development environments}
\usage{
cloud9_describe_environments(environmentIds)
}
\arguments{
\item{environmentIds}{[required] The IDs of individual environments to get information about.}
}
\description{
Gets information about Cloud9 development environments.

See \url{https://paws-r.github.io/docs/cloud9/describe_environments.html} for full documentation.
}
\keyword{internal}
