% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_tag_resource}
\alias{cloud9_tag_resource}
\title{Adds tags to an AWS Cloud9 development environment}
\usage{
cloud9_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the AWS Cloud9 development environment
to add tags to.}

\item{Tags}{[required] The list of tags to add to the given AWS Cloud9 development environment.}
}
\value{
An empty list.
}
\description{
Adds tags to an AWS Cloud9 development environment.

Tags that you add to an AWS Cloud9 environment by using this method will
NOT be automatically propagated to underlying resources.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
