% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_untag_resource}
\alias{cloud9_untag_resource}
\title{Removes tags from an AWS Cloud9 development environment}
\usage{
cloud9_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the AWS Cloud9 development environment
to remove tags from.}

\item{TagKeys}{[required] The tag names of the tags to remove from the given AWS Cloud9
development environment.}
}
\value{
An empty list.
}
\description{
Removes tags from an AWS Cloud9 development environment.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
