% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud9_operations.R
\name{cloud9_update_environment_membership}
\alias{cloud9_update_environment_membership}
\title{Changes the settings of an existing environment member for an Cloud9
development environment}
\usage{
cloud9_update_environment_membership(environmentId, userArn, permissions)
}
\arguments{
\item{environmentId}{[required] The ID of the environment for the environment member whose settings you
want to change.}

\item{userArn}{[required] The Amazon Resource Name (ARN) of the environment member whose settings
you want to change.}

\item{permissions}{[required] The replacement type of environment member permissions you want to
associate with this environment member. Available values include:
\itemize{
\item \code{read-only}: Has read-only access to the environment.
\item \code{read-write}: Has read-write access to the environment.
}}
}
\description{
Changes the settings of an existing environment member for an Cloud9 development environment.

See \url{https://paws-r.github.io/docs/cloud9/update_environment_membership.html} for full documentation.
}
\keyword{internal}
