% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeartifact_operations.R
\name{codeartifact_list_tags_for_resource}
\alias{codeartifact_list_tags_for_resource}
\title{Gets information about Amazon Web Services tags for a specified Amazon
Resource Name (ARN) in CodeArtifact}
\usage{
codeartifact_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to get tags for.}
}
\description{
Gets information about Amazon Web Services tags for a specified Amazon Resource Name (ARN) in CodeArtifact.

See \url{https://paws-r.github.io/docs/codeartifact/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
