% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_delete_builds}
\alias{codebuild_batch_delete_builds}
\title{Deletes one or more builds}
\usage{
codebuild_batch_delete_builds(ids)
}
\arguments{
\item{ids}{[required] The IDs of the builds to delete.}
}
\description{
Deletes one or more builds.

See \url{https://www.paws-r-sdk.com/docs/codebuild_batch_delete_builds/} for full documentation.
}
\keyword{internal}
