% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_build_batches}
\alias{codebuild_batch_get_build_batches}
\title{Retrieves information about one or more batch builds}
\usage{
codebuild_batch_get_build_batches(ids)
}
\arguments{
\item{ids}{[required] An array that contains the batch build identifiers to retrieve.}
}
\description{
Retrieves information about one or more batch builds.

See \url{https://paws-r.github.io/docs/codebuild/batch_get_build_batches.html} for full documentation.
}
\keyword{internal}
