% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_batch_get_report_groups}
\alias{codebuild_batch_get_report_groups}
\title{Returns an array of report groups}
\usage{
codebuild_batch_get_report_groups(reportGroupArns)
}
\arguments{
\item{reportGroupArns}{[required] An array of report group ARNs that identify the report groups to return.}
}
\description{
Returns an array of report groups.

See \url{https://paws-r.github.io/docs/codebuild/batch_get_report_groups.html} for full documentation.
}
\keyword{internal}
