% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_create_report_group}
\alias{codebuild_create_report_group}
\title{Creates a report group}
\usage{
codebuild_create_report_group(name, type, exportConfig, tags = NULL)
}
\arguments{
\item{name}{[required] The name of the report group.}

\item{type}{[required] The type of report group.}

\item{exportConfig}{[required] A \code{ReportExportConfig} object that contains information about where the
report group test results are exported.}

\item{tags}{A list of tag key and value pairs associated with this report group.

These tags are available for use by Amazon Web Services services that
support CodeBuild report group tags.}
}
\description{
Creates a report group. A report group contains a collection of reports.

See \url{https://paws-r.github.io/docs/codebuild/create_report_group.html} for full documentation.
}
\keyword{internal}
