% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_build_batch}
\alias{codebuild_delete_build_batch}
\title{Deletes a batch build}
\usage{
codebuild_delete_build_batch(id)
}
\arguments{
\item{id}{[required] The identifier of the batch build to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  statusCode = "string",
  buildsDeleted = list(
    "string"
  ),
  buildsNotDeleted = list(
    list(
      id = "string",
      statusCode = "string"
    )
  )
)
}
}
\description{
Deletes a batch build.
}
\section{Request syntax}{
\preformatted{svc$delete_build_batch(
  id = "string"
)
}
}

\keyword{internal}
