% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_project}
\alias{codebuild_delete_project}
\title{Deletes a build project}
\usage{
codebuild_delete_project(name)
}
\arguments{
\item{name}{[required] The name of the build project.}
}
\description{
Deletes a build project. When you delete a project, its builds are not deleted.

See \url{https://paws-r.github.io/docs/codebuild/delete_project.html} for full documentation.
}
\keyword{internal}
