% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_report}
\alias{codebuild_delete_report}
\title{Deletes a report}
\usage{
codebuild_delete_report(arn)
}
\arguments{
\item{arn}{[required] The ARN of the report to delete.}
}
\value{
An empty list.
}
\description{
Deletes a report.
}
\section{Request syntax}{
\preformatted{svc$delete_report(
  arn = "string"
)
}
}

\keyword{internal}
