% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_resource_policy}
\alias{codebuild_delete_resource_policy}
\title{Deletes a resource policy that is identified by its resource ARN}
\usage{
codebuild_delete_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that is associated with the resource policy.}
}
\value{
An empty list.
}
\description{
Deletes a resource policy that is identified by its resource ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_resource_policy(
  resourceArn = "string"
)
}
}

\keyword{internal}
