% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_delete_source_credentials}
\alias{codebuild_delete_source_credentials}
\title{Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source
credentials}
\usage{
codebuild_delete_source_credentials(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the token.}
}
\description{
Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials.

See \url{https://www.paws-r-sdk.com/docs/codebuild_delete_source_credentials/} for full documentation.
}
\keyword{internal}
