% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_get_report_group_trend}
\alias{codebuild_get_report_group_trend}
\title{Get report group trend}
\usage{
codebuild_get_report_group_trend(reportGroupArn, numOfReports,
  trendField)
}
\arguments{
\item{reportGroupArn}{[required]}

\item{numOfReports}{}

\item{trendField}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  stats = list(
    average = "string",
    max = "string",
    min = "string"
  ),
  rawData = list(
    list(
      reportArn = "string",
      data = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Get report group trend
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_report_group_trend(
  reportGroupArn = "string",
  numOfReports = 123,
  trendField = "PASS_RATE"|"DURATION"|"TOTAL"|"LINE_COVERAGE"|"LINES_COVERED"|"LINES_MISSED"|"BRANCH_COVERAGE"|"BRANCHES_COVERED"|"BRANCHES_MISSED"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
