% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_get_resource_policy}
\alias{codebuild_get_resource_policy}
\title{Gets a resource policy that is identified by its resource ARN}
\usage{
codebuild_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that is associated with the resource policy.}
}
\description{
Gets a resource policy that is identified by its resource ARN.

See \url{https://www.paws-r-sdk.com/docs/codebuild_get_resource_policy/} for full documentation.
}
\keyword{internal}
