% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_invalidate_project_cache}
\alias{codebuild_invalidate_project_cache}
\title{Resets the cache for a project}
\usage{
codebuild_invalidate_project_cache(projectName)
}
\arguments{
\item{projectName}{[required] The name of the CodeBuild build project that the cache is reset for.}
}
\description{
Resets the cache for a project.

See \url{https://paws-r.github.io/docs/codebuild/invalidate_project_cache.html} for full documentation.
}
\keyword{internal}
