% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_curated_environment_images}
\alias{codebuild_list_curated_environment_images}
\title{Gets information about Docker images that are managed by CodeBuild}
\usage{
codebuild_list_curated_environment_images()
}
\description{
Gets information about Docker images that are managed by CodeBuild.

See \url{https://paws-r.github.io/docs/codebuild/list_curated_environment_images.html} for full documentation.
}
\keyword{internal}
