% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_curated_environment_images}
\alias{codebuild_list_curated_environment_images}
\title{Gets information about Docker images that are managed by CodeBuild}
\usage{
codebuild_list_curated_environment_images()
}
\description{
Gets information about Docker images that are managed by CodeBuild.

See \url{https://www.paws-r-sdk.com/docs/codebuild_list_curated_environment_images/} for full documentation.
}
\keyword{internal}
