% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_list_source_credentials}
\alias{codebuild_list_source_credentials}
\title{Returns a list of SourceCredentialsInfo objects}
\usage{
codebuild_list_source_credentials()
}
\value{
A list with the following syntax:\preformatted{list(
  sourceCredentialsInfos = list(
    list(
      arn = "string",
      serverType = "GITHUB"|"BITBUCKET"|"GITHUB_ENTERPRISE",
      authType = "OAUTH"|"BASIC_AUTH"|"PERSONAL_ACCESS_TOKEN"
    )
  )
)
}
}
\description{
Returns a list of \code{SourceCredentialsInfo} objects.
}
\section{Request syntax}{
\preformatted{svc$list_source_credentials()
}
}

\keyword{internal}
