% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_put_resource_policy}
\alias{codebuild_put_resource_policy}
\title{Stores a resource policy for the ARN of a Project or ReportGroup object}
\usage{
codebuild_put_resource_policy(policy, resourceArn)
}
\arguments{
\item{policy}{[required] A JSON-formatted resource policy. For more information, see \href{https://docs.aws.amazon.com/codebuild/latest/userguide/project-sharing.html#project-sharing-share}{Sharing a Project}
and \href{https://docs.aws.amazon.com/codebuild/latest/userguide/report-groups-sharing.html#report-groups-sharing-share}{Sharing a Report Group}
in the \emph{CodeBuild User Guide}.}

\item{resourceArn}{[required] The ARN of the \code{Project} or \code{ReportGroup} resource you want to associate
with a resource policy.}
}
\description{
Stores a resource policy for the ARN of a \code{Project} or \code{ReportGroup} object.

See \url{https://paws-r.github.io/docs/codebuild/put_resource_policy.html} for full documentation.
}
\keyword{internal}
