% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_retry_build_batch}
\alias{codebuild_retry_build_batch}
\title{Restarts a failed batch build}
\usage{
codebuild_retry_build_batch(
  id = NULL,
  idempotencyToken = NULL,
  retryType = NULL
)
}
\arguments{
\item{id}{Specifies the identifier of the batch build to restart.}

\item{idempotencyToken}{A unique, case sensitive identifier you provide to ensure the
idempotency of the \code{\link[=codebuild_retry_build_batch]{retry_build_batch}}
request. The token is included in the
\code{\link[=codebuild_retry_build_batch]{retry_build_batch}} request and is valid
for five minutes. If you repeat the
\code{\link[=codebuild_retry_build_batch]{retry_build_batch}} request with the same
token, but change a parameter, CodeBuild returns a parameter mismatch
error.}

\item{retryType}{Specifies the type of retry to perform.}
}
\description{
Restarts a failed batch build. Only batch builds that have failed can be retried.

See \url{https://www.paws-r-sdk.com/docs/codebuild_retry_build_batch/} for full documentation.
}
\keyword{internal}
