% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebuild_operations.R
\name{codebuild_update_report_group}
\alias{codebuild_update_report_group}
\title{Updates a report group}
\usage{
codebuild_update_report_group(arn, exportConfig = NULL, tags = NULL)
}
\arguments{
\item{arn}{[required] The ARN of the report group to update.}

\item{exportConfig}{Used to specify an updated export type. Valid values are:
\itemize{
\item \code{S3}: The report results are exported to an S3 bucket.
\item \code{NO_EXPORT}: The report results are not exported.
}}

\item{tags}{An updated list of tag key and value pairs associated with this report
group.

These tags are available for use by Amazon Web Services services that
support CodeBuild report group tags.}
}
\description{
Updates a report group.

See \url{https://www.paws-r-sdk.com/docs/codebuild_update_report_group/} for full documentation.
}
\keyword{internal}
