% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_batch_get_commits}
\alias{codecommit_batch_get_commits}
\title{Returns information about the contents of one or more commits in a
repository}
\usage{
codecommit_batch_get_commits(commitIds, repositoryName)
}
\arguments{
\item{commitIds}{[required] The full commit IDs of the commits to get information about.

You must supply the full SHA IDs of each commit. You cannot use
shortened SHA IDs.}

\item{repositoryName}{[required] The name of the repository that contains the commits.}
}
\description{
Returns information about the contents of one or more commits in a repository.

See \url{https://paws-r.github.io/docs/codecommit/batch_get_commits.html} for full documentation.
}
\keyword{internal}
