% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_batch_get_repositories}
\alias{codecommit_batch_get_repositories}
\title{Returns information about one or more repositories}
\usage{
codecommit_batch_get_repositories(repositoryNames)
}
\arguments{
\item{repositoryNames}{[required] The names of the repositories to get information about.

The length constraint limit is for each string in the array. The array
itself can be empty.}
}
\description{
Returns information about one or more repositories.

See \url{https://paws-r.github.io/docs/codecommit/batch_get_repositories.html} for full documentation.
}
\keyword{internal}
