% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_create_branch}
\alias{codecommit_create_branch}
\title{Creates a branch in a repository and points the branch to a commit}
\usage{
codecommit_create_branch(repositoryName, branchName, commitId)
}
\arguments{
\item{repositoryName}{[required] The name of the repository in which you want to create the new branch.}

\item{branchName}{[required] The name of the new branch to create.}

\item{commitId}{[required] The ID of the commit to point the new branch to.}
}
\description{
Creates a branch in a repository and points the branch to a commit.

See \url{https://paws-r.github.io/docs/codecommit/create_branch.html} for full documentation.
}
\keyword{internal}
