% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_approval_rule_template}
\alias{codecommit_delete_approval_rule_template}
\title{Deletes a specified approval rule template}
\usage{
codecommit_delete_approval_rule_template(approvalRuleTemplateName)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the approval rule template to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  approvalRuleTemplateId = "string"
)
}
}
\description{
Deletes a specified approval rule template. Deleting a template does not
remove approval rules on pull requests already created with the
template.
}
\section{Request syntax}{
\preformatted{svc$delete_approval_rule_template(
  approvalRuleTemplateName = "string"
)
}
}

\keyword{internal}
