% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_branch}
\alias{codecommit_delete_branch}
\title{Deletes a branch from a repository, unless that branch is the default
branch for the repository}
\usage{
codecommit_delete_branch(repositoryName, branchName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branch to be deleted.}

\item{branchName}{[required] The name of the branch to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  deletedBranch = list(
    branchName = "string",
    commitId = "string"
  )
)
}
}
\description{
Deletes a branch from a repository, unless that branch is the default
branch for the repository.
}
\section{Request syntax}{
\preformatted{svc$delete_branch(
  repositoryName = "string",
  branchName = "string"
)
}
}

\keyword{internal}
