% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_comment_content}
\alias{codecommit_delete_comment_content}
\title{Deletes the content of a comment made on a change, file, or commit in a
repository}
\usage{
codecommit_delete_comment_content(commentId)
}
\arguments{
\item{commentId}{[required] The unique, system-generated ID of the comment. To get this ID, use
\code{\link[=codecommit_get_comments_for_compared_commit]{get_comments_for_compared_commit}}
or
\code{\link[=codecommit_get_comments_for_pull_request]{get_comments_for_pull_request}}.}
}
\description{
Deletes the content of a comment made on a change, file, or commit in a repository.

See \url{https://paws-r.github.io/docs/codecommit/delete_comment_content.html} for full documentation.
}
\keyword{internal}
