% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_repository}
\alias{codecommit_delete_repository}
\title{Deletes a repository}
\usage{
codecommit_delete_repository(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  repositoryId = "string"
)
}
}
\description{
Deletes a repository. If a specified repository was already deleted, a
null repository ID is returned.

Deleting a repository also deletes all associated objects and metadata.
After a repository is deleted, all future push calls to the deleted
repository fail.
}
\section{Request syntax}{
\preformatted{svc$delete_repository(
  repositoryName = "string"
)
}
}

\keyword{internal}
