% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_repository}
\alias{codecommit_delete_repository}
\title{Deletes a repository}
\usage{
codecommit_delete_repository(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to delete.}
}
\description{
Deletes a repository. If a specified repository was already deleted, a null repository ID is returned.

See \url{https://paws-r.github.io/docs/codecommit/delete_repository.html} for full documentation.
}
\keyword{internal}
