% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_approval_rule_template}
\alias{codecommit_get_approval_rule_template}
\title{Returns information about a specified approval rule template}
\usage{
codecommit_get_approval_rule_template(approvalRuleTemplateName)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the approval rule template for which you want to get
information.}
}
\description{
Returns information about a specified approval rule template.

See \url{https://www.paws-r-sdk.com/docs/codecommit_get_approval_rule_template/} for full documentation.
}
\keyword{internal}
