% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_blob}
\alias{codecommit_get_blob}
\title{Returns the base-64 encoded content of an individual blob in a
repository}
\usage{
codecommit_get_blob(repositoryName, blobId)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the blob.}

\item{blobId}{[required] The ID of the blob, which is its SHA-1 pointer.}
}
\description{
Returns the base-64 encoded content of an individual blob in a repository.

See \url{https://paws-r.github.io/docs/codecommit/get_blob.html} for full documentation.
}
\keyword{internal}
