% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_branch}
\alias{codecommit_get_branch}
\title{Returns information about a repository branch, including its name and
the last commit ID}
\usage{
codecommit_get_branch(repositoryName = NULL, branchName = NULL)
}
\arguments{
\item{repositoryName}{The name of the repository that contains the branch for which you want
to retrieve information.}

\item{branchName}{The name of the branch for which you want to retrieve information.}
}
\description{
Returns information about a repository branch, including its name and the last commit ID.

See \url{https://paws-r.github.io/docs/codecommit/get_branch.html} for full documentation.
}
\keyword{internal}
