% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_branch}
\alias{codecommit_get_branch}
\title{Returns information about a repository branch, including its name and
the last commit ID}
\usage{
codecommit_get_branch(repositoryName, branchName)
}
\arguments{
\item{repositoryName}{The name of the repository that contains the branch for which you want
to retrieve information.}

\item{branchName}{The name of the branch for which you want to retrieve information.}
}
\value{
A list with the following syntax:\preformatted{list(
  branch = list(
    branchName = "string",
    commitId = "string"
  )
)
}
}
\description{
Returns information about a repository branch, including its name and
the last commit ID.
}
\section{Request syntax}{
\preformatted{svc$get_branch(
  repositoryName = "string",
  branchName = "string"
)
}
}

\keyword{internal}
