% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_pull_request}
\alias{codecommit_get_pull_request}
\title{Gets information about a pull request in a specified repository}
\usage{
codecommit_get_pull_request(pullRequestId)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID of the pull request. To get this ID, use
\code{\link[=codecommit_list_pull_requests]{list_pull_requests}}.}
}
\description{
Gets information about a pull request in a specified repository.

See \url{https://paws-r.github.io/docs/codecommit/get_pull_request.html} for full documentation.
}
\keyword{internal}
