% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_pull_request_approval_states}
\alias{codecommit_get_pull_request_approval_states}
\title{Gets information about the approval states for a specified pull request}
\usage{
codecommit_get_pull_request_approval_states(pullRequestId, revisionId)
}
\arguments{
\item{pullRequestId}{[required] The system-generated ID for the pull request.}

\item{revisionId}{[required] The system-generated ID for the pull request revision.}
}
\description{
Gets information about the approval states for a specified pull request. Approval states only apply to pull requests that have one or more approval rules applied to them.

See \url{https://paws-r.github.io/docs/codecommit/get_pull_request_approval_states.html} for full documentation.
}
\keyword{internal}
