% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_repository}
\alias{codecommit_get_repository}
\title{Returns information about a repository}
\usage{
codecommit_get_repository(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to get information about.}
}
\description{
Returns information about a repository.

See \url{https://www.paws-r-sdk.com/docs/codecommit_get_repository/} for full documentation.
}
\keyword{internal}
