% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_get_repository_triggers}
\alias{codecommit_get_repository_triggers}
\title{Gets information about triggers configured for a repository}
\usage{
codecommit_get_repository_triggers(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository for which the trigger is configured.}
}
\description{
Gets information about triggers configured for a repository.

See \url{https://www.paws-r-sdk.com/docs/codecommit_get_repository_triggers/} for full documentation.
}
\keyword{internal}
