% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_list_branches}
\alias{codecommit_list_branches}
\title{Gets information about one or more branches in a repository}
\usage{
codecommit_list_branches(repositoryName, nextToken = NULL)
}
\arguments{
\item{repositoryName}{[required] The name of the repository that contains the branches.}

\item{nextToken}{An enumeration token that allows the operation to batch the results.}
}
\description{
Gets information about one or more branches in a repository.

See \url{https://paws-r.github.io/docs/codecommit/list_branches.html} for full documentation.
}
\keyword{internal}
