% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_list_repositories_for_approval_rule_template}
\alias{codecommit_list_repositories_for_approval_rule_template}
\title{Lists all repositories associated with the specified approval rule
template}
\usage{
codecommit_list_repositories_for_approval_rule_template(
  approvalRuleTemplateName,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the approval rule template for which you want to list
repositories that are associated with that template.}

\item{nextToken}{An enumeration token that, when provided in a request, returns the next
batch of the results.}

\item{maxResults}{A non-zero, non-negative integer used to limit the number of returned
results.}
}
\description{
Lists all repositories associated with the specified approval rule template.

See \url{https://paws-r.github.io/docs/codecommit/list_repositories_for_approval_rule_template.html} for full documentation.
}
\keyword{internal}
