% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_post_comment_reply}
\alias{codecommit_post_comment_reply}
\title{Posts a comment in reply to an existing comment on a comparison between
commits or a pull request}
\usage{
codecommit_post_comment_reply(inReplyTo, clientRequestToken = NULL, content)
}
\arguments{
\item{inReplyTo}{[required] The system-generated ID of the comment to which you want to reply. To
get this ID, use
\code{\link[=codecommit_get_comments_for_compared_commit]{get_comments_for_compared_commit}}
or
\code{\link[=codecommit_get_comments_for_pull_request]{get_comments_for_pull_request}}.}

\item{clientRequestToken}{A unique, client-generated idempotency token that, when provided in a
request, ensures the request cannot be repeated with a changed
parameter. If a request is received with the same parameters and a token
is included, the request returns information about the initial request
that used that token.}

\item{content}{[required] The contents of your reply to a comment.}
}
\description{
Posts a comment in reply to an existing comment on a comparison between commits or a pull request.

See \url{https://paws-r.github.io/docs/codecommit/post_comment_reply.html} for full documentation.
}
\keyword{internal}
