% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_approval_rule_template_description}
\alias{codecommit_update_approval_rule_template_description}
\title{Updates the description for a specified approval rule template}
\usage{
codecommit_update_approval_rule_template_description(
  approvalRuleTemplateName,
  approvalRuleTemplateDescription
)
}
\arguments{
\item{approvalRuleTemplateName}{[required] The name of the template for which you want to update the description.}

\item{approvalRuleTemplateDescription}{[required] The updated description of the approval rule template.}
}
\description{
Updates the description for a specified approval rule template.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_approval_rule_template_description/} for full documentation.
}
\keyword{internal}
