% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_approval_rule_template_name}
\alias{codecommit_update_approval_rule_template_name}
\title{Updates the name of a specified approval rule template}
\usage{
codecommit_update_approval_rule_template_name(
  oldApprovalRuleTemplateName, newApprovalRuleTemplateName)
}
\arguments{
\item{oldApprovalRuleTemplateName}{[required] The current name of the approval rule template.}

\item{newApprovalRuleTemplateName}{[required] The new name you want to apply to the approval rule template.}
}
\value{
A list with the following syntax:\preformatted{list(
  approvalRuleTemplate = list(
    approvalRuleTemplateId = "string",
    approvalRuleTemplateName = "string",
    approvalRuleTemplateDescription = "string",
    approvalRuleTemplateContent = "string",
    ruleContentSha256 = "string",
    lastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    creationDate = as.POSIXct(
      "2015-01-01"
    ),
    lastModifiedUser = "string"
  )
)
}
}
\description{
Updates the name of a specified approval rule template.
}
\section{Request syntax}{
\preformatted{svc$update_approval_rule_template_name(
  oldApprovalRuleTemplateName = "string",
  newApprovalRuleTemplateName = "string"
)
}
}

\keyword{internal}
