% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_comment}
\alias{codecommit_update_comment}
\title{Replaces the contents of a comment}
\usage{
codecommit_update_comment(commentId, content)
}
\arguments{
\item{commentId}{[required] The system-generated ID of the comment you want to update. To get this
ID, use
\code{\link[=codecommit_get_comments_for_compared_commit]{get_comments_for_compared_commit}}
or
\code{\link[=codecommit_get_comments_for_pull_request]{get_comments_for_pull_request}}.}

\item{content}{[required] The updated content to replace the existing content of the comment.}
}
\description{
Replaces the contents of a comment.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_comment/} for full documentation.
}
\keyword{internal}
