% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_default_branch}
\alias{codecommit_update_default_branch}
\title{Sets or changes the default branch name for the specified repository}
\usage{
codecommit_update_default_branch(repositoryName, defaultBranchName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to set or change the default branch for.}

\item{defaultBranchName}{[required] The name of the branch to set as the default.}
}
\description{
Sets or changes the default branch name for the specified repository.

See \url{https://www.paws-r-sdk.com/docs/codecommit_update_default_branch/} for full documentation.
}
\keyword{internal}
