% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_update_repository_description}
\alias{codecommit_update_repository_description}
\title{Sets or changes the comment or description for a repository}
\usage{
codecommit_update_repository_description(repositoryName,
  repositoryDescription)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to set or change the comment or description
for.}

\item{repositoryDescription}{The new comment or description for the specified repository. Repository
descriptions are limited to 1,000 characters.}
}
\value{
An empty list.
}
\description{
Sets or changes the comment or description for a repository.

The description field for a repository accepts all HTML characters and
all valid Unicode characters. Applications that do not HTML-encode the
description and display it in a webpage can expose users to potentially
malicious code. Make sure that you HTML-encode the description field in
any application that uses this API to display the repository description
on a webpage.
}
\section{Request syntax}{
\preformatted{svc$update_repository_description(
  repositoryName = "string",
  repositoryDescription = "string"
)
}
}

\keyword{internal}
