% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_application_revisions}
\alias{codedeploy_batch_get_application_revisions}
\title{Gets information about one or more application revisions}
\usage{
codedeploy_batch_get_application_revisions(applicationName, revisions)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application about which to get revision
information.}

\item{revisions}{[required] An array of \code{RevisionLocation} objects that specify information to get
about the application revisions, including type and location. The
maximum number of \code{RevisionLocation} objects you can specify is 25.}
}
\description{
Gets information about one or more application revisions. The maximum number of application revisions that can be returned is 25.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_batch_get_application_revisions/} for full documentation.
}
\keyword{internal}
