% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_deployment_groups}
\alias{codedeploy_batch_get_deployment_groups}
\title{Gets information about one or more deployment groups}
\usage{
codedeploy_batch_get_deployment_groups(applicationName, deploymentGroupNames)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the applicable
IAM user or AWS account.}

\item{deploymentGroupNames}{[required] The names of the deployment groups.}
}
\description{
Gets information about one or more deployment groups.

See \url{https://paws-r.github.io/docs/codedeploy/batch_get_deployment_groups.html} for full documentation.
}
\keyword{internal}
