% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_batch_get_on_premises_instances}
\alias{codedeploy_batch_get_on_premises_instances}
\title{Gets information about one or more on-premises instances}
\usage{
codedeploy_batch_get_on_premises_instances(instanceNames)
}
\arguments{
\item{instanceNames}{[required] The names of the on-premises instances about which to get information.
The maximum number of instance names you can specify is 25.}
}
\description{
Gets information about one or more on-premises instances. The maximum number of on-premises instances that can be returned is 25.

See \url{https://paws-r.github.io/docs/codedeploy/batch_get_on_premises_instances.html} for full documentation.
}
\keyword{internal}
