% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_create_application}
\alias{codedeploy_create_application}
\title{Creates an application}
\usage{
codedeploy_create_application(applicationName, computePlatform, tags)
}
\arguments{
\item{applicationName}{[required] The name of the application. This name must be unique with the
applicable IAM user or AWS account.}

\item{computePlatform}{The destination platform type for the deployment (\code{Lambda}, \code{Server}, or
\code{ECS}).}

\item{tags}{The metadata that you apply to CodeDeploy applications to help you
organize and categorize them. Each tag consists of a key and an optional
value, both of which you define.}
}
\value{
A list with the following syntax:\preformatted{list(
  applicationId = "string"
)
}
}
\description{
Creates an application.
}
\section{Request syntax}{
\preformatted{svc$create_application(
  applicationName = "string",
  computePlatform = "Server"|"Lambda"|"ECS",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
