% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_create_deployment}
\alias{codedeploy_create_deployment}
\title{Deploys an application revision through the specified deployment group}
\usage{
codedeploy_create_deployment(
  applicationName,
  deploymentGroupName = NULL,
  revision = NULL,
  deploymentConfigName = NULL,
  description = NULL,
  ignoreApplicationStopFailures = NULL,
  targetInstances = NULL,
  autoRollbackConfiguration = NULL,
  updateOutdatedInstancesOnly = NULL,
  fileExistsBehavior = NULL,
  deploymentMode = NULL,
  overrideAlarmConfiguration = NULL
)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon Web Services account.}

\item{deploymentGroupName}{The name of the deployment group.}

\item{revision}{The type and location of the revision to deploy.}

\item{deploymentConfigName}{The name of a deployment configuration associated with the user or Amazon Web Services account.

If not specified, the value configured in the deployment group is used as the default. If the deployment group does not have a deployment configuration associated with it, \code{CodeDeployDefault}.\code{OneAtATime} is used by default.}

\item{description}{A comment about the deployment.}

\item{ignoreApplicationStopFailures}{If true, then if an \code{ApplicationStop}, \code{BeforeBlockTraffic}, or \code{AfterBlockTraffic} deployment lifecycle event to an instance fails, then the deployment continues to the next deployment lifecycle event. For example, if \code{ApplicationStop} fails, the deployment continues with \code{DownloadBundle}. If \code{BeforeBlockTraffic} fails, the deployment continues with \code{BlockTraffic}. If \code{AfterBlockTraffic} fails, the deployment continues with \code{ApplicationStop}.

If false or not specified, then if a lifecycle event fails during a deployment to an instance, that deployment fails. If deployment to that instance is part of an overall deployment and the number of healthy hosts is not less than the minimum number of healthy hosts, then a deployment to the next instance is attempted.

During a deployment, the CodeDeploy agent runs the scripts specified for \code{ApplicationStop}, \code{BeforeBlockTraffic}, and \code{AfterBlockTraffic} in the AppSpec file from the previous successful deployment. (All other scripts are run from the AppSpec file in the current deployment.) If one of these scripts contains an error and does not run successfully, the deployment can fail.

If the cause of the failure is a script from the last successful deployment that will never run successfully, create a new deployment and use \code{ignoreApplicationStopFailures} to specify that the \code{ApplicationStop}, \code{BeforeBlockTraffic}, and \code{AfterBlockTraffic} failures should be ignored.}

\item{targetInstances}{Information about the instances that belong to the replacement environment in a blue/green deployment.}

\item{autoRollbackConfiguration}{Configuration information for an automatic rollback that is added when a deployment is created.}

\item{updateOutdatedInstancesOnly}{Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.}

\item{fileExistsBehavior}{Information about how CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The \code{fileExistsBehavior} parameter takes any of the following values:
\itemize{
\item DISALLOW: The deployment fails. This is also the default behavior if no option is specified.
\item OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.
\item RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.
}}

\item{deploymentMode}{The deployment mode to use for the deployment. When set to STANDARD (the default), the deployment runs the standard set of deployment lifecycle events. When set to RESTART, an EC2/On-premises in-place deployment runs a shortened set of lifecycle events to quickly restart the application on the target instances.}

\item{overrideAlarmConfiguration}{Allows you to specify information about alarms associated with a deployment. The alarm configuration that you specify here will override the alarm configuration at the deployment group level. Consider overriding the alarm configuration if you have set up alarms at the deployment group level that are causing deployment failures. In this case, you would call \code{\link[=codedeploy_create_deployment]{create_deployment}} to create a new deployment that uses a previous application revision that is known to work, and set its alarm configuration to turn off alarm polling. Turning off alarm polling ensures that the new deployment proceeds without being blocked by the alarm that was generated by the previous, failed, deployment.

If you specify an \code{overrideAlarmConfiguration}, you need the \code{\link[=codedeploy_update_deployment_group]{update_deployment_group}} IAM permission when calling \code{\link[=codedeploy_create_deployment]{create_deployment}}.}
}
\description{
Deploys an application revision through the specified deployment group.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_create_deployment/} for full documentation.
}
\keyword{internal}
