% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_application}
\alias{codedeploy_delete_application}
\title{Deletes an application}
\usage{
codedeploy_delete_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon
Web Services account.}
}
\description{
Deletes an application.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_delete_application/} for full documentation.
}
\keyword{internal}
