% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_application}
\alias{codedeploy_delete_application}
\title{Deletes an application}
\usage{
codedeploy_delete_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}
}
\description{
Deletes an application.

See \url{https://paws-r.github.io/docs/codedeploy/delete_application.html} for full documentation.
}
\keyword{internal}
