% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_deployment_config}
\alias{codedeploy_delete_deployment_config}
\title{Deletes a deployment configuration}
\usage{
codedeploy_delete_deployment_config(deploymentConfigName)
}
\arguments{
\item{deploymentConfigName}{[required] The name of a deployment configuration associated with the user or
Amazon Web Services account.}
}
\description{
Deletes a deployment configuration.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_delete_deployment_config/} for full documentation.
}
\keyword{internal}
