% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_deployment_group}
\alias{codedeploy_delete_deployment_group}
\title{Deletes a deployment group}
\usage{
codedeploy_delete_deployment_group(applicationName, deploymentGroupName)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon
Web Services account.}

\item{deploymentGroupName}{[required] The name of a deployment group for the specified application.}
}
\description{
Deletes a deployment group.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_delete_deployment_group/} for full documentation.
}
\keyword{internal}
