% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_git_hub_account_token}
\alias{codedeploy_delete_git_hub_account_token}
\title{Deletes a GitHub account connection}
\usage{
codedeploy_delete_git_hub_account_token(tokenName)
}
\arguments{
\item{tokenName}{The name of the GitHub account connection to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  tokenName = "string"
)
}
}
\description{
Deletes a GitHub account connection.
}
\section{Request syntax}{
\preformatted{svc$delete_git_hub_account_token(
  tokenName = "string"
)
}
}

\keyword{internal}
