% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_git_hub_account_token}
\alias{codedeploy_delete_git_hub_account_token}
\title{Deletes a GitHub account connection}
\usage{
codedeploy_delete_git_hub_account_token(tokenName = NULL)
}
\arguments{
\item{tokenName}{The name of the GitHub account connection to delete.}
}
\description{
Deletes a GitHub account connection.

See \url{https://paws-r.github.io/docs/codedeploy/delete_git_hub_account_token.html} for full documentation.
}
\keyword{internal}
