% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_delete_resources_by_external_id}
\alias{codedeploy_delete_resources_by_external_id}
\title{Deletes resources linked to an external ID}
\usage{
codedeploy_delete_resources_by_external_id(externalId)
}
\arguments{
\item{externalId}{The unique ID of an external resource (for example, a CloudFormation
stack ID) that is linked to one or more CodeDeploy resources.}
}
\value{
An empty list.
}
\description{
Deletes resources linked to an external ID.
}
\section{Request syntax}{
\preformatted{svc$delete_resources_by_external_id(
  externalId = "string"
)
}
}

\keyword{internal}
