% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_application}
\alias{codedeploy_get_application}
\title{Gets information about an application}
\usage{
codedeploy_get_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}
}
\value{
A list with the following syntax:\preformatted{list(
  application = list(
    applicationId = "string",
    applicationName = "string",
    createTime = as.POSIXct(
      "2015-01-01"
    ),
    linkedToGitHub = TRUE|FALSE,
    gitHubAccountName = "string",
    computePlatform = "Server"|"Lambda"|"ECS"
  )
)
}
}
\description{
Gets information about an application.
}
\section{Request syntax}{
\preformatted{svc$get_application(
  applicationName = "string"
)
}
}

\keyword{internal}
