% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_application}
\alias{codedeploy_get_application}
\title{Gets information about an application}
\usage{
codedeploy_get_application(applicationName)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon
Web Services account.}
}
\description{
Gets information about an application.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_get_application/} for full documentation.
}
\keyword{internal}
