% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment}
\alias{codedeploy_get_deployment}
\title{Gets information about a deployment}
\usage{
codedeploy_get_deployment(deploymentId)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment associated with the IAM user or Amazon Web
Services account.}
}
\description{
Gets information about a deployment.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_get_deployment/} for full documentation.
}
\keyword{internal}
