% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment_config}
\alias{codedeploy_get_deployment_config}
\title{Gets information about a deployment configuration}
\usage{
codedeploy_get_deployment_config(deploymentConfigName)
}
\arguments{
\item{deploymentConfigName}{[required] The name of a deployment configuration associated with the IAM user or
AWS account.}
}
\description{
Gets information about a deployment configuration.

See \url{https://paws-r.github.io/docs/codedeploy/get_deployment_config.html} for full documentation.
}
\keyword{internal}
