% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment_instance}
\alias{codedeploy_get_deployment_instance}
\title{Gets information about an instance as part of a deployment}
\usage{
codedeploy_get_deployment_instance(deploymentId, instanceId)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{instanceId}{[required] The unique ID of an instance in the deployment group.}
}
\description{
Gets information about an instance as part of a deployment.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_get_deployment_instance/} for full documentation.
}
\keyword{internal}
