% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment_target}
\alias{codedeploy_get_deployment_target}
\title{Returns information about a deployment target}
\usage{
codedeploy_get_deployment_target(deploymentId = NULL, targetId = NULL)
}
\arguments{
\item{deploymentId}{The unique ID of a deployment.}

\item{targetId}{The unique ID of a deployment target.}
}
\description{
Returns information about a deployment target.

See \url{https://paws-r.github.io/docs/codedeploy/get_deployment_target.html} for full documentation.
}
\keyword{internal}
