% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_on_premises_instance}
\alias{codedeploy_get_on_premises_instance}
\title{Gets information about an on-premises instance}
\usage{
codedeploy_get_on_premises_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance about which to get information.}
}
\description{
Gets information about an on-premises instance.

See \url{https://paws-r.github.io/docs/codedeploy/get_on_premises_instance.html} for full documentation.
}
\keyword{internal}
