% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_applications}
\alias{codedeploy_list_applications}
\title{Lists the applications registered with the IAM user or Amazon Web
Services account}
\usage{
codedeploy_list_applications(nextToken = NULL)
}
\arguments{
\item{nextToken}{An identifier returned from the previous list applications call. It can
be used to return the next set of applications in the list.}
}
\description{
Lists the applications registered with the IAM user or Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_list_applications/} for full documentation.
}
\keyword{internal}
