% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_deployment_configs}
\alias{codedeploy_list_deployment_configs}
\title{Lists the deployment configurations with the IAM user or Amazon Web
Services account}
\usage{
codedeploy_list_deployment_configs(nextToken = NULL)
}
\arguments{
\item{nextToken}{An identifier returned from the previous
\code{\link[=codedeploy_list_deployment_configs]{list_deployment_configs}} call. It
can be used to return the next set of deployment configurations in the
list.}
}
\description{
Lists the deployment configurations with the IAM user or Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_list_deployment_configs/} for full documentation.
}
\keyword{internal}
