% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_deployment_groups}
\alias{codedeploy_list_deployment_groups}
\title{Lists the deployment groups for an application registered with the
Amazon Web Services user or Amazon Web Services account}
\usage{
codedeploy_list_deployment_groups(applicationName, nextToken = NULL)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon
Web Services account.}

\item{nextToken}{An identifier returned from the previous list deployment groups call. It
can be used to return the next set of deployment groups in the list.}
}
\description{
Lists the deployment groups for an application registered with the Amazon Web Services user or Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_list_deployment_groups/} for full documentation.
}
\keyword{internal}
