% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_deployment_targets}
\alias{codedeploy_list_deployment_targets}
\title{Returns an array of target IDs that are associated a deployment}
\usage{
codedeploy_list_deployment_targets(
  deploymentId = NULL,
  nextToken = NULL,
  targetFilters = NULL
)
}
\arguments{
\item{deploymentId}{The unique ID of a deployment.}

\item{nextToken}{A token identifier returned from the previous
\code{\link[=codedeploy_list_deployment_targets]{list_deployment_targets}} call. It
can be used to return the next set of deployment targets in the list.}

\item{targetFilters}{A key used to filter the returned targets. The two valid values are:
\itemize{
\item \code{TargetStatus} - A \code{TargetStatus} filter string can be \code{Failed},
\code{InProgress}, \code{Pending}, \code{Ready}, \code{Skipped}, \code{Succeeded}, or
\code{Unknown}.
\item \code{ServerInstanceLabel} - A \code{ServerInstanceLabel} filter string can be
\code{Blue} or \code{Green}.
}}
}
\description{
Returns an array of target IDs that are associated a deployment.

See \url{https://paws-r.github.io/docs/codedeploy/list_deployment_targets.html} for full documentation.
}
\keyword{internal}
