% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_list_on_premises_instances}
\alias{codedeploy_list_on_premises_instances}
\title{Gets a list of names for one or more on-premises instances}
\usage{
codedeploy_list_on_premises_instances(
  registrationStatus = NULL,
  tagFilters = NULL,
  nextToken = NULL
)
}
\arguments{
\item{registrationStatus}{The registration status of the on-premises instances:
\itemize{
\item \code{Deregistered}: Include deregistered on-premises instances in the
resulting list.
\item \code{Registered}: Include registered on-premises instances in the
resulting list.
}}

\item{tagFilters}{The on-premises instance tags that are used to restrict the on-premises
instance names returned.}

\item{nextToken}{An identifier returned from the previous list on-premises instances
call. It can be used to return the next set of on-premises instances in
the list.}
}
\description{
Gets a list of names for one or more on-premises instances.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_list_on_premises_instances/} for full documentation.
}
\keyword{internal}
