% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_register_application_revision}
\alias{codedeploy_register_application_revision}
\title{Registers with CodeDeploy a revision for the specified application}
\usage{
codedeploy_register_application_revision(
  applicationName,
  description = NULL,
  revision
)
}
\arguments{
\item{applicationName}{[required] The name of an CodeDeploy application associated with the user or Amazon
Web Services account.}

\item{description}{A comment about the revision.}

\item{revision}{[required] Information about the application revision to register, including type
and location.}
}
\description{
Registers with CodeDeploy a revision for the specified application.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_register_application_revision/} for full documentation.
}
\keyword{internal}
