% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_register_on_premises_instance}
\alias{codedeploy_register_on_premises_instance}
\title{Registers an on-premises instance}
\usage{
codedeploy_register_on_premises_instance(
  instanceName,
  iamSessionArn = NULL,
  iamUserArn = NULL
)
}
\arguments{
\item{instanceName}{[required] The name of the on-premises instance to register.}

\item{iamSessionArn}{The ARN of the IAM session to associate with the on-premises instance.}

\item{iamUserArn}{The ARN of the user to associate with the on-premises instance.}
}
\description{
Registers an on-premises instance.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_register_on_premises_instance/} for full documentation.
}
\keyword{internal}
