% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_remove_tags_from_on_premises_instances}
\alias{codedeploy_remove_tags_from_on_premises_instances}
\title{Removes one or more tags from one or more on-premises instances}
\usage{
codedeploy_remove_tags_from_on_premises_instances(tags, instanceNames)
}
\arguments{
\item{tags}{[required] The tag key-value pairs to remove from the on-premises instances.}

\item{instanceNames}{[required] The names of the on-premises instances from which to remove tags.}
}
\description{
Removes one or more tags from one or more on-premises instances.

See \url{https://paws-r.github.io/docs/codedeploy/remove_tags_from_on_premises_instances.html} for full documentation.
}
\keyword{internal}
