% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_stop_deployment}
\alias{codedeploy_stop_deployment}
\title{Attempts to stop an ongoing deployment}
\usage{
codedeploy_stop_deployment(deploymentId, autoRollbackEnabled = NULL)
}
\arguments{
\item{deploymentId}{[required] The unique ID of a deployment.}

\item{autoRollbackEnabled}{Indicates, when a deployment is stopped, whether instances that have
been updated should be rolled back to the previous version of the
application revision.}
}
\description{
Attempts to stop an ongoing deployment.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_stop_deployment/} for full documentation.
}
\keyword{internal}
