% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_update_application}
\alias{codedeploy_update_application}
\title{Changes the name of an application}
\usage{
codedeploy_update_application(
  applicationName = NULL,
  newApplicationName = NULL
)
}
\arguments{
\item{applicationName}{The current name of the application you want to change.}

\item{newApplicationName}{The new name to give the application.}
}
\description{
Changes the name of an application.

See \url{https://www.paws-r-sdk.com/docs/codedeploy_update_application/} for full documentation.
}
\keyword{internal}
