% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_update_application}
\alias{codedeploy_update_application}
\title{Changes the name of an application}
\usage{
codedeploy_update_application(applicationName, newApplicationName)
}
\arguments{
\item{applicationName}{The current name of the application you want to change.}

\item{newApplicationName}{The new name to give the application.}
}
\value{
An empty list.
}
\description{
Changes the name of an application.
}
\section{Request syntax}{
\preformatted{svc$update_application(
  applicationName = "string",
  newApplicationName = "string"
)
}
}

\keyword{internal}
