% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_describe_code_review}
\alias{codegurureviewer_describe_code_review}
\title{Returns the metadata associated with the code review along with its
status}
\usage{
codegurureviewer_describe_code_review(CodeReviewArn)
}
\arguments{
\item{CodeReviewArn}{[required] The Amazon Resource Name (ARN) of the
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_CodeReview.html}{CodeReview}
object.}
}
\description{
Returns the metadata associated with the code review along with its status.

See \url{https://www.paws-r-sdk.com/docs/codegurureviewer_describe_code_review/} for full documentation.
}
\keyword{internal}
