% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurureviewer_operations.R
\name{codegurureviewer_untag_resource}
\alias{codegurureviewer_untag_resource}
\title{Removes a tag from an associated repository}
\usage{
codegurureviewer_untag_resource(resourceArn, TagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the
\href{https://docs.aws.amazon.com/codeguru/latest/reviewer-api/API_RepositoryAssociation.html}{RepositoryAssociation}
object. You can retrieve this ARN by calling
\code{\link[=codegurureviewer_list_repository_associations]{list_repository_associations}}.}

\item{TagKeys}{[required] A list of the keys for each tag you want to remove from an associated
repository.}
}
\description{
Removes a tag from an associated repository.

See \url{https://www.paws-r-sdk.com/docs/codegurureviewer_untag_resource/} for full documentation.
}
\keyword{internal}
