% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegurusecurity_operations.R
\name{codegurusecurity_update_account_configuration}
\alias{codegurusecurity_update_account_configuration}
\title{Use to update the encryption configuration for an account}
\usage{
codegurusecurity_update_account_configuration(encryptionConfig)
}
\arguments{
\item{encryptionConfig}{[required] The customer-managed KMS key ARN you want to use for encryption. If not
specified, CodeGuru Security will use an AWS-managed key for encryption.
If you previously specified a customer-managed KMS key and want CodeGuru
Security to use an AWS-managed key for encryption instead, pass nothing.}
}
\description{
Use to update the encryption configuration for an account.

See \url{https://www.paws-r-sdk.com/docs/codegurusecurity_update_account_configuration/} for full documentation.
}
\keyword{internal}
