% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_delete_pipeline}
\alias{codepipeline_delete_pipeline}
\title{Deletes the specified pipeline}
\usage{
codepipeline_delete_pipeline(name)
}
\arguments{
\item{name}{[required] The name of the pipeline to be deleted.}
}
\description{
Deletes the specified pipeline.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_delete_pipeline/} for full documentation.
}
\keyword{internal}
