% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_get_third_party_job_details}
\alias{codepipeline_get_third_party_job_details}
\title{Requests the details of a job for a third party action}
\usage{
codepipeline_get_third_party_job_details(jobId, clientToken)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID used for identifying the job.}

\item{clientToken}{[required] The clientToken portion of the clientId and clientToken pair used to
verify that the calling entity is allowed access to the job and its
details.}
}
\description{
Requests the details of a job for a third party action. Used for partner actions only.

See \url{https://paws-r.github.io/docs/codepipeline/get_third_party_job_details.html} for full documentation.
}
\keyword{internal}
