% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_action_types}
\alias{codepipeline_list_action_types}
\title{Gets a summary of all CodePipeline action types associated with your
account}
\usage{
codepipeline_list_action_types(
  actionOwnerFilter = NULL,
  nextToken = NULL,
  regionFilter = NULL
)
}
\arguments{
\item{actionOwnerFilter}{Filters the list of action types to those created by a specified entity.}

\item{nextToken}{An identifier that was returned from the previous list action types
call, which can be used to return the next set of action types in the
list.}

\item{regionFilter}{The Region to filter on for the list of action types.}
}
\description{
Gets a summary of all CodePipeline action types associated with your account.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_list_action_types/} for full documentation.
}
\keyword{internal}
