% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_pipeline_executions}
\alias{codepipeline_list_pipeline_executions}
\title{Gets a summary of the most recent executions for a pipeline}
\usage{
codepipeline_list_pipeline_executions(
  pipelineName,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline for which you want to get execution summary
information.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned nextToken
value. Pipeline history is limited to the most recent 12 months, based
on pipeline execution start times. Default value is 100.}

\item{nextToken}{The token that was returned from the previous
\code{\link[=codepipeline_list_pipeline_executions]{list_pipeline_executions}}
call, which can be used to return the next set of pipeline executions in
the list.}
}
\description{
Gets a summary of the most recent executions for a pipeline.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_list_pipeline_executions/} for full documentation.
}
\keyword{internal}
