% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_list_pipelines}
\alias{codepipeline_list_pipelines}
\title{Gets a summary of all of the pipelines associated with your account}
\usage{
codepipeline_list_pipelines(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous list pipelines call.
It can be used to return the next set of pipelines in the list.}

\item{maxResults}{The maximum number of pipelines to return in a single call. To retrieve
the remaining pipelines, make another call with the returned nextToken
value. The minimum value you can specify is 1. The maximum accepted
value is 1000.}
}
\description{
Gets a summary of all of the pipelines associated with your account.

See \url{https://paws-r.github.io/docs/codepipeline/list_pipelines.html} for full documentation.
}
\keyword{internal}
