% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_poll_for_third_party_jobs}
\alias{codepipeline_poll_for_third_party_jobs}
\title{Determines whether there are any third party jobs for a job worker to
act on}
\usage{
codepipeline_poll_for_third_party_jobs(actionTypeId, maxBatchSize = NULL)
}
\arguments{
\item{actionTypeId}{[required] Represents information about an action type.}

\item{maxBatchSize}{The maximum number of jobs to return in a poll for jobs call.}
}
\description{
Determines whether there are any third party jobs for a job worker to act on. Used for partner actions only.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_poll_for_third_party_jobs/} for full documentation.
}
\keyword{internal}
