% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_register_webhook_with_third_party}
\alias{codepipeline_register_webhook_with_third_party}
\title{Configures a connection between the webhook that was created and the
external tool with events to be detected}
\usage{
codepipeline_register_webhook_with_third_party(webhookName = NULL)
}
\arguments{
\item{webhookName}{The name of an existing webhook created with PutWebhook to register with
a supported third party.}
}
\description{
Configures a connection between the webhook that was created and the external tool with events to be detected.

See \url{https://paws-r.github.io/docs/codepipeline/register_webhook_with_third_party.html} for full documentation.
}
\keyword{internal}
