% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_start_pipeline_execution}
\alias{codepipeline_start_pipeline_execution}
\title{Starts the specified pipeline}
\usage{
codepipeline_start_pipeline_execution(name, clientRequestToken)
}
\arguments{
\item{name}{[required] The name of the pipeline to start.}

\item{clientRequestToken}{The system-generated unique ID used to identify a unique execution
request.}
}
\value{
A list with the following syntax:\preformatted{list(
  pipelineExecutionId = "string"
)
}
}
\description{
Starts the specified pipeline. Specifically, it begins processing the
latest commit to the source location specified as part of the pipeline.
}
\section{Request syntax}{
\preformatted{svc$start_pipeline_execution(
  name = "string",
  clientRequestToken = "string"
)
}
}

\keyword{internal}
