% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_untag_resource}
\alias{codepipeline_untag_resource}
\title{Removes tags from an AWS resource}
\usage{
codepipeline_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove tags from.}

\item{tagKeys}{[required] The list of keys for the tags to be removed from the resource.}
}
\value{
An empty list.
}
\description{
Removes tags from an AWS resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
