% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_update_action_type}
\alias{codepipeline_update_action_type}
\title{Updates an action type that was created with any supported integration
model, where the action type is to be used by customers of the action
type provider}
\usage{
codepipeline_update_action_type(actionType)
}
\arguments{
\item{actionType}{[required] The action type definition for the action type to be updated.}
}
\description{
Updates an action type that was created with any supported integration model, where the action type is to be used by customers of the action type provider. Use a JSON file with the action definition and \code{\link[=codepipeline_update_action_type]{update_action_type}} to provide the full structure.

See \url{https://www.paws-r-sdk.com/docs/codepipeline_update_action_type/} for full documentation.
}
\keyword{internal}
