% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_delete_user_profile}
\alias{codestar_delete_user_profile}
\title{Deletes a user profile in AWS CodeStar, including all personal
preference data associated with that profile, such as display name and
email address}
\usage{
codestar_delete_user_profile(userArn)
}
\arguments{
\item{userArn}{[required] The Amazon Resource Name (ARN) of the user to delete from AWS CodeStar.}
}
\value{
A list with the following syntax:\preformatted{list(
  userArn = "string"
)
}
}
\description{
Deletes a user profile in AWS CodeStar, including all personal
preference data associated with that profile, such as display name and
email address. It does not delete the history of that user, for example
the history of commits made by that user.
}
\section{Request syntax}{
\preformatted{svc$delete_user_profile(
  userArn = "string"
)
}
}

\keyword{internal}
