% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_describe_project}
\alias{codestar_describe_project}
\title{Describes a project and its resources}
\usage{
codestar_describe_project(id)
}
\arguments{
\item{id}{[required] The ID of the project.}
}
\description{
Describes a project and its resources.

See \url{https://www.paws-r-sdk.com/docs/codestar_describe_project/} for full documentation.
}
\keyword{internal}
