% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_describe_user_profile}
\alias{codestar_describe_user_profile}
\title{Describes a user in AWS CodeStar and the user attributes across all
projects}
\usage{
codestar_describe_user_profile(userArn)
}
\arguments{
\item{userArn}{[required] The Amazon Resource Name (ARN) of the user.}
}
\description{
Describes a user in AWS CodeStar and the user attributes across all projects.

See \url{https://www.paws-r-sdk.com/docs/codestar_describe_user_profile/} for full documentation.
}
\keyword{internal}
