% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_list_tags_for_project}
\alias{codestar_list_tags_for_project}
\title{Gets the tags for a project}
\usage{
codestar_list_tags_for_project(id, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{id}{[required] The ID of the project to get tags for.}

\item{nextToken}{Reserved for future use.}

\item{maxResults}{Reserved for future use.}
}
\description{
Gets the tags for a project.

See \url{https://www.paws-r-sdk.com/docs/codestar_list_tags_for_project/} for full documentation.
}
\keyword{internal}
