% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_list_user_profiles}
\alias{codestar_list_user_profiles}
\title{Lists all the user profiles configured for your AWS account in AWS
CodeStar}
\usage{
codestar_list_user_profiles(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The continuation token for the next set of results, if the results
cannot be returned in one response.}

\item{maxResults}{The maximum number of results to return in a response.}
}
\description{
Lists all the user profiles configured for your AWS account in AWS CodeStar.

See \url{https://www.paws-r-sdk.com/docs/codestar_list_user_profiles/} for full documentation.
}
\keyword{internal}
