% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestar_operations.R
\name{codestar_update_project}
\alias{codestar_update_project}
\title{Updates a project in AWS CodeStar}
\usage{
codestar_update_project(id, name = NULL, description = NULL)
}
\arguments{
\item{id}{[required] The ID of the project you want to update.}

\item{name}{The name of the project you want to update.}

\item{description}{The description of the project, if any.}
}
\description{
Updates a project in AWS CodeStar.

See \url{https://www.paws-r-sdk.com/docs/codestar_update_project/} for full documentation.
}
\keyword{internal}
