% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarconnections_operations.R
\name{codestarconnections_update_sync_configuration}
\alias{codestarconnections_update_sync_configuration}
\title{Updates the sync configuration for your connection and a specified
external Git repository}
\usage{
codestarconnections_update_sync_configuration(
  Branch = NULL,
  ConfigFile = NULL,
  RepositoryLinkId = NULL,
  ResourceName,
  RoleArn = NULL,
  SyncType,
  PublishDeploymentStatus = NULL,
  TriggerResourceUpdateOn = NULL
)
}
\arguments{
\item{Branch}{The branch for the sync configuration to be updated.}

\item{ConfigFile}{The configuration file for the sync configuration to be updated.}

\item{RepositoryLinkId}{The ID of the repository link for the sync configuration to be updated.}

\item{ResourceName}{[required] The name of the Amazon Web Services resource for the sync configuration
to be updated.}

\item{RoleArn}{The ARN of the IAM role for the sync configuration to be updated.}

\item{SyncType}{[required] The sync type for the sync configuration to be updated.}

\item{PublishDeploymentStatus}{Whether to enable or disable publishing of deployment status to source
providers.}

\item{TriggerResourceUpdateOn}{When to trigger Git sync to begin the stack update.}
}
\description{
Updates the sync configuration for your connection and a specified external Git repository.

See \url{https://www.paws-r-sdk.com/docs/codestarconnections_update_sync_configuration/} for full documentation.
}
\keyword{internal}
