% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestarnotifications_operations.R
\name{codestarnotifications_delete_target}
\alias{codestarnotifications_delete_target}
\title{Deletes a specified target for notifications}
\usage{
codestarnotifications_delete_target(TargetAddress, ForceUnsubscribeAll = NULL)
}
\arguments{
\item{TargetAddress}{[required] The Amazon Resource Name (ARN) of the Amazon Q Developer in chat
applications topic or Amazon Q Developer in chat applications client to
delete.}

\item{ForceUnsubscribeAll}{A Boolean value that can be used to delete all associations with this
Amazon Q Developer in chat applications topic. The default value is
FALSE. If set to TRUE, all associations between that target and every
notification rule in your Amazon Web Services account are deleted.}
}
\description{
Deletes a specified target for notifications.

See \url{https://www.paws-r-sdk.com/docs/codestarnotifications_delete_target/} for full documentation.
}
\keyword{internal}
