% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_describe_recovery_snapshots}
\alias{drs_describe_recovery_snapshots}
\title{Lists all Recovery Snapshots for a single Source Server}
\usage{
drs_describe_recovery_snapshots(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL,
  order = NULL,
  sourceServerID
)
}
\arguments{
\item{filters}{A set of filters by which to return Recovery Snapshots.}

\item{maxResults}{Maximum number of Recovery Snapshots to retrieve.}

\item{nextToken}{The token of the next Recovery Snapshot to retrieve.}

\item{order}{The sorted ordering by which to return Recovery Snapshots.}

\item{sourceServerID}{[required] Filter Recovery Snapshots by Source Server ID.}
}
\description{
Lists all Recovery Snapshots for a single Source Server.

See \url{https://www.paws-r-sdk.com/docs/drs_describe_recovery_snapshots/} for full documentation.
}
\keyword{internal}
