% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_initialize_service}
\alias{drs_initialize_service}
\title{Initialize Elastic Disaster Recovery}
\usage{
drs_initialize_service()
}
\description{
Initialize Elastic Disaster Recovery.

See \url{https://www.paws-r-sdk.com/docs/drs_initialize_service/} for full documentation.
}
\keyword{internal}
