% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_list_tags_for_resource}
\alias{drs_list_tags_for_resource}
\title{List all tags for your Elastic Disaster Recovery resources}
\usage{
drs_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags should be returned.}
}
\description{
List all tags for your Elastic Disaster Recovery resources.

See \url{https://www.paws-r-sdk.com/docs/drs_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
