% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_put_launch_action}
\alias{drs_put_launch_action}
\title{Puts a resource launch action}
\usage{
drs_put_launch_action(
  resourceId,
  actionCode,
  order,
  actionId,
  optional,
  active,
  name,
  actionVersion,
  category,
  parameters = NULL,
  description
)
}
\arguments{
\item{resourceId}{[required]}

\item{actionCode}{[required] Launch action code.}

\item{order}{[required]}

\item{actionId}{[required]}

\item{optional}{[required] Whether the launch will not be marked as failed if this action fails.}

\item{active}{[required] Whether the launch action is active.}

\item{name}{[required]}

\item{actionVersion}{[required]}

\item{category}{[required]}

\item{description}{[required]}
}
\description{
Puts a resource launch action.

See \url{https://www.paws-r-sdk.com/docs/drs_put_launch_action/} for full documentation.
}
\keyword{internal}
