% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_retry_data_replication}
\alias{drs_retry_data_replication}
\title{WARNING: RetryDataReplication is deprecated}
\usage{
drs_retry_data_replication(sourceServerID)
}
\arguments{
\item{sourceServerID}{[required] The ID of the Source Server whose data replication should be retried.}
}
\description{
WARNING: RetryDataReplication is deprecated. Causes the data replication initiation sequence to begin immediately upon next Handshake for the specified Source Server ID, regardless of when the previous initiation started. This command will work only if the Source Server is stalled or is in a DISCONNECTED or STOPPED state.

See \url{https://www.paws-r-sdk.com/docs/drs_retry_data_replication/} for full documentation.
}
\keyword{internal}
