% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs_operations.R
\name{drs_update_launch_configuration_template}
\alias{drs_update_launch_configuration_template}
\title{Updates an existing Launch Configuration Template by ID}
\usage{
drs_update_launch_configuration_template(
  copyPrivateIp = NULL,
  copyTags = NULL,
  exportBucketArn = NULL,
  launchConfigurationTemplateID,
  launchDisposition = NULL,
  licensing = NULL,
  targetInstanceTypeRightSizingMethod = NULL
)
}
\arguments{
\item{copyPrivateIp}{Copy private IP.}

\item{copyTags}{Copy tags.}

\item{exportBucketArn}{S3 bucket ARN to export Source Network templates.}

\item{launchConfigurationTemplateID}{[required] Launch Configuration Template ID.}

\item{launchDisposition}{Launch disposition.}

\item{licensing}{Licensing.}

\item{targetInstanceTypeRightSizingMethod}{Target instance type right-sizing method.}
}
\description{
Updates an existing Launch Configuration Template by ID.

See \url{https://www.paws-r-sdk.com/docs/drs_update_launch_configuration_template/} for full documentation.
}
\keyword{internal}
