% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis_operations.R
\name{fis_list_experiments}
\alias{fis_list_experiments}
\title{Lists your experiments}
\usage{
fis_list_experiments(
  maxResults = NULL,
  nextToken = NULL,
  experimentTemplateId = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}

\item{nextToken}{The token for the next page of results.}

\item{experimentTemplateId}{The ID of the experiment template.}
}
\description{
Lists your experiments.

See \url{https://www.paws-r-sdk.com/docs/fis_list_experiments/} for full documentation.
}
\keyword{internal}
