% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_associate_lenses}
\alias{wellarchitected_associate_lenses}
\title{Associate a lens to a workload}
\usage{
wellarchitected_associate_lenses(WorkloadId, LensAliases)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAliases}{[required]}
}
\description{
Associate a lens to a workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_associate_lenses/} for full documentation.
}
\keyword{internal}
