% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_lens_share}
\alias{wellarchitected_create_lens_share}
\title{Create a lens share}
\usage{
wellarchitected_create_lens_share(LensAlias, SharedWith, ClientRequestToken)
}
\arguments{
\item{LensAlias}{[required]}

\item{SharedWith}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a lens share.

See \url{https://paws-r.github.io/docs/wellarchitected/create_lens_share.html} for full documentation.
}
\keyword{internal}
