% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_milestone}
\alias{wellarchitected_create_milestone}
\title{Create a milestone for an existing workload}
\usage{
wellarchitected_create_milestone(WorkloadId, MilestoneName, ClientRequestToken)
}
\arguments{
\item{WorkloadId}{[required]}

\item{MilestoneName}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a milestone for an existing workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_milestone/} for full documentation.
}
\keyword{internal}
