% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_profile}
\alias{wellarchitected_create_profile}
\title{Create a profile}
\usage{
wellarchitected_create_profile(
  ProfileName,
  ProfileDescription,
  ProfileQuestions,
  ClientRequestToken,
  Tags = NULL
)
}
\arguments{
\item{ProfileName}{[required] Name of the profile.}

\item{ProfileDescription}{[required] The profile description.}

\item{ProfileQuestions}{[required] The profile questions.}

\item{ClientRequestToken}{[required]}

\item{Tags}{The tags assigned to the profile.}
}
\description{
Create a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_profile/} for full documentation.
}
\keyword{internal}
