% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_profile_share}
\alias{wellarchitected_create_profile_share}
\title{Create a profile share}
\usage{
wellarchitected_create_profile_share(
  ProfileArn,
  SharedWith,
  ClientRequestToken
)
}
\arguments{
\item{ProfileArn}{[required] The profile ARN.}

\item{SharedWith}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Create a profile share.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_profile_share/} for full documentation.
}
\keyword{internal}
