% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_create_review_template}
\alias{wellarchitected_create_review_template}
\title{Create a review template}
\usage{
wellarchitected_create_review_template(
  TemplateName,
  Description,
  Lenses,
  Notes = NULL,
  Tags = NULL,
  ClientRequestToken
)
}
\arguments{
\item{TemplateName}{[required] Name of the review template.}

\item{Description}{[required] The review template description.}

\item{Lenses}{[required] Lenses applied to the review template.}

\item{Notes}{}

\item{Tags}{The tags assigned to the review template.}

\item{ClientRequestToken}{[required]}
}
\description{
Create a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_create_review_template/} for full documentation.
}
\keyword{internal}
