% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_delete_lens}
\alias{wellarchitected_delete_lens}
\title{Delete an existing lens}
\usage{
wellarchitected_delete_lens(LensAlias, ClientRequestToken, LensStatus)
}
\arguments{
\item{LensAlias}{[required]}

\item{ClientRequestToken}{[required]}

\item{LensStatus}{[required] The status of the lens to be deleted.}
}
\description{
Delete an existing lens.

See \url{https://paws-r.github.io/docs/wellarchitected/delete_lens.html} for full documentation.
}
\keyword{internal}
