% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_delete_workload}
\alias{wellarchitected_delete_workload}
\title{Delete an existing workload}
\usage{
wellarchitected_delete_workload(WorkloadId, ClientRequestToken)
}
\arguments{
\item{WorkloadId}{[required]}

\item{ClientRequestToken}{[required]}
}
\description{
Delete an existing workload.

See \url{https://paws-r.github.io/docs/wellarchitected/delete_workload.html} for full documentation.
}
\keyword{internal}
