% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_disassociate_lenses}
\alias{wellarchitected_disassociate_lenses}
\title{Disassociate a lens from a workload}
\usage{
wellarchitected_disassociate_lenses(WorkloadId, LensAliases)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAliases}{[required]}
}
\description{
Disassociate a lens from a workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_disassociate_lenses/} for full documentation.
}
\keyword{internal}
