% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_answer}
\alias{wellarchitected_get_answer}
\title{Get the answer to a specific question in a workload review}
\usage{
wellarchitected_get_answer(
  WorkloadId,
  LensAlias,
  QuestionId,
  MilestoneNumber = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}

\item{MilestoneNumber}{}
}
\description{
Get the answer to a specific question in a workload review.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_answer/} for full documentation.
}
\keyword{internal}
