% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens}
\alias{wellarchitected_get_lens}
\title{Get an existing lens}
\usage{
wellarchitected_get_lens(LensAlias, LensVersion = NULL)
}
\arguments{
\item{LensAlias}{[required]}

\item{LensVersion}{The lens version to be retrieved.}
}
\description{
Get an existing lens.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_lens/} for full documentation.
}
\keyword{internal}
