% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_lens_review}
\alias{wellarchitected_get_lens_review}
\title{Get lens review}
\usage{
wellarchitected_get_lens_review(WorkloadId, LensAlias, MilestoneNumber = NULL)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{MilestoneNumber}{}
}
\description{
Get lens review.

See \url{https://paws-r.github.io/docs/wellarchitected/get_lens_review.html} for full documentation.
}
\keyword{internal}
