% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_review_template_answer}
\alias{wellarchitected_get_review_template_answer}
\title{Get review template answer}
\usage{
wellarchitected_get_review_template_answer(TemplateArn, LensAlias, QuestionId)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}
}
\description{
Get review template answer.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_review_template_answer/} for full documentation.
}
\keyword{internal}
