% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_get_review_template_lens_review}
\alias{wellarchitected_get_review_template_lens_review}
\title{Get a lens review associated with a review template}
\usage{
wellarchitected_get_review_template_lens_review(TemplateArn, LensAlias)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{LensAlias}{[required]}
}
\description{
Get a lens review associated with a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_get_review_template_lens_review/} for full documentation.
}
\keyword{internal}
