% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_import_lens}
\alias{wellarchitected_import_lens}
\title{Import a new custom lens or update an existing custom lens}
\usage{
wellarchitected_import_lens(
  LensAlias = NULL,
  JSONString,
  ClientRequestToken,
  Tags = NULL
)
}
\arguments{
\item{LensAlias}{}

\item{JSONString}{[required] The JSON representation of a lens.}

\item{ClientRequestToken}{[required]}

\item{Tags}{Tags to associate to a lens.}
}
\description{
Import a new custom lens or update an existing custom lens.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_import_lens/} for full documentation.
}
\keyword{internal}
