% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_answers}
\alias{wellarchitected_list_answers}
\title{List of answers for a particular workload and lens}
\usage{
wellarchitected_list_answers(
  WorkloadId,
  LensAlias,
  PillarId = NULL,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  QuestionPriority = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{QuestionPriority}{The priority of the question.}
}
\description{
List of answers for a particular workload and lens.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_answers/} for full documentation.
}
\keyword{internal}
