% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_review_improvements}
\alias{wellarchitected_list_lens_review_improvements}
\title{List lens review improvements}
\usage{
wellarchitected_list_lens_review_improvements(
  WorkloadId,
  LensAlias,
  PillarId = NULL,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List lens review improvements.

See \url{https://paws-r.github.io/docs/wellarchitected/list_lens_review_improvements.html} for full documentation.
}
\keyword{internal}
