% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_reviews}
\alias{wellarchitected_list_lens_reviews}
\title{List lens reviews for a particular workload}
\usage{
wellarchitected_list_lens_reviews(
  WorkloadId,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List lens reviews for a particular workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_lens_reviews/} for full documentation.
}
\keyword{internal}
