% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_reviews}
\alias{wellarchitected_list_lens_reviews}
\title{List lens reviews}
\usage{
wellarchitected_list_lens_reviews(
  WorkloadId,
  MilestoneNumber = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{MilestoneNumber}{}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List lens reviews.

See \url{https://paws-r.github.io/docs/wellarchitected/list_lens_reviews.html} for full documentation.
}
\keyword{internal}
