% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lens_shares}
\alias{wellarchitected_list_lens_shares}
\title{List the lens shares associated with the lens}
\usage{
wellarchitected_list_lens_shares(
  LensAlias,
  SharedWithPrefix = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Status = NULL
)
}
\arguments{
\item{LensAlias}{[required]}

\item{SharedWithPrefix}{The Amazon Web Services account ID or IAM role with which the lens is
shared.}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}

\item{Status}{}
}
\description{
List the lens shares associated with the lens.

See \url{https://paws-r.github.io/docs/wellarchitected/list_lens_shares.html} for full documentation.
}
\keyword{internal}
