% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_lenses}
\alias{wellarchitected_list_lenses}
\title{List the available lenses}
\usage{
wellarchitected_list_lenses(
  NextToken = NULL,
  MaxResults = NULL,
  LensType = NULL,
  LensStatus = NULL,
  LensName = NULL
)
}
\arguments{
\item{NextToken}{}

\item{MaxResults}{}

\item{LensType}{The type of lenses to be returned.}

\item{LensStatus}{The status of lenses to be returned.}

\item{LensName}{}
}
\description{
List the available lenses.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_lenses/} for full documentation.
}
\keyword{internal}
