% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_notifications}
\alias{wellarchitected_list_notifications}
\title{List lens notifications}
\usage{
wellarchitected_list_notifications(
  WorkloadId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List lens notifications.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_notifications/} for full documentation.
}
\keyword{internal}
