% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_profile_notifications}
\alias{wellarchitected_list_profile_notifications}
\title{List profile notifications}
\usage{
wellarchitected_list_profile_notifications(
  WorkloadId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadId}{}

\item{NextToken}{}

\item{MaxResults}{}
}
\description{
List profile notifications.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_profile_notifications/} for full documentation.
}
\keyword{internal}
