% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_review_template_answers}
\alias{wellarchitected_list_review_template_answers}
\title{List the answers of a review template}
\usage{
wellarchitected_list_review_template_answers(
  TemplateArn,
  LensAlias,
  PillarId = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The ARN of the review template.}

\item{LensAlias}{[required]}

\item{PillarId}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
List the answers of a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_review_template_answers/} for full documentation.
}
\keyword{internal}
