% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_tags_for_resource}
\alias{wellarchitected_list_tags_for_resource}
\title{List the tags for a resource}
\usage{
wellarchitected_list_tags_for_resource(WorkloadArn)
}
\arguments{
\item{WorkloadArn}{[required]}
}
\description{
List the tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
