% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_list_workloads}
\alias{wellarchitected_list_workloads}
\title{Paginated list of workloads}
\usage{
wellarchitected_list_workloads(
  WorkloadNamePrefix = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{WorkloadNamePrefix}{}

\item{NextToken}{}

\item{MaxResults}{The maximum number of results to return for this request.}
}
\description{
Paginated list of workloads.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_list_workloads/} for full documentation.
}
\keyword{internal}
