% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_integration}
\alias{wellarchitected_update_integration}
\title{Update integration features}
\usage{
wellarchitected_update_integration(
  WorkloadId,
  ClientRequestToken,
  IntegratingService
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{ClientRequestToken}{[required]}

\item{IntegratingService}{[required] Which integrated service to update.}
}
\description{
Update integration features.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_integration/} for full documentation.
}
\keyword{internal}
