% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_lens_review}
\alias{wellarchitected_update_lens_review}
\title{Update lens review}
\usage{
wellarchitected_update_lens_review(
  WorkloadId,
  LensAlias,
  LensNotes = NULL,
  PillarNotes = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{LensNotes}{}

\item{PillarNotes}{}
}
\description{
Update lens review.

See \url{https://paws-r.github.io/docs/wellarchitected/update_lens_review.html} for full documentation.
}
\keyword{internal}
