% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_lens_review}
\alias{wellarchitected_update_lens_review}
\title{Update lens review for a particular workload}
\usage{
wellarchitected_update_lens_review(
  WorkloadId,
  LensAlias,
  LensNotes = NULL,
  PillarNotes = NULL,
  JiraConfiguration = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{LensNotes}{}

\item{PillarNotes}{}

\item{JiraConfiguration}{Configuration of the Jira integration.}
}
\description{
Update lens review for a particular workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_lens_review/} for full documentation.
}
\keyword{internal}
