% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_review_template_answer}
\alias{wellarchitected_update_review_template_answer}
\title{Update a review template answer}
\usage{
wellarchitected_update_review_template_answer(
  TemplateArn,
  LensAlias,
  QuestionId,
  SelectedChoices = NULL,
  ChoiceUpdates = NULL,
  Notes = NULL,
  IsApplicable = NULL,
  Reason = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The review template ARN.}

\item{LensAlias}{[required]}

\item{QuestionId}{[required]}

\item{SelectedChoices}{}

\item{ChoiceUpdates}{A list of choices to be updated.}

\item{Notes}{}

\item{IsApplicable}{}

\item{Reason}{The update reason.}
}
\description{
Update a review template answer.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_review_template_answer/} for full documentation.
}
\keyword{internal}
