% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_share_invitation}
\alias{wellarchitected_update_share_invitation}
\title{Update a workload or custom lens share invitation}
\usage{
wellarchitected_update_share_invitation(
  ShareInvitationId,
  ShareInvitationAction
)
}
\arguments{
\item{ShareInvitationId}{[required] The ID assigned to the share invitation.}

\item{ShareInvitationAction}{[required]}
}
\description{
Update a workload or custom lens share invitation.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_share_invitation/} for full documentation.
}
\keyword{internal}
