% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_workload}
\alias{wellarchitected_update_workload}
\title{Update an existing workload}
\usage{
wellarchitected_update_workload(
  WorkloadId,
  WorkloadName = NULL,
  Description = NULL,
  Environment = NULL,
  AccountIds = NULL,
  AwsRegions = NULL,
  NonAwsRegions = NULL,
  PillarPriorities = NULL,
  ArchitecturalDesign = NULL,
  ReviewOwner = NULL,
  IsReviewOwnerUpdateAcknowledged = NULL,
  IndustryType = NULL,
  Industry = NULL,
  Notes = NULL,
  ImprovementStatus = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{WorkloadName}{}

\item{Description}{}

\item{Environment}{}

\item{AccountIds}{}

\item{AwsRegions}{}

\item{NonAwsRegions}{}

\item{PillarPriorities}{}

\item{ArchitecturalDesign}{}

\item{ReviewOwner}{}

\item{IsReviewOwnerUpdateAcknowledged}{Flag indicating whether the workload owner has acknowledged that the
\emph{Review owner} field is required.

If a \strong{Review owner} is not added to the workload within 60 days of
acknowledgement, access to the workload is restricted until an owner is
added.}

\item{IndustryType}{}

\item{Industry}{}

\item{Notes}{}

\item{ImprovementStatus}{}
}
\description{
Update an existing workload.

See \url{https://paws-r.github.io/docs/wellarchitected/update_workload.html} for full documentation.
}
\keyword{internal}
