% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_update_workload_share}
\alias{wellarchitected_update_workload_share}
\title{Update a workload share}
\usage{
wellarchitected_update_workload_share(ShareId, WorkloadId, PermissionType)
}
\arguments{
\item{ShareId}{[required]}

\item{WorkloadId}{[required]}

\item{PermissionType}{[required]}
}
\description{
Update a workload share.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_update_workload_share/} for full documentation.
}
\keyword{internal}
