% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_upgrade_lens_review}
\alias{wellarchitected_upgrade_lens_review}
\title{Upgrade lens review for a particular workload}
\usage{
wellarchitected_upgrade_lens_review(
  WorkloadId,
  LensAlias,
  MilestoneName,
  ClientRequestToken = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{MilestoneName}{[required]}

\item{ClientRequestToken}{}
}
\description{
Upgrade lens review for a particular workload.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_upgrade_lens_review/} for full documentation.
}
\keyword{internal}
