% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_upgrade_lens_review}
\alias{wellarchitected_upgrade_lens_review}
\title{Upgrade lens review}
\usage{
wellarchitected_upgrade_lens_review(
  WorkloadId,
  LensAlias,
  MilestoneName,
  ClientRequestToken = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{LensAlias}{[required]}

\item{MilestoneName}{[required]}

\item{ClientRequestToken}{}
}
\description{
Upgrade lens review.

See \url{https://paws-r.github.io/docs/wellarchitected/upgrade_lens_review.html} for full documentation.
}
\keyword{internal}
