% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_upgrade_profile_version}
\alias{wellarchitected_upgrade_profile_version}
\title{Upgrade a profile}
\usage{
wellarchitected_upgrade_profile_version(
  WorkloadId,
  ProfileArn,
  MilestoneName = NULL,
  ClientRequestToken = NULL
)
}
\arguments{
\item{WorkloadId}{[required]}

\item{ProfileArn}{[required] The profile ARN.}

\item{MilestoneName}{}

\item{ClientRequestToken}{}
}
\description{
Upgrade a profile.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_upgrade_profile_version/} for full documentation.
}
\keyword{internal}
