% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellarchitected_operations.R
\name{wellarchitected_upgrade_review_template_lens_review}
\alias{wellarchitected_upgrade_review_template_lens_review}
\title{Upgrade the lens review of a review template}
\usage{
wellarchitected_upgrade_review_template_lens_review(
  TemplateArn,
  LensAlias,
  ClientRequestToken = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The ARN of the review template.}

\item{LensAlias}{[required]}

\item{ClientRequestToken}{}
}
\description{
Upgrade the lens review of a review template.

See \url{https://www.paws-r-sdk.com/docs/wellarchitected_upgrade_review_template_lens_review/} for full documentation.
}
\keyword{internal}
