% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_batch_get_traces}
\alias{xray_batch_get_traces}
\title{You cannot find traces through this API if Transaction Search is enabled
since trace is not indexed in X-Ray}
\usage{
xray_batch_get_traces(TraceIds, NextToken = NULL)
}
\arguments{
\item{TraceIds}{[required] Specify the trace IDs of requests for which to retrieve segments.}

\item{NextToken}{Pagination token.}
}
\description{
You cannot find traces through this API if Transaction Search is enabled since trace is not indexed in X-Ray.

See \url{https://www.paws-r-sdk.com/docs/xray_batch_get_traces/} for full documentation.
}
\keyword{internal}
