% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_batch_get_traces}
\alias{xray_batch_get_traces}
\title{Retrieves a list of traces specified by ID}
\usage{
xray_batch_get_traces(TraceIds, NextToken = NULL)
}
\arguments{
\item{TraceIds}{[required] Specify the trace IDs of requests for which to retrieve segments.}

\item{NextToken}{Pagination token.}
}
\description{
Retrieves a list of traces specified by ID. Each trace is a collection of segment documents that originates from a single request. Use \code{\link[=xray_get_trace_summaries]{get_trace_summaries}} to get a list of trace IDs.

See \url{https://paws-r.github.io/docs/xray/batch_get_traces.html} for full documentation.
}
\keyword{internal}
