% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_delete_group}
\alias{xray_delete_group}
\title{Deletes a group resource}
\usage{
xray_delete_group(GroupName = NULL, GroupARN = NULL)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN of the group that was generated on creation.}
}
\description{
Deletes a group resource.

See \url{https://paws-r.github.io/docs/xray/delete_group.html} for full documentation.
}
\keyword{internal}
