% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_delete_sampling_rule}
\alias{xray_delete_sampling_rule}
\title{Deletes a sampling rule}
\usage{
xray_delete_sampling_rule(RuleName = NULL, RuleARN = NULL)
}
\arguments{
\item{RuleName}{The name of the sampling rule. Specify a rule by either name or ARN, but
not both.}

\item{RuleARN}{The ARN of the sampling rule. Specify a rule by either name or ARN, but
not both.}
}
\description{
Deletes a sampling rule.

See \url{https://paws-r.github.io/docs/xray/delete_sampling_rule.html} for full documentation.
}
\keyword{internal}
