% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_encryption_config}
\alias{xray_get_encryption_config}
\title{Retrieves the current encryption configuration for X-Ray data}
\usage{
xray_get_encryption_config()
}
\description{
Retrieves the current encryption configuration for X-Ray data.

See \url{https://www.paws-r-sdk.com/docs/xray_get_encryption_config/} for full documentation.
}
\keyword{internal}
