% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_encryption_config}
\alias{xray_get_encryption_config}
\title{Retrieves the current encryption configuration for X-Ray data}
\usage{
xray_get_encryption_config()
}
\description{
Retrieves the current encryption configuration for X-Ray data.

See \url{https://paws-r.github.io/docs/xray/get_encryption_config.html} for full documentation.
}
\keyword{internal}
