% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_groups}
\alias{xray_get_groups}
\title{Retrieves all active group details}
\usage{
xray_get_groups(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token.}
}
\description{
Retrieves all active group details.

See \url{https://paws-r.github.io/docs/xray/get_groups.html} for full documentation.
}
\keyword{internal}
