% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_sampling_statistic_summaries}
\alias{xray_get_sampling_statistic_summaries}
\title{Retrieves information about recent sampling results for all sampling
rules}
\usage{
xray_get_sampling_statistic_summaries(NextToken = NULL)
}
\arguments{
\item{NextToken}{Pagination token.}
}
\description{
Retrieves information about recent sampling results for all sampling rules.

See \url{https://paws-r.github.io/docs/xray/get_sampling_statistic_summaries.html} for full documentation.
}
\keyword{internal}
