% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_sampling_targets}
\alias{xray_get_sampling_targets}
\title{Requests a sampling quota for rules that the service is using to sample
requests}
\usage{
xray_get_sampling_targets(
  SamplingStatisticsDocuments,
  SamplingBoostStatisticsDocuments = NULL
)
}
\arguments{
\item{SamplingStatisticsDocuments}{[required] Information about rules that the service is using to sample requests.}

\item{SamplingBoostStatisticsDocuments}{Information about rules that the service is using to boost sampling
rate.}
}
\description{
Requests a sampling quota for rules that the service is using to sample requests.

See \url{https://www.paws-r-sdk.com/docs/xray_get_sampling_targets/} for full documentation.
}
\keyword{internal}
