% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_list_resource_policies}
\alias{xray_list_resource_policies}
\title{Returns the list of resource policies in the target Amazon Web Services
account}
\usage{
xray_list_resource_policies(NextToken = NULL)
}
\arguments{
\item{NextToken}{Not currently supported.}
}
\description{
Returns the list of resource policies in the target Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/xray_list_resource_policies/} for full documentation.
}
\keyword{internal}
