% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_update_sampling_rule}
\alias{xray_update_sampling_rule}
\title{Modifies a sampling rule's configuration}
\usage{
xray_update_sampling_rule(SamplingRuleUpdate)
}
\arguments{
\item{SamplingRuleUpdate}{[required] The rule and fields to change.}
}
\description{
Modifies a sampling rule's configuration.

See \url{https://paws-r.github.io/docs/xray/update_sampling_rule.html} for full documentation.
}
\keyword{internal}
